[Web Payment] Do not store PaymentAppService in BrowserContext
Before this patch, the PaymentAppService (which contains
PaymentAppFactories) was owned by BrowserContext, but that ownership
appears to be used only for injection of test app factories in unit
tests. That ownership structure persisted PaymentAppService until
browser shutdown, even when no payment was being performed any more.

This patch makes the PaymentRequestState on desktop and the
PaymentAppServiceBridge on Android own an instance of PaymentAppService.

After this patch, PaymentAppService on Desktop is deleted on iframe
reload or navigate (for example), when no payment apps are found, as
well as when PaymentResponse.complete() is called. On Android,
PaymentAppService is deleted when the factories have finished producing
payment apps.

Bug: 1382826
Change-Id: I280f55e998c996f0e99c67e8fc1bf3bc0401ad56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4017573
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1072935}