Use Rust parsing from base/json via unidirectional FFI bindings.
Previously this was implemented with bi-directional FFI. We wrote a
wrapper Rust type around the C++ base::Value implementation. This
removes the wrapper type and all binding generation from Rust to C++,
instead only generating C++-to-Rust bindings. This also means
removing the `base` Rust module and making the //base component
no-longer a `mixed_component`.

Function pointers are used to get back to C++ in order to construct
the C++ base::Value tree. However due to lack of support for actual
function pointers in the interop binding generator (for now), we use
a structure of function pointers and methods, and then have CXX
generate access to the methods, which simply call through to the
function pointers.

To avoid UB in C++, the base::Value& casted to a ContextPointer& must
not be used in that form.

Toward standardizing the Rust build flag situation, we always include
//build/rust:buildflags into the build, and allow C++ files to then
conditionally use rust internally, instead of forcing the decisions
to be made in many GN targets. To do this, we must put the other
parts of //build/rust/BUILD.gn behind the `toolchain_has_rust`
conditional.

We simplify the GN flags in base/BUILD.gn to not check both
`enable_rust` and `toolchain_has_rust` since the latter implies the
former already.

Bug: 1069271, 1381552
Change-Id: I1b4fe4ba3346710a75969827f5f4230038718093
Cq-Include-Trybots: luci.chromium.try:linux-rust-x64-rel,linux-rust-x64-dbg,android-rust-arm-dbg
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3930926
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Auto-Submit: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1072939}