bookmarks: fixes bug in DropBookmarks() when copying
DropBookmarks() has a code path that looks up the index of a
node. The node supplied to GetIndexOf() is the original node. This
is the wrong node to use if a copy is created (because copying
creates a new node). This fixes the copy path to increment the
index.

Bug: b:257217681

Change-Id: If982e624be373a301898dad21cdfbb8a40c3debb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4033027
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1072972}