[Lens]Add back button navigation support for Lens camera
The back button will be enabled if the current page is a Lens camera
experience that is enabled (this checks feature flags, etc.) and the
user is at the bottom of the navigation stack. In that case, we open the
Lens camera and close the current tab (regardless of what the user does
after opening the camera).

Bug: 1353430
Change-Id: I82375a1f7b915eef9aebe720874d4005b98017a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3965534
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Jason Hu <hujasonx@google.com>
Cr-Commit-Position: refs/heads/main@{#1072973}