WebUI: Remove unnecessary dependencies on js_library mojo targets
<target_name>_webui_js is actually a js_library() target, and is not
responsible for generating any JS files (including .mojom-webui.js
files). However, it is used as a proxy for the target generating such
files in a lot of places.

Things like copy() or grit() that are directly handling the generated
mojo files should instead be relying on mojo targets that are
responsible for generating those files, which have the form
<mojo_target_name>_js__generator.

This is in preparation for removing generation of such targets on
non-ChromeOS platforms.

Bug: 1316438
Change-Id: Ia902fc5f888bf11b3f3d84406034e15dc151d972
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4029046
Reviewed-by: Connie Wan <connily@chromium.org>
Reviewed-by: Florin Malita <fmalita@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Auto-Submit: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: Connie Wan <connily@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073047}