subframe-reuse: remove experiment arms
The subframe-reuse feature was enabled by default in
crrev.com/c/3811446, which landed in version 107.0.5292.0.

At least a milestone has passed since the feature was enabled by default
(per the guidance in go/finch-best-practices), so the experiment code
can start to be removed.

This change removes the various experiment arms whose behavior was
not enabled by default (as they did not perform as well as the
default "constant delay" behavior). The base::Feature itself will
be cleaned up in a subsequent CL.

Bug: 894253
Change-Id: If33d1c35d2ed21359bdfd9457b90f63f48a5d66d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4007041
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Cr-Commit-Position: refs/heads/main@{#1073093}