Mock out low level connection details in ChromotingSessionServicesClientTest
Isolated connections are being deprecated and will not work as we expect
after switching to MojoIpcz. One big difference between the two is that
isolated connection supports self-to-self IPCs while non-isolated
connection doesn't, and the client will hang when trying to connect to
the server.

In crrev.com/c/4027301, I turned NamedMojoIpcServerTest multi-process,
but it's tricky to turn ChromotingHostServicesClientTest multi-process,
since the subprocess (server) would need to signal the client that it's
ready to receive connections, which is difficult to do with the existing
library. So instead I just put all the low level stuff into a
ServerConnector callback and mock it out in unit tests.

The deleted test comment also called for verifying that the client can
reconnect after the server is restarted. The test on the low level has
actually been added to NamedMojoIpcServerTest in crrev.com/c/4027301, so
it doesn't need to be tested in ChromotingSessionServicesClientTest.

See also: go/crd-ipcz

Bug: 1378803
Change-Id: Id0c5f17b22879a2082772ea8096c51607c41c942
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4032576
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073119}