LockScreen: Prevent System Web App Manager from existing on LS app profile
The lock screen app profile is intended only for installing/running
regular web apps for use on the lock screen (on CrOS). System web apps
should not be used with that profile, but were inadvertently enabled on
it because they use the same check as web apps to instantiate their
keyed service (in SWAMF::GetBrowserContextToUse).

Fixing the above led to an audit of usages of SWAM::Get and therefore
also fixing crbug.com/1353262 so SWAM doesn't exist in Kiosk either.

Bug: 1006642, 1353262
Fixed: 1353262
Change-Id: I9ecaf0b8571cb0e60b045494355b80fc555e95fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3937645
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Richard Chui <richui@chromium.org>
Reviewed-by: Jiewei Qian <qjw@chromium.org>
Commit-Queue: Glen Robertson <glenrob@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073132}