Resize partial text label after receiving actual source language
The initial partial translation uses "Detected Language" as the source 
by default, which can be a much longer string than the actual source 
language. The text label for translated text needs to be resized after 
receiving the actual source language string to accommodate this 
difference, otherwise the bubble is shown artificially wide.
Note that the width of the loading view will still be affected by this 
issue, as at that point we don't know the source language yet.

Bug: 1384670
Change-Id: I1e982784c7bc117bf778000d0c523a7965f5e42a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4026299
Reviewed-by: Basia Zimirska <basiaz@google.com>
Commit-Queue: Anthony Cui <cuianthony@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073142}