view-transitions: Safely handle concurrent transition in local frames
The current stack can only support one transition tied to a compositor.
This is an issue if multiple local frames tied to the same compositor
trigger a transition concurrently.

Safely handle this by skipping an existing transition if another one is
initiated within the same local frame tree. A transition initiated by
a parent will take priority over child frames.

The test for this is chromium specific, instead of a WPT, since this is
unspec'd behaviour. Its a limitation of the current stack that will be
eventually improved.

Bug: 1376158
Change-Id: If173e39e87efc7b63e26821e6608ce16ffa771e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4022033
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Auto-Submit: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073210}