Reland "Remove kSkipAccessibilityPaintChecks from Ash WindowCycleTabSlider"
This is a reland of commit 695b0f2e4bc5940423ea9e2b8bae29eb5fa8416c.

It had been reverted due to the DCHECK being reached during manual
testing. See b/257301699.

The reason the DCHECK failed is that the active button selector claims
to be focusable. This appears to have been done to improve the visual
appearance of the UI; the selector is not end-user focusable. If this
view needs to claim focusability, it needs a valid role and accessible
name in order to not be flagged by the accessibility paint checks.

Give the active button selector a generic role (kPane) and set its name
to explicitly empty, which is appropriate for focusable views which are
not end-user focusable. Add test coverage for the paint check failure.

Bug: b/257301699
Bug: 1348081, 1218186

Original change's description:
> Remove kSkipAccessibilityPaintChecks from Ash WindowCycleTabSlider
>
> It appears that whatever was causing the paint checks to fail is no
> longer an issue.
>
> Bug: 1348081, 1218186
> Change-Id: I4015a5b44be4517d0aa439fb2577037940d88dfe
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3955757
> Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
> Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1060443}

Change-Id: Ic5ea8de20ffcfa04fc7a9e8fc430061ee994b7b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4020278
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073252}