[Metadata][Clank] Append dimension specs to card art URL
The card art icons are served via a FIFE URL which can be formatted to
serve icons of a specific required size. The native side passes the raw
URL (without dimension params) to Clank. We have to append the dimension
params on the Java side considering the following 2 criteria:
1. Autofill component. The URL is used by the Dropdown, Keyboard
Accessory, and the Credit Card Accessory Sheet, and they do NOT share
the same dimension specs.
2. Phone's display density. The URL params are to be formatted in
pixels, so we need to calculate the pixel dimension corresponding to the
icon size in dp.

Card art icon in Keyboard Accessory (32 x 20):
https://screenshot.googleplex.com/77JtiEFu3Hkv3KS
Card art icon in KA (64 x 40, just to show that we can fetch images of
different sizes by appending params to the FIFE URL):
https://screenshot.googleplex.com/4H6jEfJX3XoHVeG

Note: To get the phone's display density, we need a context (a handle to
the system). AutofillPopupBridge's addToAutofillSuggestionArray method
makes the call to fetch the icon, but currently does not have access to
the context as it is static. Make it an instance method so that it has
access to the context.

Bug: 1380501
Change-Id: Id62c752c38ec2c4977f302632efce586a35a3e01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3997634
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Reviewed-by: Siddharth Shah <siashah@chromium.org>
Commit-Queue: Vishwas Uppoor <vishwasuppoor@google.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1073259}