Fix COOP bypass that uses reload redirection.
Current code is not well suited to handle history navigations to
unexpected targets. This can be surprising when for example a reload
ends up with a redirect or an error (see https://crbug.com/1239540).

Instead, we base ourselves on the response, with up to date security
information, and then decide if we should use the destination
SiteInstance.

Bug: 1374705
Change-Id: I43f8bf55456a93aa01ec2cab6e7b55f6da403e3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4016449
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Arthur Hemery <ahemery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073311}