[Managed Prefs] Cookies settings UI disabled for sites managed by policy
This CL blocks the UI that allows the user to change cookies exceptions controlled by policy, which as far as we can tell is only a UI bug without any downstream effects. A pop-up with a remove button appeared when the user clicked on the site preference.
With this CL, the preference now shows that the site is managed. A toast is also printed when the user tries clicking on it, and the pop-up is no longer shown.

Screenshots: https://bugs.chromium.org/p/chromium/issues/detail?id=1323274#c16

Bug: 1323273, 1323274
Change-Id: I2d7329ce02895e79e89ec42b05645ad356a81647
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3988093
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073420}