[iOS] Minor cleanup of new tab animation code
This refactors and cleans up the new tab animation code a bit. This
change was broken out of the change that removes the animation when
opening a new tab from an external intent (e.g. the search widget).
These changes were not necessary in that branch, so I've made them
here instead.

Bug: 1213172
Change-Id: Iad3f36baf2582bafe4549978f1e62f6a1c7a2a9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4003065
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1073456}