Only Create PrivacyMetricsService for Regular Profiles
In the process, rename metrics that are emitted on profile open. (Deprecate the metrics and replace them with a version 2.)  For one obsolete metric, I didn't replace it but instead revised the histogram description to explain the problem with it, in case people look at its old data.

Bug: 1284659
Change-Id: I041dd0c9c6b7cf38001b2f1866ce43fb59289c94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4020815
Reviewed-by: Sean Harrison <harrisonsean@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Mark Pearson <mpearson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073532}