CullRectUpdater optimization for the root layer
Previously we always updated the contents cull rect of the root layer.
This is not a big deal, but it's better to avoid that. Also fix the
scroll enough logic for the root layer. Previously we didn't update
LastCullRectUpdateScrollPosition for the root layer, which wasn't a
problem as we always updated the contents cull rect of the root layer.

Bug: 1253726
Change-Id: I07f9ad61ea20c18c85603b73c760f19f670acc5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4036945
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073616}