Fix crash in TouchSelectionMenuViews::DisconnectOwner()
There are two code paths that end up in closing a touch selection menu.
One that we explicitly call `TouchSelectionMenuViews::CloseMenu()` in
response to a user action on the menu and one that the window is closed
for some other reason (e.g. clicking outside the menu) which goes
through `TouchSelectionMenuViews::WindowClosing()`. Both of these paths
disconnect the menu from its owner (menu runner). The first path
eventually ends up calling the second one, so we guard the second path
by early returning on a null `owner_`. However, our assumption was that
the second path cannot be hit after the first one, so it was guarded
with a DCHECK. Apparently, this is not the case and there is a race
condition between the two paths in which we might end up calling the
second path and disconnecting the owner and then calling the second path
and hitting the DCHECK. This CL replaces the DCHECK with an early return
to gracefully handle this case.

Bug: 1343326
Change-Id: I54ba69ee8dcfec44999440eb161eb13181d87531
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4034920
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073710}