[ReadingList] Avoid inheritance from storage to ModelTypeSyncBridge
Conceptually the storage doesn't need to support any sync-related
functionality, so the interface ReadingListModelStorage should not have
a dependency to sync interfaces (ModelTypeSyncBridge).

Removing the dependency to ModelTypeSyncBridge in a first patch isn't
particularly simple, so as a first step in this direction, inheritance
is replaced with a getter, which test-only implementations can trivially
return null for (TestReadingListStorage).

Change-Id: I1236deffb048d735b1a31b6d31368efc0a38f737
Bug: 1386158
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4037270
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073979}