Fix population of alternative state name map when sync is started.
Currently, the autocomplete, Autofill and ContactInfo sync bridges emit
the NotifyThatSyncHasStarted(model_type) event after the initial sync.

This event is solely consumed by the PersonalDataManagerCleaner, which
uses it to (among other things) to populate the alternative state name
map, if that hasn't been done yet.

Since the logic currently doesn't check which model type was started, it
can be triggered from the autocomplete sync bridge, before the profiles
are even available. This leaves the alternative state name map empty.

With this CL, the initialization only happens after profiles from both
sources (the Autofill and the ContactInfo sync bridge) have been loaded.
This is done by adding to booleans to the PersonalDataManagerCleaner's
state.

Ideally, these booleans should be reset if sync is disabled. Since we
don't have such an event, this is not done. This is not a regression,
since the alternative state name map already cannot be populated
multiple times due to an `is_alternative_state_name_map_populated()` if.
So even in the current state, when sync is disabled and re-enabled, we
wouldn't initialize again. A TODO statement concerning this fact is
added.

Bug: 1313616
Change-Id: I06f350264daaf7df43b3c2e302e29b42c3c4efe8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4020388
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1073980}