iwa: Disallow primary-url Web Bundle section
The optional primary-url section of Web Bundles "records a single URL
identifying the primary URL of the bundle. The URL MUST refer to a
resource with representations contained in the bundle itself."

Chromium's `WebBundleParser` supports this section, hence the
`SignedWebBundleReader` should also support it, given that the
`SignedWebBundleReader` is supposed to be a generic reader for Web
Bundles that happen to be signed.

However, in the context of Isolated Web Apps, such a primary URL does
not make sense, given that Isolated Web Apps also have a web app
manifest potentially containing a `start_url` field. Therefore, this CL
updates the Isolated Web App validation code to validate that a primary
URL is not set and updates some tests that were setting primary URLs.

Bug: 1385394
Change-Id: I82b584f9be40d9d2f20aa40a96979eff3ec80015
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4030813
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Phillis Tang <phillis@chromium.org>
Commit-Queue: Christian Flach <cmfcmf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074064}