Don't cache observer iterator in WebContentsObserver.
If a class that listens to a WebContentsObserver triggers another
WebContentsObserver event (eg. by changing the focus of the WebContents)
then any observers after that observer in the observerlist won't
receive the first event at all.

Fix this by not caching the iterator for the observers so it doesn't
get reset when a new event is triggered.

Bug: 1385302
Change-Id: I7216da6c2a319528caa14a73e8effb5bb3667f97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4036825
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074239}