shortcuts: Use LayoutInfo in AcceleratorSubsection
This CL stores each row's LayoutInfo as part of the AcceleratorRowData
in the AcceleratorSubsection element. This is necessary for the next
downstream CL, in which the layoutInfo.style is passed into
AcceleratorRow. Later on, LayoutInfo is used here to pass the
accelerator text property to AcceleratorRow.

As part of this CL, I've also renamed the type AcceleratorRowData (from
Accelerator) and other variables to improve clarity. Also,
AcceleratorRowData did not need to be exported so I removed
the `export` keyword. Finally, since LayoutInfo is now part of
AcceleratorRowData, the 'source', 'action', and 'description'
properties were removed since they can be derived from LayoutInfo.

Also, in accelerator_subsection.html, I removed the unused class
"acceleratorRow".

Bug: b:216049298
Test: browser_tests --gtest_filter=ShortcutCustomizationApp*
Change-Id: Icb1c9fc1e9b1db6ef7c670565288a91ac99ee845
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4037832
Commit-Queue: Camden Bickel <cambickel@google.com>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074247}