Additional tests to ensure Ed25519 keys are valid
Most of the test we have for the Ed25519 algorithm's operations rely on
the import / export API to manage the keys. We have specific tests for
the generateKey API, but we don't verify the obtained keys are correct
to be used in the algorithm's operations.

The new tests ensure that keys created using the generateKey API are
valid for the 'sign' and 'verify' operations.

Bug: 1370697

Change-Id: I0499ff809fee295e14b5e699f93cbe4619fb88eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4031830
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074283}