immersive fullscreen mac: workaround for blank top chrome
NSTitlebarAccessoryViewController has a bug when moving from a
fullScreenMinHeight of zero to >zero. A blank rect will be displayed
instead of the hosted view. A radar and dts have been opened with Apple.

This CL works around this behavior by keeping a small sliver of a
NSTitlebarAccessoryViewController on screen when the toolbar is hidden.
This sliver (0.5px) is not readily perceivable because it blends in with
the menu bar's black background.

Bug: 1369643
Change-Id: I6a92a75d004f3795b7150ca07b2ada1263de5a66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4036047
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Tom Burgin <bur@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074285}