[Ntp Redesign] Handling NPE when currentStream is null
We are seeing a NPE when section headers are being bound and header tab is selected. mCurrentStream is showing up as null - could be because this is set during bindStream that might be happening after header setup. Calling updateLayout() when stream is bound and adding null checks on stream.

Bug: 1386057
Change-Id: Ia962b12c9b48eff2ed92ecf98edc9f09649f7145
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4037086
Reviewed-by: Cathy Li <chili@chromium.org>
Commit-Queue: Sirisha Kavuluru <skavuluru@google.com>
Cr-Commit-Position: refs/heads/main@{#1074325}