[Cormorant] Add dynamic title and url in the profile section
This CL allows the code to take in the Title and URL received from the CreatorBridge, and dynamically display them in the Profile section. It also converts the WebFeedId to a string before sending it through the Bridge, and changes the URL to be in a String format as opposed to GURL. Also, this CL removes the temporary activity title.

Screenshot:
- Follow: https://screenshot.googleplex.com/3BPS9gsHoiNQRPH
- Following: https://screenshot.googleplex.com/8MPh6wvf5dkQAgo

As an added bonus, our example creator matches the mocks too!

Bug: 1377071
Change-Id: Ifbe0be3c72aac414a553cea554d341cb24877ab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4028789
Commit-Queue: Wissem Gamra <wiss@chromium.org>
Reviewed-by: Jeffrey Cohen <jeffreycohen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074336}