[Office upload] Store whether the setup has ever been completed
So that we don't run the setup every time, use a preference to store
whether the setup has previously run to completion or not.

To allow easy testing, add a button to the chrome://files-internals
page to unset this pref, and show the current status of the pref
there too.

Bug: b:251046333
Change-Id: I1359655beb9eac867087c4f9f5cfb7c1d36048c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4036949
Commit-Queue: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Dominic Schulz <dominicschulz@google.com>
Cr-Commit-Position: refs/heads/main@{#1074374}