Files app: Don't throw Error in volumeManagerUtil.validateError
Just log a warning message instead.

Throwing an Error results in an uncaught error (in promise) and causes
other issues.

For example, if the user tries to open (ie mount) an archive, and the
mount operation returns an unexpected error string, the exception thrown
by validateError prevents Files App from displaying an error message to
the user. When this happens, the mount operation is seen as never
terminating.

On the other side, by letting this (unexpected) error string bubble up
the normal way, a generic error message is shown to the user, which
seems a preferred outcome.

BUG=None

Change-Id: I50ec3a8221833067104df3d0194273208e8db252
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4044724
Commit-Queue: François Degros <fdegros@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074376}