bluetooth: Update wpt_internal tests to use garbageCollect utility
These tests previously relied on the runGarbageCollection utility method
which was removed from external web platform tests in favor of a common
utility method for all tests. This CL updates the wpt_internal tests to
use the same utility as the external tests and re-enables the tests.

Bug: 1386241
Change-Id: I344baeb150077510f608810b428a5358360fcde3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4044108
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074410}