WebGPU: Throw RangeError when deviceCreateBuffer() returns nullptr
This patch makes Blink throw a RangeError when deviceCreateBuffer()
returns nullptr.

According to the latest WebGPU SPEC, a RangeError should be generated
when we fail to initialize an active buffer mapping when
mappedAtCreation == true, and when an OOM happens on dawn_wire::client
when creating buffer with mappedAtCreation == true, deviceCreateBuffer()
should return a nullptr.

Note that this CL is landed before we land the CL in Dawn so that there
won't be any potential null pointer deference in Blink.

Bug: dawn:1586
Change-Id: I6c5236f057f13fb59f30db57026406b666b67c75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4041633
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>
Reviewed-by: Austin Eng <enga@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074411}