[PDF Viewer] Fix navigation tests for invalid URLs
When the expected URL for a navigation test is undefined, the test
succeeds, regardless of whether or not navigation occurs. The expected
behavior when the expected URL is undefined is for navigation to not
occur.

Properly check that navigation does not occur if the expected URL is
undefined.

Change-Id: Idbb9e1c98683d4c34e6ee1fa3b907e0c8ab510f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4039502
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074414}