ozone/wayland: fix potential crash with unsupported buffer formats
The hardcoded RGBA_8888 in SkiaOutputSurfaceImplOnGpu might cause
crash if WaylandOverlayDelegation is enabled. This depends on if
the Wayland implementation supports RGBA_8888.

On Linux, taking Gnome Mutter as example, the support of RGBA_8888
was introduced about a year ago [1], which means systems running
old versions of Mutter (like Ubuntu 20.04) will definitely crash.

In this CL, BGRA_8888 is used instead if the default RGBA_8888 is
not supported.

[1] https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1999

Bug: 1371420
Co-Author: Maksim Sisov <msisov@igalia.com>
Change-Id: I35667aae4d25e1078f9f053ccd1381540c8ed74b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3959355
Reviewed-by: Peter McNeeley <petermcneeley@chromium.org>
Commit-Queue: Jiahe Zhang <jiahe.zhang@intel.com>
Reviewed-by: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074438}