[Files]: Open query input via method not CSS.
The old code relies on CSS class being present and goes around the
controller element to open the query input. That results in the query
input being hidden. This change exposes openSearch method on the search
container. It then uses the fact that search command has access to the
fileManager object, on which the public property ui exposes
searchContainer to it. Thus it opens the query input by asking the
search container to do that, rather than by tweaking focus on a private
input element.

Bug: b:259857934
Change-Id: I4edc31327b214feee21088f39fe67585ab5a331e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4041928
Commit-Queue: Bo Majewski <majewski@chromium.org>
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074445}