Always remove origin ID from list when taking one
Noticed when working on the code that if an invalid entry ends up on the
list, it is not removed so that subsequent attempts will try with the
same invalid entry. It should be removed from the list so that future
calls will try with another (hopefully valid) entry.

Bug: none
Test: MediaDrmOriginIdManagerTest unit_tests pass
Change-Id: I6ff128f34d3ce5da70b652c1b59422f3a71fa938
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4039834
Commit-Queue: John Rummell <jrummell@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074469}