[iOS] fix ordering of StoreKit tab helper init.
The current ordering of initialization in BrowserCoordinator sets
TabHelper delegates before coordinators are created. This means that
for tabs that are present during a cold start, there will be no StoreKit
or PassKit coordinators when the tab helpers for those features have
delegates attached.

This CL is a stopgap fix, which adds a second round of webstate delegate
attachment after coordinators are started, along with DCHECKs to catch
assigning nil delegates to these tab helpers.

The intent is that this CL can be cherrypicked into M109. A more complete
cleanup plan is discussed in the TODO bug (crbug.com/1392109) referenced
in the CL.

Bug: 1380980
Change-Id: Ic460770d2718dc15a58aaa4ef422160527abb0bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4043122
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074513}