Do not log the optional `tts` field in trace data tests
Because this field is optional, it may be included in some instances and
not in others. In particular, it's been seen to not be included when
running on windows. As such, this CL omits it from the trace data tests
logs to prevent flakiness.

Fixed: 1385497
Change-Id: I5374caf8334de116657a61228fa826de94a1b030
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4031055
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074539}