Only log the first BeginFrame ocurrence in frame lifecycle test
This is a reland of commit 8a086eb012eb84f0d4dd8288c9b05094720603b9

By mistake, the test was logging all the "BeginFrame" events it
encountered in the trace it gathered. This is problematic because we
don't know with precision the amount of "BeginFrame" events that will be
dispatched during the test execution, which introduces non determinism
to the test.

As such, this CL changes the test so that only the first occurance is
logged, since we know at least one should occur and we only care about
its shape.

Bug: 1383130


Original change's description:
> Add tests for frame lifecycle trace events
>
> Tests for trace event data related to frame lifecycle analysis in DevTools Perf panel. This includes Screenshot trace events.
>
> Bug: 1383130
> Change-Id: I0857c791cfa6b678c04cd0829cbd1c5215dbf07e
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4030386
> Reviewed-by: Simon Zünd <szuend@chromium.org>
> Commit-Queue: Andres Olivares <andoli@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1074072}

Bug: 1383130
Change-Id: I1a55ad696fd890febf0a6b45b7013a7b5196365c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4046302
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074543}