DigitalAssetLinksHandler should take url::Origin instead of std::string
Digital Asset Links live in an origin's .well-known directory so
conceptually this already is a requirement, so most consumers of this
class already store the origin in either an url::Origin or GURL object.
This aligns the DigitalAssetLinksHandler method signatures, while
maintaining DAL terminology, and removes a bunch of unnecessary copies.

Change-Id: I8fc50ec44ada91558f10b9d65c1a988c9d042060
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4042668
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Commit-Queue: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074561}