[SPC] Refactor SecurePaymentAuthenticatorCreate*Test tests
This CL merges together a number of overlapping credential creation
tests:

- SuccessfulEnrollment
- CredentialType
- CreatePaymentCredential
- CreatePaymentCredentialTwice

In doing so, it introduces a single JS entrypoint for creating a
credential, which returns a JSON object string with the created
credential's information. This is then parsed on the C++ side into a new
struct, PaymentCredentialInfo, and used in tests.

Bug: 1385134
Change-Id: Iea0f6c8fa628058eabeec1847a14c924619f32e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4043642
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074619}