Reland: CrOS Settings: Use RouteObserverMixin
This is a reland of crrev.com/c/3942946. This CL was initially
reverted because it was part of a CL chain where the base CL was the
culprit and needed to be reverted. See crrev.com/c/3943835. The base
CL was relanded so this CL can safely be relanded too.

Replaces all uses of RouteObserverBehavior with the mixin version in
all current TS code for ChromeOS Settings.

Bug: chromium:1315757
Test: Sanity check of all subpages
Change-Id: I0f6a7fabe0d8647626e2c473eb4cc1e0cdfb3b50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4043493
Commit-Queue: Wes Okuhara <wesokuhara@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074716}