saved_desks: Handle dialog controller not being available.
This CL changes calls to saved_desk_util::GetSavedDeskDialogController
to handle the case where it might return null instead of just assuming
that it will always be there.

A SavedDeskDialogController object is bound to the lifetime of its
owning OverviewSession. We had code in the saved desk UI that assumed
that the dialog controller was always available. The saved desk UI is
owned by the overview session, so on the face of it this seems like a
fair assumption.

However, when overview mode is exiting, the ownership of the saved desk
UI is transferred to an animation. This means that the overview session
can be gone while saved desk UI components are still alive. The above
assumption therefore doesn't necessarily hold. There may be other such
cases.

BUG=b:258306298

Change-Id: I79fc6d6742b52f4bb1bfedcbed4c4488be164846
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4034804
Commit-Queue: Daniel Andersson <dandersson@chromium.org>
Reviewed-by: Richard Chui <richui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074753}