[ChromeCart] Check cart load success
This CL fixes a potential UAF bug. When ChromeCart database (backed by
SessionProtoDB) fails to initialize, the callbacks from load attempts
will be posted to thread pool[1]. This could lead to the callback to
not be canceled and still execute after the component that initializes
the callback (e.g. CartService, CartHandler) is destroyed, and the
callback execution could potentially trigger UAF. This CL fixes this
issue by adding checks in the corresponding callbacks, and early return
if the load was failed.


[1] https://source.chromium.org/chromium/chromium/src/+/main:components/session_proto_db/session_proto_db.h;l=205-208;drc=7b5df6d4083453cd77a2fd7a1ca1ffce54154f4b;bpv=1;bpt=0

Bug: 1385831, 1385709
Change-Id: I75dfa658e988c16ed8a97839a7f57e5d615d0bf6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4044017
Reviewed-by: David Maunder <davidjm@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074781}