Avoid quadratic behavior of image decoders.
With DCHECKs enabled, image decoding was quadratic in image size.
This is because the removed DCHECK_EQ was calling Data()
a linear number of times in the image byte size, and
Data() write the entire image onto a new buffer: each call was
thus allocating a linear number of bytes.

Bug: 1392800
Change-Id: Ic3addaad9edaa37d9c263e68e3ac4e7debe24ac8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4047191
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Luca Versari <veluca@google.com>
Cr-Commit-Position: refs/heads/main@{#1074817}