🤠 Disable MainActivityWithURLTest#testLaunchActivity on phones
This test needs to be updated to account for Start showing on startup
when there are no tabs and the NTP is set as the homepage.

Disabling for now while we sort out how to cover the various startup
scenarios.

Bug: 1392547
Change-Id: Ie05245676ff9d4e476bdf0463b897deddbc4cb39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4048908
Commit-Queue: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Shu Yang <shuyng@google.com>
Owners-Override: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074823}