Extend use-winrt-midi-api flag
Web MIDI for Windows was implemented on top of the Win32 WinMM API,
which is now pretty old and doesn't provide hot-plug related features.

Another backend for Windows 10+ was implemented on WinRT, which is
selected by this flag.

Until we can enable the new backend by default we need this flag.

Change-Id: I40349c7e0e4606017e7afc9dbb0ebfb486fe5ab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4037915
Commit-Queue: Michael Wilson <mjwilson@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074828}