WebUI: Make dialog title an h2 element
This CL updates cr-dialog's title slot to be automatically a h2 element
to adhere to accessibility guidelines. This CL also removes existing
h2 elements from some uses of cr-dialog and also demotes other headings
to h3 for headings that are not the title.

Fixed: 1312612
Change-Id: I8e73f9a01fa79ef9712ee35fe444b453c39ef999
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4044083
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074847}