[Android Tab Groups] Force Tab Groups to be contiguous
TabGroups are expected to be contiguous; however, this is not strictly
enforced.

addTab and moveTab do not check for Tab Groups. This can result in
tabs that are part of a group being positioned disjoint from other
members of their group OR tabs that are not part of a group can be
interposed between members of a group.

This CL fixes this problem by having all addTab actions check
TabGroupModelFilter (if it exists) to ensure any tabs are positioned
correctly relative to existing groups. This should also fix any cases
where this is already a problem upon tab restore.

This CL also removes the only usage of moveTab that is not aware of Tab
Groups and replaces it with setting the index as part of creating the
tab.

Emit a histogram on tab state initialization to check how often there
are issues.

Logic
- An ungrouped tab is added inside a tab group it gets pushed to after
  the group.
- A grouped tab placed with its group has no change occur, but if it
  is placed away from its group it is moved to be adjacent to the
  closest member of the group.

Bug: 1383067
Change-Id: I67538b42e23c40941b5ce1744cae5bb000f8debf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4022011
Reviewed-by: Fred Mello <fredmello@chromium.org>
Reviewed-by: Yue Zhang <yuezhanggg@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074896}