[WebXR] Fix bug with cancelling ARCore Install
Previously install was checked by re-requesting an installation and
expecting the ArCore APK to throw an exception since it wasn't requested
by the user. However, at some point it looks like this stopped throwing,
as in certain cases if the user cancelled an ongoing ARCore install an
exception wasn't thrown, despite the prompt still being up. The page
would then receive notification that installation had succeeded, try to
start a session, and then fail. The page would then become
non-interactive because what looks like the new AR surface would not
be cleaned up properly. When attempting an install after a refresh,
Chrome would crash, I believe because not all of the status was properly
cleaned up.

This fixes the latter issue (the crash) sepculatively by cleaning up the
pending install state when the native-side of the install utils are
destroyed, but also prevents it from getting in that state by fixing the
first issue as well.

I had briefly considered simply checking the return value from
requestInstall rather than relying on it to throw; however, this had the
side effect of forcing the installation prompt to reappear after the
initial cancellation. By checking the availability again, we allow the
prompt to be dismissed. There had previously been warnings that this
prompt may take a while; but those have since been removed, and this is
an asynchronous operation anyways, which most users shouldn't see, so
the slight potential for extra time is a reasonable trade-off for the
better experience.

Fixed: 1280475
Change-Id: Ib6d7359af3af53151bfaec6b0e8e7dbfb3146260
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4049686
Commit-Queue: Piotr Bialecki <bialpio@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074940}