qs: Adjust tray detailed view spacing for quick settings revamp
Modify several functions in TrayPopupUtil to take a "wide" parameter
that configures the layout for quick settings revamp. In particular,
the TriView START and END containers have their minimum sizes set
narrower (for example, to tightly contain a 20x20 pixel icon). Then
wide insets are added to compensate for the wider quick settings (for
example, 24 pixels of left inset).

An opposite change is made to TrayToggleButton, allowing it to be
sized "small", so it fits in the narrower TriView END container.

The detailed pages that have been converted to QsRevamp are switched
to use "wide" mode when the QsRevamp flag is enabled (for example,
Bluetooth, a11y, IME). Other detailed pages are left as-is. Also,
calendar view continues to use the existing layout.

Screenshots (the font corruption in some of these is a problem with
my emulator, not the code):
https://screenshot.googleplex.com/8Rrjku4AFASAZh3
https://screenshot.googleplex.com/5S6pahyFuw5TSPc
https://screenshot.googleplex.com/6J63pWSe56QCqxq
https://screenshot.googleplex.com/4E4No55XUDUmkBQ
https://screenshot.googleplex.com/Bmrs8hi6usyPDcB
https://screenshot.googleplex.com/C5R4xpQ3KZtxXar

Bug: b:252872600, b:252870873, b:253091169
Change-Id: I1abe501ae41d3ddabf0ee0c0c6bb748424211f92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4049761
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074946}