Clicking Ash shelf icon shows a reduced list of tabs
When it comes to clicking shelf icons, apps (YouTube, Map,
YouTube Music, Gmail, ...) behave differently from the Ash browser. In both cases, clicking the shelf icon would reveal a list of tabs of the app that is being clicked. The similarity ends here. For apps, this list should only contain tabs whose web contents belong to their original app. It's possible within an app the user wandered into completely unrelated web contents. For tabs showing those unrelated contents, it should not be present in the aforementioned list of tabs. In the meantime, this "filtering" mechanism shouldn't happen if the shelf icon in question is the Ash browser, for obvious reasons.

This CL makes sure no "filtering the list of tabs" is happening when
clicking the Ash browser shelf icon, so that the list of tabs being
rendered is complete.

Bug: 1328064
Change-Id: Ie325cf8f46d7f7d8e034cbf03675be720cbcb457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4040848
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Di Wu <diwux@google.com>
Cr-Commit-Position: refs/heads/main@{#1074992}