LacrosTablet: Adjust bubble if offscreen for Lacros
This CL sets PlatformStyle::kAdjustBubbleIfOffscreen true on Lacros.
If this is false, bubble position is not adjusted properly.

For example, crbug.com/1362816 is caused by this issue.
BubbleFrameView::GetUpdatedWindowBounds() didn't update the bounds to
proper position so that Lacros textfield bubble window is positioned
outside of the screen.

Bug: 1362816
Change-Id: I1cf125c9a5343ac97f225ebb29f1b002e5266f8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4045529
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075078}