Add policies to control the DNS over HTTPS identifiers feature
This allows to set the policies DnsOverHttpsTemplatesWithIdentifiers and
DnsOverHttpsSalt to allow including variables in the template. Those
variables will then be evaluated when generating the final URI to be
passed to the DNS over HTTPS server. The server can use the content of
the variables for making logging or routing decisions.

The feature is only supported on ChromeOS. For other OSs the original
policy DnsOverHttpsTemplates will be used (if set) which does not allow
including variables. The cross platform policy DnsOverHttpsTemplates
was not extended with identifiers to preserve existing behavior and
because the client request asks for machine identifiers which are not
available on other platforms.

This CL only adds the policies, for the implementation see the followup CL:4013345.

This CL was originally reviewed as CL:3913063.

Bug: b/230468360

Change-Id: I41b976f35f4d9cccca3c8b05b4e2765c55d4570e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3990794
Reviewed-by: Eric Orth <ericorth@chromium.org>
Reviewed-by: Roland Bock <rbock@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Andreea Costinas <acostinas@google.com>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075090}