[Privacy Hub] Keep app name for camera hw switch notification updated
The notification informing the user that the camera privacy switch on
their device is active has an optional app name that can run out of
sync. The CameraClientType change listener isn't invoked when an
app is closed but only when the HAL client type (i.e. all apps) of one
stops using the camera.

Through the capability access changes we can track when an app stops
accessing the camera and then can update the application name in the
notification if necessary.

Bug: b:258666453
Change-Id: I1207bd52d3d2562edc91fab34fb510fcedc3ea92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4027576
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Auto-Submit: Christoph Schlosser <cschlosser@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075105}