Mark css-pseudo/text-selection.html as passing.
This was marked as timing out on Windows as part of 0c3318c39
for some reason, even though that import didn't actually change
the test. It seems to consistently pass on the builders, so marking
it as timeout actually causes a flake.

Fixed: 1392080

Change-Id: I187aebfa5a07782bb48a313cf8d565cdc292218c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4050289
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075115}