[FilesTiering] Add skeleton class for DriveFsPinManager
This adds skeleton code for the DriveFsPinManager which will manage and
maintain the "bulk pinning" functionality. This also adds a new pref
that maintains the enablement of this feature, toggling it currently
does nothing, but later will serve as the mechanism to start and stop
the bulk pinning action.

Bug: b:259454320
Change-Id: I790775a01218493a3184cc0f3066540b4649631d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4036471
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Auto-Submit: Ben Reich <benreich@chromium.org>
Reviewed-by: Bo Majewski <majewski@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075127}