Load kAccount profiles into PDM.
This CL adds a feature flag kAutofillAccountProfilesUnionView to load
kAccount profiles from AutofillTable into the PDM. In the PDM, they are
stored in a separate vector account_profiles_ besides the existing
web_profiles_ (= kLocalOrSyncable profiles).

The implementation of PDM's CRUD operations is updated to work on the
correct vector, under the assumption that GUIDs are unique globally.

A new accessor GetProfilesFromSource(source) is added. It is used to
restrict suggestions, deduplications and imports to kLocalOrSyncable
profiles. Subsequent CLs will incorporate kAccount profiles there.

The implementation of SetProfiles() is simplified with a helper function
SetProfilesFromSource(source).

Bug: 1348294
Change-Id: I85da36dd1534192e245f3343be16912d34b964a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4032149
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1075145}