[GLImagePbufferBacking] Remove support for overlays
GLImagePbuffer cannot be presented in an overlay, as it is not of an
appropriate GLImage type [1]. For this reason, PbufferPictureBuffer does
not support overlays: PbufferPictureBuffer::AllowOverlay() returns
false. As a result, (a) GLImagePbufferBacking will never be asked to
produce an overlay, and (b) even if it was asked to do so, the
representation that it gives out in ProduceOverlay() would not result in
an overlay being presented due to [1].

This CL strips out all code from GLImagePbufferBacking that is related
to producing overlay representations.

[1] https://source.chromium.org/chromium/chromium/src/+/main:ui/gl/swap_chain_presenter.cc;l=1100;drc=603337a74bf04efd536b251a7f2b4eb44fe153a9;bpv=1;bpt=1

Bug: 1384438
Change-Id: Ib736c715cc0dcc9b85449a00d793c74101193957
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4037624
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075155}