ios: Fix EG test closeTabs in tearDown.
EG tests were closing tabs, but for tests where the following
run triggers a restart would not give enough time for all the tab
save to sync to disk. This was causing flake when the ending
test was loading a slow page, and the next test triggered a restart.
On restart, the app would try to restore the slow page, and the test
would fail to idle on app startup.

Change-Id: If85ccb99d7237dcc849c6261478fe4dd07bf7761
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4048290
Auto-Submit: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075156}