webauthn: filter sync devices on ID, not DeviceInfo timestamp.
DeviceInfo information is not valid forever: phones will reject
connections if the information is more than three weeks old.
Corresponding to this, desktops won't use information if it's more than
two weeks old. However, that was based on the modification time of the
DeviceInfo entity and, if you very rarely use Clank on a device, it's
possible that the DeviceInfo got updated, but updated caBLEV2 info
wasn't ready in time and thus that remains the same.

This change instead uses the fact that the pairing IDs in sync
information are actually timestamps to get the real time that the
caBLEv2 information was generated when the desktop is considering if
sync information is timely.

Bug: b/258189086
Change-Id: I6b47e9aa76615c4eaff483e59acaecdc16be2d91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4021688
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075204}