Migrating VA-API fake driver to Chrome.
Moving this code from ChromeOS code base to Chrome. It's moving from
ChromeOS for two main reasons:

1. It's easier to submit code to Chrome. This is because it has a less
flaky CQ since there is less real hardware involved.
2. The tests that need the fake driver are all in Chrome.

The largest edit made from the original code was changing the type of
the iterative value in some of the for loops from size_t to int. This
was because its stop condition was comparing it to an int, which was
throwing errors. Some smaller changes were: converted some for
loops to be range for lops, white space changes, converted a NULL to a
nullptr, adding extern "C" to __vaDriverInit_1_0(), and changed the
vendor string from "fake" to "libfake".

The //testing/ directory is edited to make ChromeCQ trybots run with
this fake driver instead of the old one in ChromeOS.

Link to original directory:
https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/platform/libva-fake-driver/

NOPRESUBMIT=true

Change-Id: Iffab7cc2c6cd793725818f5227523f2b1e9dc51d
Bug: b:259120742
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4024753
Reviewed-by: Pilar Molina Lopez <pmolinalopez@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Grayson LaFleur <graysonlafleur@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075231}