Restore a permission prompt on tab change.
This CL fixes a bug where we resolve a permission request if a user
opened a new tab and started editing the location bar view. In that case, when the user switched back to the first tab, the permission request will be automatically ignored.

Bug: 1379249,931657,1003747
Change-Id: I8cd9d573be9c9baba467f4be4bae3c41da103427
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3999327
Commit-Queue: Illia Klimov <elklm@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Ravjit Uppal <ravjit@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075251}