Add alternative_of to css_properties.json5
This CL makes it possible for the same CSS property to be represented
by multiple CSSProperty classes. The CSSProperty class in effect is
determined by runtime flags.

This is done by looking up the "alternative" property of a given "main"
property during UnresolvedCSSPropertyID, and returning the alternative
property if it's enabled.

In order to not leave the alternative property code unexercised, the
"-alternative-animation" property is introduced, which differs from the
main "animation" shorthand in that it also accepts "animation-timeline".
This is not terribly useful on its own, since the situation of
"an additional longhand behind a flag" could already be handled by
StylePropertyShorthand. However:

 * A subsequent CL will extend -alternative-animation in a way which
   can't be represented by the current runtime-flag-aware shorthand
   expansion.
 * The above mentioned expansion does not support trials, so using
   an alternative property is probably preferred anyway.

Bug: 1375994
Change-Id: I85f99a1e5eaf1e597010ac9585563d24d81c9ee2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4041147
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075253}