[Payments] Clean up some basic-card only test-support code
This CL removes a set of methods that are no longer called in
PaymentRequestDialogView and PaymentRequestBrowsertestBase. There are
still some related methods that can be cleaned up in
PaymentRequestBrowsertestBase, but the methods are still used in
compiled-but-disabled tests that need to be migrated away from
basic-card for us to remove the code.

Bug: 1209835
Change-Id: Iaf58909fa6fdb4427be611daeaa29dc81444ce79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4052440
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Auto-Submit: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075303}