Fix presentation callbacks in PresentationTimeRecorder
`PresentationTimeRecorder` uses presentation callbacks to determine when
a frame is presented on the screen. This class is mainly used to measure
latency of different UI interactions in ChromeOS and report them as UMA
histograms. Currently, the callbacks used in this class will be called
even if the next submitted frame fails to successfully present on the
screen. This CL fixes this by using callbacks that would be called when
a subsequent frame successfully gets on the screen.

Bug: 1378612
Change-Id: Ib76134905cbefa7d4ebf52d4a98c28bab25598f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4037545
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075305}