Visual Debugger UI tweaks
Text placed at the top of the frame can't be seen because it is drawn
outside of the canvas, so expand the drawing region with some padding.

Focus the annotation textbox when the filter dialog is opened, and save the filter and close the dialog when enter is pressed.

Bug: 1393118
Change-Id: Idb8718345265fccdaf58d3e210738af836666efd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4049764
Commit-Queue: River Gilhuly <rivr@chromium.org>
Reviewed-by: Peter McNeeley <petermcneeley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075356}