media/gpu/v4l2/test: Implement buffer management for VP8 decoder
This CL implements buffer managerment in CAPTURE queue for VP8 stateless
decoder. As there are differences in refreshing different reference
frames, this CL adds helper functions to determine if the buffer holding
the reference frame is no longer in use and is safe to be reused for
CAPTURE queue.

BUG=b:254902467, b:246647671
TEST=tast -verbose run dut video.PlatformDecoding.v4l2_stateless_vp8* on
asurada

Change-Id: Id98236725e7b4c0718cbafe7719245a9fd4f63be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4043596
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Steve Cho <stevecho@chromium.org>
Commit-Queue: Hang Nguyen <hnt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075403}