[Sticky-Header]Fix sort panel lag problem
We used to have this bug because we call the expand method on the real header, but we couldn't finish the animation on the sticky header. That’s why we could observe lag on our sticky header.

We fix this problem by setting the invisible header’s sort panel to visible/gone directly without the animation.

Before:
http://shortn/_Agzof9GY2S
After:
http://shortn/_K1ZTp8Wd1i

Bug: b:241271058
Change-Id: Ib5063d1a8d38465102bc9b3ad758d2dd5b8a1ad1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4053520
Commit-Queue: Cathy Wei <cathywei@google.com>
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Reviewed-by: Cathy Li <chili@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075407}