[Launch Lacros at login screen] Ensure Lacros blocks where expected
This commit is part of a series, meant to provide the capability
of pre-launching Lacros at login screen for performance reasons.
Design doc: http://go/lacros-loginscreen-design

Previous relevant CLs:
- [1] http://crrev/c/3793529 (Pre-launch Lacros (under flag))

Lacros's browser process can now block at login screen as soon as it
tries to access initialization parameters that are only available
post-login. For the purpose of clarity, we want that point to be
easily recognizable.

This commit ensure that the first access to post-login parameters
is correctly and clearly specified.

Bug: 1363933
Change-Id: I4de4c6d16f5aaf7e66b92607d82370dcd3b7665e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4045382
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Andrea Orru <andreaorru@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075444}