Make the InputMethodSettings mojom object extensible
This is an attempt to essentially reland
https://chromium-review.googlesource.com/c/chromium/src/+/4030016
incrementally.

The end goal of this change is to add the mojo API to fetch and send the
Japanese settings in decoder initalization inside of the
ConnectionFactory::ConnectToInputMethod function -
https://crsrc.org/c/chromeos/ash/services/ime/public/mojom/connection_factory.mojom;drc=5406e4a6f8cec5d7fcdae10b1c71f8efd7cd4be7;l=20

This CL avoids needing to create a new extensible InputMethodSettings and then
replacing the old struct with the new struct.

BUG=b/232341104

Change-Id: I8982e363b13af8198e88514e9e903934fe569ef1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4045844
Auto-Submit: Keith Lee <keithlee@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Commit-Queue: Keith Lee <keithlee@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075458}