Re-add lacros PowerSaveBlocker dependency on crosapi::PowerAsh
https://crrev.com/c/3234446 modified the lacros PowerSaveBlocker to
prevent screensaver suspension using the wayland idle inhibitor. The
previous behavior which depended on crosapi::PowerAsh was left as a
fallback.

https://crrev.com/c/3638668 removed the crosapi::PowerAsh fallback
behavior and all related dependencies from the lacros PowerSaveBlocker.

This CL restores the lacros PowerSaveBlocker to the version depending on
crosapi::PowerAsh, and it removes the wayland idle inhibitor dependency.

This change fixes an issue with the lacros PowerSaveBlocker where it
does not respect the provided WakeLockType. It's not currently possible
to support that behavior using the idle inhibitor protocol, so the
implementation needs to be reverted to using the PowerAsh API.

Bug: 260134089
Change-Id: Iab4e1dd7450a83dc29a2d911005b20b6f7f490d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4052103
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075520}