[GLImagePbufferBacking] Eliminate final calls into |image_| ivar
This CL removes the remaining calls that GLImagePbufferBacking makes
into its |image_| ivar. This ivar is a GLImagePbufferBacking, and all
of the methods called have stub implementations:

- BindTexImage() just returns true.
- ReleaseTexImage() does nothing.
- GetNativePixmap() returns nullptr.
- OnMemoryDump() does nothing.

We either eliminate these calls or replace them with their inlined
implementations as suitable, and collapse all related code in
GLImagePbufferBacking that can be simplified as a result. Two notes
with respect to this simplification:

- GLTextureImageBackingHelper::ScopedRestoreTexture and
gl::ScopedTextureBinder exactly undo the changes made in their
constructors within their destructors (as one would expect), so
eliminating the usage of these within scopes that are not accessing GL
is a no-op.
- With the above fact, the entire functionality of CopyOrBindIfNeeded()
can be collapsed down to marking |passthrough_texture_| as not pending
binding.

With this change, the only remaining purpose that
GLImagePbufferBacking's holding on to |image_| serves is to scope a
reference to the lifetime of the SharedImage instance.

Bug: 1384438
Change-Id: I58482ebddf59252d49b06314e7e58ea5b4e17b69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4036573
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075521}