[ReadingList] Fix ReadingListModelLoaded() notified too early
ReadingListModel should notify observers about the model being loaded
only after sync metadata is also loaded (ModelReadyToSync() invoked).
Otherwise, there is a theoretical risk that changes to the model during
this gap may never sync (in practice, changes shortly after browser
startup).

The issue is minor and unlikely to reproduce in the real world.

For the behavior to be testable, ordering of invocation needs to be
enforced in unit-tests. To achieve this, the previous home-grown mock
is replaced with gMock mocks, which also makes unit-tests more standard.

Change-Id: I71368a88a6d3ec3b793d27bd6c5199b6d9680717
Bug: 1386158
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055260
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075579}