Fix double move of JsSandboxIsolateCallback
Transition the callback to being held in a scoped_refptr, as it needs to
be shared between both the promise resolve and reject handlers.

Add tests for the asynchronous promise resolve and reject code paths. (The synchronous ones are already covered in existing tests.)

Bug: 1347094
Change-Id: If4429b890fb5e7643c1f99f848880c24b7cbcbac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055551
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Abhijith Nair <abhijithnair@chromium.org>
Commit-Queue: Ashley Newson <ashleynewson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075597}