Decouple lock screen online reauth UI from password sync
This cl extracts management of LockScreenStartReauthDialog from
InSessionPasswordSyncManager. Previously lock screen online reauth flow
always went through InSessionPasswordSyncManager::CreateAndShowDialog()
even though it could be caused by timeout policy unrelated to password
sync. Also InSessionPasswordSyncManager stored a unique pointer to
LockScreenStartReauthDialog creating an illusion that it owns it, even
though in fact the dialog triggers self-destruction from
LockScreenStartReauthDialog::OnDialogClosed method which is typical for
a SystemWebDialogDelegate. This cl makes it clear that
LockScreenStartReauthDialog is a singleton independent of
InSessionPasswordSyncManager.

This is a refactoring, not a behavioral change.

Bug: 1377014
Change-Id: I0c584d89ae776660b44721ba335e2fac98595af0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4051112
Reviewed-by: Roman Sorokin <rsorokin@google.com>
Reviewed-by: Tomas Gunnarsson <tommi@chromium.org>
Reviewed-by: Mohammed Abdon <mohammedabdon@chromium.org>
Commit-Queue: Andrey Davydov <andreydav@google.com>
Cr-Commit-Position: refs/heads/main@{#1075600}