webauthn: CONSTRAINT_ERROR from Play Services is not CREDENTIAL_EXCLUDED.
The mapping from CONSTRAINT_ERROR to CREDENTIAL_EXCLUDED has existed for
a long time, but it's not true of current Play Services and I'm not sure
that it ever was. Pertinently, Play Services has added passkey support,
which involves new cases that result in a CONSTRAINT_ERROR. But
CREDENTIAL_EXCLUDED is a meaningful error: it indicates to a site that a
registration was moot because a credential already exists. We shouldn't
return it unless that is really the case.

Change-Id: Ia77e0894c91b4aa799a9214d537b04eb4eaa3a91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4053723
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Auto-Submit: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075623}