[Idle Timeout] Add IdleTimeoutActions policy
Also, rename IdleProfileCloseTimeout to IdleTimeout.

The IdleTimeoutActions policy is in the same group as IdleTimeout. It's
a string-enum-list with the following values:

- close_browsers
- show_profile_picker

Update ActionRunner, so instead of the using a hardcoded
value (close_browsers + show_profile_picker) it picks up the policy/pref
value.

Bug: 1326685, 1316551
Change-Id: If68bc3dbf6d43ee5dc8f4c7a88b97702b6921d68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3996874
Reviewed-by: Owen Min <zmin@chromium.org>
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075629}