[KeyboardAccessory] Stop animator and restart with the initial position
Before this CL:
If KA is rerequested to be shown during an animation, the already
running animation is cancelled and the RecyclerView's current location
will be used as the final position.

After this CL:
If an animation is running, final position of the currently running
animation is used as the final position.


Change-Id: Ie106d1f1c32979bffeb54af28ca037eb7104f939
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055704
Commit-Queue: Adem Derinel <derinel@google.com>
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075715}