Fix extension set proxies in Lacros incognito
Incognito mode should not apply proxies set by extensions active in the
main Lacros profile.

This CL ensures that extension set proxies do not apply to Lacros incognito when an extension in the Lacros main profile controls the proxy.

Lacros incognito listens for changes to the system proxy via a mojo pipeline to Ash. When Ash receives a proxy config from a Lacros extension, it sends an update to all Lacros profiles (including incognito) that the *system proxy* has changed. This CL ensures that incognito ignores the update when the proxy config source is an extension.

Extensions running in Ash-the-browser do not propagate the extension proxy config to other profiles as a system setting.

Bug: 1378108
Change-Id: Ie1eef84f6d226c91d2c141fa744340ea6013f2bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4002973
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Andreea Costinas <acostinas@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Roland Bock <rbock@google.com>
Cr-Commit-Position: refs/heads/main@{#1075743}