Remove SVGElement::RebuildAllIncomingReferences
The intent of function is to notify dependent elements when an 'id' they
are targeting may need to be re-resolved. We are however relying on the
id target observers for handling this since quite a long time ago.

This gets rid of the only user of SvgAttributeChanged() outside of
attribute mutation paths.

Bug: 1392697
Change-Id: I23736a7e4ff37e5b5d1676d752531d3d52a60f69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4043147
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1075745}