[FPS][Android] Remove fps delete string and its usage
We introduced a new delete string in the First Party Sets context
(Page Info) that only make sense if the button deletion semantics was
extended to the whole FPS group. Since this is not the case this CL
remove that string and its usage.

Bug: 1349370
Change-Id: Ic9cc879886d14c05950882c0ee10b626b50f11ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055533
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Nicola Tommasi <tommasin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075754}