[FedCM] Introduce dedicated class for fetching manifest
This CL introduces FederatedManifestRequester which fetches the
manifests and manifest lists for all of the endpoints and calls the
callback when done. This class was introduced in order to correctly
notify (in a follow up CL) the metrics endpoints of all of the IDPs as
result of FederatedAuthRequestImpl::CancelTokenRequest() in the
multi-IDP case.

BUG=1361642

Change-Id: I38d0292b90ca77d50860d56284b7ae4262b04af5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4052080
Commit-Queue: Peter Kotwicz <pkotwicz@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075815}