Revive some autocomplete histograms
This CL moves autocomplete histograms (in the sense of autocomplete
for forms, not in the sense of omnibox autocomplete) next to autofill
histograms. They were overseen because they were in a different place.

The CL also revives two histograms that I would like to keep:
- Autocomplete.DaysSinceLastUse
- Autocomplete.Events
It removes code for expired and unused histograms:
- Autocomplete.Cleanup
- Autofill.AutocompleteQuery
- Autofill.AutocompleteSuggestions

The chante to validate_format.py is necessary because there is an
Omnibox related histogram with prefix "Autocomplete." in
others/histograms.xml.

Fixed: 1384720
Change-Id: I6714fc162445aa1dff83295696f16e0e67f95081
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4047295
Commit-Queue: Dominic Battré <battre@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Matthias Körber <koerber@google.com>
Cr-Commit-Position: refs/heads/main@{#1075850}