Files F2: Add the selection to the Store
Use requestAnimationFrame() to rate limit how often the selection
event is pushed to the store. This emulates the CHANGE_THROTTLED event
for the selection, but happens more often, rAF every ~16ms instead
every 250ms. The rAF frequency is enough to get started, we can fine
tune later if needed.

For the Store the goal is to only have the throttled event, the
non-throttled event will be only local to the file list/grid.

Bug: b:228112698
Change-Id: I8e66d708db530a3500473b4f4a0fc9dfa8540aef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4041772
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076037}