WebNN: Implement the synchronous compute method of MLGraph
This CL implements MLGraph::ComputeSync() that is called by MLContext to
implement the computeSync() method for the dedicated worker.
ComputeSync() has the similar function as ComputeAsync(). The difference
is if there are no validation errors, it calls ComputeSyncImpl()
implemented by an MLGraph backend, such as MLGraphXnnpack, and executes
the compiled platform graph synchronously in the caller's thread.

This CL also implements MLGraphTestBase::ComputeGraph() that helps test
the validation steps for both asynchronous and synchronous compute
methods.

Bug: 1273291
Change-Id: Ic19a83338af3865979ecae4ddfdea0ce39741a61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4050197
Reviewed-by: Jiewei Qian <qjw@chromium.org>
Commit-Queue: ningxin hu <ningxin.hu@intel.com>
Cr-Commit-Position: refs/heads/main@{#1076052}