Move all non-main Populate.*Document methods to NotRestoredReasonBuilder.
This is a no-op prefactor for https://crrev.com/c/4044207

They are on BackForwardCacheImpl currently but do not access anything
from that class.

Moving PopulateReasonsForMainDocument is a bit trickier as it does use some things from BackForwardCacheImpl. Also, the builder manages the recursive process and this is not part of that so, maybe shouldn't move. Since it's not needed for the follow-on CL, I am not doing it at this time.

Bug: 1382407
Change-Id: I77011b9820561a94e9980b44e8587def6c8db57a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4060369
Auto-Submit: Fergal Daly <fergal@chromium.org>
Reviewed-by: Yuzu Saijo <yuzus@chromium.org>
Commit-Queue: Fergal Daly <fergal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076074}