[mparch] Move FTN access to navigator.cc on creating WebBundleHandleTracker
This CL is a part of the effort to split RFH and FTN.
It introduces MaybeCreateWebBundleHandleTracker helper in
navigator.cc and moves FTN access to find the RFHI's parent or opener
from RFHI::MaybeCreateWebBundleHandleTracker() to that helper.

It removes WebBundleHandleTracker parameter from
Navigator::OnBeginNavigation() and uses
MaybeCreateWebBundleHandleTracker() to create the tracker.

Bug: 1179502
Change-Id: Ie21a093fbcb77e6557a3afc17ebdad3595181542
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4054766
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1076145}