[PA] Put ASAN BRP behind a buildflag
This CL trivially puts the declaration of ASAN BRP behind the
corresponding buildflag and asserts that ASAN BRP is only built with
Chromium. This is necessary because we intend to put `raw_ptr`
(including the implementation of ASAN BRP) inside the PA codebase, but
we do not intend to disentangle ASAN BRP from the rest of Chromium.

Bug: 1371729
Change-Id: I2e894bb6b0415f682a3dccb9ee17d028b7b70698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4060588
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Bartek Nowierski <bartekn@chromium.org>
Auto-Submit: Kalvin Lee <kdlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076174}