Re-enable some tests in TabSelectionEditorTest
The test disabled on Android M is safe to re-enable as Android M support
is deprecated and we replaced the LayerDrawable that was causing the
issue in the first place.

The test that is disabled was marked safe to re-enable by test reviver.
If it flakes again we can just re-disable and look for the root cause.

Bug: 1237368, 1374370
Change-Id: I4874add22fe38652c0ccb5cd09afab0c00ec31d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4051961
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Commit-Queue: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076188}