Add a browser test of ContentMainDelegate entry points
This tests that ContentMainRunnerImpl invokes the ContentMainDelegate
entry points in the expected order, with the expected parameters.

Also adds some entry points that were missed in Android browser
tests.

This is mainly a base for several followup CL's that will test more
behaviours:

* Returning true or false from ShouldCreateFeatureList and
  ShouldInitializeMojo
* Returning early exit codes from PreBrowserMain,
  PostEarlyInitialization, etc.

This is intended to document and verify the semantics of
ContentMainDelegate which can be unclear, and exercise edge
behaviours of the delegate. This will help to catch cases where
implementations of ContentMainDelegate add hidden assumptions or
break existing assumptions.

Bug: 1340599
Change-Id: Ifd52f0370a80cb292ecf8540dbdee82d345166b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3967194
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076243}