[power_sampler] Add "sample age" field to battery sampler.
This field contains the time delta between when power_sampler emits
the sample and when macOS read the battery state. Since macOS reports
the time at which it read the battery state in seconds since epoch,
the resolution of this new field cannot be more precise than "seconds".
When samples are emitted on IOPMPowerSource notification, we expect
the "sample age" to be 0 or 1 second.

Change-Id: I4bf572c48fe3507f61ffdf2b466cf45bedfd7bbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4054584
Reviewed-by: Olivier Li <olivierli@chromium.org>
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076321}