Cleanup XR Session Shutdown paths
Attempgs to cleanup/reorganize some of the VR Shutdown paths and make
things a little bit clearer. Specifically:

1) Removes on_presentation_ended from the CompositorBase::RequestSession
interface. The only remaining caller was openxr, which simply passed a
base::DoNothing. It's possible that future refactors to OpenXr::
ShutdownSession will want a callback at the end; but this is better
handled by passing such a callback to ExitPresent directly.

2) Removes VrDeviceBase::ShutdownSession. This forces children to
implement this method directly, rather than relying on the default
backwards compatible, but not necessarily precise behavior.

3) Implements ShutdownSession for VrOrientationDevice, this is still
just an immediate return of the session; but there's nothing really
blocking on the device side if an inline session is still outstanding
like there is for an immersive session.

4) Implements ShutdownSession for OpenXRDevice. It still returns a bit
too early; but it at least now actually kicks off a shutdown rather than
waiting for the PresentationProvider pipe to be disconnected; which is
the current way that an OpenXRDevice gets notified of a shutdown. This
introduces a little bit of refactoring to share code with the existing
shutdown path.

5) Adds an "already exited" check to OpenXrRenderLoop to avoid multiple
calls. All of the operations should be resilient to multiple calls, but
it saves a little cognitive overhead for developers.

Future work in this area should address not calling the ShutdownSession
callback in OpenXRDevice until the Session is fully shutdown. It should
be possible to further simplify the session shutdown flow by removing

all the way in to the device; which could then mimic the "OnExitPresent"
call by disconnecting the XRSessionClient. This has the added benefit of
remvoing an unneeded hope for the VisibilityStateChanged method.

XrRuntimeEventListener: :OnExitPresent and passing the XRSessionClient
Change-Id: I7df95eda6d30668187cca1d5827086282bd4eda5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4054022
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076371}