Drop call to HasIntrinsicSize() in StyleImage::ApplyZoom
It should always be true for the cases where it's used, and it will
probably be better to have a caller check a predicate like this if it
ought to be skipped. This will allow improving the correctness of
StyleImage's HasIntrinsicSize() to fix other uses.

Since this removes that last reference to |this| from said function,
make it static. Do some minor clean up.

Bug: 935057
Change-Id: Iea513d3ebdbd11200566d78e4120845ddf95c869
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4035241
Auto-Submit: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076383}