Pass the list of configuration options for Cronet engine across JNI boundary as a proto.
With us unwrapping the experimental options into separate fields the
number of parameters being passed from the builder is about to grow
significantly, and this should help Cronet developers maintain sanity
when modifying the list.

There's no ideal solution to the problem - having a single method
with 50 parameters is a nightmare to maintain, accessing fields of
objects across languages is verbose, and protos require serialization
across language boundary. That being said, there's no danger of verion
skew (it's all the same implementation target), the de/serialization
is tightly contained, and the cost of serialization of the proto is tiny
in context of creating a new Cronet engine.

Bug: 1393287

Change-Id: I0d7468cece65c474bef8ec5fafd557dd9a2aa426
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4051188
Reviewed-by: Stefano Duo <stefanoduo@google.com>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Dan Stahr <danstahr@google.com>
Cr-Commit-Position: refs/heads/main@{#1076391}