locks: Migrate DOMException to `RejectWith-`
When throwing an exception `e`, it's important to make a stack trace
available as `e.stack` as long as possible.  However, DOMException's
constructor doesn't create this `stack` property, so the direct use of
DOMException's constructor is highly discouraged. So this patch migrates
`DOMException` constructor to `ScriptPromiseResolver::RejectWith-`.

Bug: 991544
Change-Id: I15f9d77a8364cf7e91665784619359f1c24ccadc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055212
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Jinho Bang <zino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076399}