MojoIpcz: Eliminate DataPipe deadlocks
This eliminates unsafe DataPipe and MojoTrap re-entrancy caused by
flush operations on a DataPipe while the DataPipe's or MojoTrap's
internal lock is held.

The general fix is to ensure such locks are released before issuing
flush operations, and in some cases to forego flushing in favor of
basic queries of the pipe's state.

Fixed: 1386093
Change-Id: I99c865c8b2282bbfcf3c02301b6971744ba95953
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4052941
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#1076436}