Record the last completed state of the ParentAccessDialog to a histogram
The ParentAccessStateTracker stores the last encountered screen and
records it to two histograms-- a general Parent Access histogram, and a
flow-specific histogram (ie web approvals). Metrics are only recorded
for instances of the ParentAccess webUI that correspond to a dialog
(i.e. not accessing chrome://parent-access directly in the browser).

Also rename the FaviconAvailability histogram to follow the naming
conventions for families histograms. This is an unlaunched feature so
very little data is stored in the old histogram name.

Bug: b:258722110
Change-Id: I2baa06938d4a1b3c65bcbf9f1cdafb82d321d2bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4029663
Reviewed-by: Dan S <danan@chromium.org>
Commit-Queue: Courtney Wong <courtneywong@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076451}