Break up presubmits in tools\perf\PRESUBMIT.py
CheckChangeOnCommit from this file is the longest presubmit function on
Linux in all of the Chromium repo by a significant margin (~460 s,
second slowest is ~90 s). This presubmit function calls _CommonChecks
which was doing seven different presubmits. With those split out we can
more easily see where the time is going.

Testing with this change makes it clear that CheckWprShaFiles accounts
for the vast majority of the time and the bug is for tracking attempts
to make it run faster.

Bug: 1393724
Change-Id: Idbc4dc6dd7156e5083d38272e42f8bf6e49d949b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4058427
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076487}