[DIPS] Add new feature param to control what action triggers DIPS.
We want to determine what behavior should trigger clearing if a site doesn't have recent user interaction. Our options are (1) the site using storage, (2) site redirecting the user quickly (bounce), or (3) the site bouncing the user while writing to storage.

This is being added as a feature param so that we can make experiment
groups to compare the effects of these actions.

Bug: 1375302
Change-Id: I42f98f3ca28943463a12e9792d12d46763c32b45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4030637
Commit-Queue: Ryan Tarpine <rtarpine@chromium.org>
Reviewed-by: Joshua Hood <jdh@chromium.org>
Reviewed-by: Ryan Tarpine <rtarpine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076490}