Metrics for the selection dialog being accepted with a CVC challenge
This CL adds metrics for the CardUnmaskAuthenticationSelectionDialog
being accepted with a CVC challenge option being chosen. These metrics
are slightly different from the SMS OTP metrics due to the fact that the
SMS OTP metrics depend on the server response from a
SelectChallengeOption call, which a CVC challenge option does not make
due to not needing to generate an OTP in the backend to send to the
user. CVC challenge options directly render the CVC input dialog, so we
must add metric logging for this new case.

Bug: 1370329, 1394173
Change-Id: Ib60aac1263feb541a6e590c044d11034be49d266
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4049897
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076507}