Update URL list for fenced frame benchmark
The ad urls currently being used for the benchmark tests do not serve
the same page between iframes and fenced frames. For a lot of them, they
load a static image for iframes but a dynamic animated page for fenced
frames.

I did an audit of a bunch of possible rich media gallery ads and picked
6 that load consistently the same way between iframes and fenced frames.

Change-Id: Ieeadb11bc1063f55d1130e98dfd0e4161dbbf4e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4053022
Commit-Queue: Liam Brady <lbrady@google.com>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076517}