Make fallback app launches work for PWAs on Mac.
On Mac OS, when an app shim fails to load chrome, it will fall back to
trying to launch chrome passing --app-id on the command line instead.
For legacy packaged apps, this would also include the profile to use on
the command line, however for PWAs the profile is unknown at this point,
so an empty string is passed. To not crash in this case, add logic to
GetStartupProfilePath to determine an appropriate profile to use for
these launches.

Since these cases should be extremely rare, it is not important that
this new logic exactly matches the normal profile selection logic in
AppShimManagerLoadAndLaunchApp. Instead simply pick an arbitrary
profile where the passed in app-id is valid.

Bug: 1379587
Include-Ci-Only-Tests: true
Change-Id: I54526dccb148e43c38cd460aeddcae2cfd21de11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4035927
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076581}