Implement keyboard shortcut for Lens Region Search.
Keyboard shortcut (Ctrl/Cmd + Shift + L) is temporary as the keyboard
shortcut has not yet been finalized. All code is flag-guarded so
keyboard behavior is unaffected when flag is disabled.

Tested on Linux, Mac, and Windows; with flag disabled and enabled; and
with search engine set to Google, Bing (third-party search engine that
supports region search), and Yahoo (does not support region search).
Availability of keyboard shortcut is identical to that of region search
when flag is enabled.

Screenshot of context menu item with keyboard shortcut label:
https://screenshot.googleplex.com/3MSyZ2kn9iGHDef

Bug: 1392456
Change-Id: Ifa4fb6297879d9ae4de3c0d9dc970c4878e52c04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4043546
Reviewed-by: Juan Mojica <juanmojica@google.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Bryan Nguyen <nguyenbryan@google.com>
Cr-Commit-Position: refs/heads/main@{#1076590}