arc: Add [Default] = None to a mojo enum
Mojo::TextInputType, which will be extended in crrev.com/c/4054753,
does not have a default value, so that might cause the crash of
the ARC windows. To prevent it, we introduce the default value to the enum first.

BUG: b:247962953
TEST: Cherry-pick ag/20559780 to rvc-arc and confirm that
you can use Android apps without crash.

Change-Id: Ib5a364ad2728cdecbfe7efe8e57bab4828969811
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4056676
Commit-Queue: Koki Ryu <kokiryu@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076614}