mojo_service_manager: Enable browser test
Use browser tests to verify the browser can be initialized with the
chromeos mojo service manager. If the browser test flag is set, the
fake service manager will be initialized. This condition will be removed
after removing the kAshUseCrOSMojoServiceManager switch.
The checking logic of kDisableMojoBroker is moved to
CreateConnectionAndPassCloser(). In real browser, this flag must be set
to use service manager. In browser tests, this flag must not be set so
the browser asks as the broker since we don't have an external broker
in browser tests.

Bug: b:249897521

LOW_COVERAGE_REASON=connection_helper.cc contains real impl which
shouldn't be used in tests.

Change-Id: I58865e4599e9a209d88133bb9174eb2912af1061
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3915810
Commit-Queue: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076690}