Introduce AudioGlitchInfo and add it to AudioConverter::ProvideInput()
This is the first step in a larger effort to plumb AudioGlitchInfo
through the whole Audio pipeline. We have split this effort into
several CLs to reduce the scope of each one.

Glitch info received by the AudioConverter needs to be delivered to
the audio converter inputs. This is done in the next CL.
This CL only adds AudioGlitchInfo into
AudioConverter::InputCallback::InputCallback(), so that we touch all
InputCallback implementations, some of which do not care about
AudioGlitchInfo. We considered other options of plumbing glitch info
only into InputCallback implementations which are interested in it, but
they turned out more complicated and more difficult to maintain. Here
is an attempt to template the AudioConverter to allow for arbitrary
types of callbacks, which ended up too complicated:
https://chromium-review.googlesource.com/c/chromium/src/+/4026183/41/media/base/audio_converter.h#49


The next CL is
https://chromium-review.googlesource.com/c/chromium/src/+/4051263

Bug: 1384499
Change-Id: I906fd378581783b39d3d85679b50bcb8acafd53c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4051143
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Reviewed-by: Yuchen Liu <yucliu@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Fredrik Hernqvist <fhernqvist@google.com>
Cr-Commit-Position: refs/heads/main@{#1076720}