[GPU] Change GLES2 cmd decoder tests to call BindImage() as in prod
gles2_cmd_decoder_unittest.cc invokes GLES2CommandDecoder::BindImage(),
passing both true and false for the value of the |can_bind_to_sampler|.
In production, |false| is always passed on Mac/Win and |true| is always
passed on other platforms. I am currently in the process of hardening
this invariant in the code, as it aids in analysis of these codepaths
and ensures that we know exactly where the flows that can result in
decoders doing lazy binding are occurring.

This CL changes gles2_cmd_decoder_unittest.cc to invoke BindImage() as
is done in prod on each platform and makes necessary adjustments to
test expectations.

Note that it looks like parameterization can now be removed entirely
from this test file. We leave that as a followup to avoid introducing
noise into this CL.

Bug: 1323341
Change-Id: I3e06af006153876f39fb2be5e319aedc7fc49566
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4061726
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076727}