Cookies Tree Model: Correctly count Quota nodes for size
The Quota node type did not correctly implement the InclusiveSize
function, leading to it incorrectly reporting 0 size and be excluded
from UI display.

The Quota node type is only used when the experimental partitioned
storage flag is enabled. It is hoped that the Cookies Tree Model will
not need to represent partitioned storage in production code, and the
new BrowsingDataModel will instead. Whether this will happen is not
immediately clear, and so the CookiesTreeModel should still be fixed.

Bug: 1382940
Change-Id: I63e87c6b00b7a5fbcab2444466d0c3a94d2b88da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4061547
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/main@{#1076747}