Prepare tests for Chrome side owner key generation
In the upcoming change Chrome will start establishing ownership on
ChromeOS, but some tests are not prepared to emulate that properly.
They would need to make sure that all the relevant fake/mock/stub
classes generate public owner key, install attributes, local state,
device policies and other files. Tests with PRE_ steps would need to
ensure that all the relevant files are persisted on disk in between
tests.

The current implementation of the tests in this CL is not emulating
the ownership logic realistically enough and it doesn't seem to
be essential for the logic under test.
Instead skip to the state when the consumer ownership is already
established to ensure that the tests continue working with the
upcoming ownership changes (without requiring addition setups).


Bug: b:244407123
Test: ParentPermissionDialogViewTest*
Change-Id: I93afb8956658a906b42891e6ae743344b160841f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4061452
Auto-Submit: Michael Ershov <miersh@google.com>
Commit-Queue: Michael Ershov <miersh@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076770}