[Enrollment] Loose VPD check for FRE.
Based on crrev.com/c/chromium/src/+/3779583/comment/64f7e768_9aaf38d2.

Before the change, AutoEnrollmentTypeChecker triggered inescapable FRE
check when RO_VPD and RW_VPD were invalid simultaneously, whilst
triggering escapable FRE when only RO_VPD was invalid.

When RO_VPD (or whole VPD) is invalid, state keys are not available,
and re-enrollment always fails. If it is enforced and inescapable, the
device is stuck on FRE, and a user cannot proceed with any other flow.

With this change, when both RO_VPD and RW_VPD are invalid, require
re-enrollment but do not enforce it so that it can be cancelled in case
of an error.

Bug: b:213325251
Test: unit_tests --gtest_filter=AutoEnrollmentTypeCheckerTest.*
Change-Id: I8a63a9a4017861c9b2d0a5712bcae6fe3a3d6034
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4051064
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Roland Bock <rbock@google.com>
Commit-Queue: Artem Sumaneev <asumaneev@google.com>
Cr-Commit-Position: refs/heads/main@{#1076771}