Use functions for histograms with dynamic strings in Android accessibility
The Android accessibility code tracks metrics of the running services
and associated AccessibilityServiceInfo objects. In a recent update
(crrev.com/c/3945884), we expanded these metrics and in doing so we
added dynamic strings to an existing macro. This is not allowed and
was intermittently causing failures on debug builds.

With this CL we update to call the histogram functions directly as
recommended in the metrics docs.

AX-Relnotes: N/A
Bug: N/A
Change-Id: Ia81c409340d254580c68c49f8815b3b1096e02ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4043867
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Amanda Lin Dietz <aldietz@google.com>
Auto-Submit: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1076919}