Share attribution-source JSON parsing between blink and content
To accomplish this, we add Mojo typemaps for the relevant C++ types
in //components/attribution_reporting, and change blink to call into
the top-level Parse method there.

This allows us to remove redundant JSON-schema-parsing implementations,
fuzzers, and unit tests, and allows us to centralize validation in the
relevant //components/attribution_reporting types themselves, with Mojo
typemapping taking the place of a number of explicit checks in
attribution_data_host_manager_impl.cc.

Finally, in addition to the simplifications that benefit us now, having
a single schema parser will make it easier to add features to the
Attribution Reporting API in the future and to address
crbug.com/1374121 by relocating all the parsing C++ and Mojo code
to //services/network.

Bug: 1379009, 1374121
Change-Id: I618c3f5ebb76183daf72a0e4b16c1efb21fd7654
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4008661
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Nan Lin <linnan@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076979}