Add PdfLocalFileAccessAllowedForDomains Policy
Enterprise customers want to be able to access file:// protocol links.
For security reasons, this should be a policy that enterprise can set
for specific domains. Add a new policy that grants an allowlist of
domains access to file:// links from the PDF Viewer.

Bug: 1302465
Change-Id: Ibbb5cfa61c795b9bce949c3ca8a17e31fa3fc5bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4003999
Commit-Queue: Andy Phan <andyphan@chromium.org>
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076995}