[Test Automation] Kombucha 1.5 verbs for WebContents instrumentation.
It is common to create a Do(...) step in order to instrument an existing
or new tab, so this CL provides verbs to do that.

It also provides a one-stop "open a new tab to this URL and instrument
it" verb, with the option to specify a target browser other than the
current one.

Tests are provided for all new functionality. Documentation is updated
and reorganized for clarity.

Bug: 1374319
Change-Id: If14549390c63183b3b64f2db0eec3ca90b1ddd40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4039422
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077006}