Fix JSTranslateWebFrameManager DCHECK
JSTranslateWebFrameManager can't properly track whether or not the
script has been injected. For example, the script may have been injected
but the WebFrame may not know that if the page was loaded from the WebKit
page cache.

Instead of tracking if the script was injected, the scripts have been
updated to always perform the necessary action or fail gracefully.

Fixed: 1383578
Change-Id: I9f1a3f45650b4e9a473a79127ce1a19b210b5a85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4053584
Reviewed-by: John Wu <jzw@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077056}