Prefer WGC Capturer to embed the cursor
An option was recently added as false by default to embed the cursor in
the capture frame. (Note that Desktop Frames have long had the ability
to indicate to the Cursor And Frame composer if the frame might already
have the cursor). However, the WGC capturer has this set as true by
default and it needs to be configured off.

This change sets the appropriate bool in Chrome to turn on the embedded
cursor, which is the default/pre-existing behavior before another change
lands in WebRTC to actually respect this bool and disable the cursor for
WGC when requested.

Fixed: 1393140
Change-Id: If0af1a9b1e7e170ecd208036e4ecfdcd77de454e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4061293
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077074}