[sheriff] Disable testControllerShowsBottomSheet
This is failing on
https://ci.chromium.org/p/chrome/builders/ci/android-arm64-official-tests.

Suspect
https://chromium-review.googlesource.com/c/chromium/src/+/4061587/15,
but not reverting that due to a chain of CLs that landed on top of it.

Suspecting only tests are broken, but please refer to the linked CL if
this is not the case.

Bug: None
Change-Id: If1a46562a085e52debb06fb594d5c420514ea5bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4063870
Owners-Override: Tommy Li <tommycli@chromium.org>
Commit-Queue: Thiabaud Engelbrecht <thiabaud@google.com>
Reviewed-by: Thiabaud Engelbrecht <thiabaud@google.com>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077086}