[Payments] Move PopulateSheetHeaderView into PaymentRequestSheetController
This CL moves PopulateSheetHeaderView from a utility function to
a method on PaymentRequestSheetController, as a first step to
implementing a new header for PaymentHandler (see bug).

The CL should be a pure refactor in terms of behavior, but the
method signature changed to move various method calls (e.g.,
to ShouldShowHeaderBackArrow()) to within PopulateSheetHeaderView
now that it has access to the PaymentRequestSheetController directly.

Bug: 1385136
Change-Id: I1c064c433de7911d3e7139f62d8b6e5f4a70512a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4024643
Auto-Submit: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Nick Burris <nburris@chromium.org>
Commit-Queue: Nick Burris <nburris@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077105}