mac: Pad DCHECK x86_64 main executable by 44kB instead of 56kB
Switching between DCHECK and non-DCHECK canaries on arm64 triggers bug
1295098 (the running Chrome is unable to open file open / save dialogs).

This is because DCHECK builds are larger which causes the internal
geometry of the universal binary to change.

This CL reduces the size of the x86_64 artificial padding of DCHECK
builds to stabilize the main executable geometry, bringing it inline
with non-DCHECK builds. See the associated bug for the calculation
details.

Bug: 1394196
Change-Id: I88eb6dcf40ccd96f8304bc37240ac502d6332e79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4062276
Auto-Submit: Tom Burgin <bur@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Mark Mentovai <mark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077107}