float: Add metrics to track how long float session lasts
In most cases, the session is the duration of the `FloatedWindowInfo`
object. The one exception is when there is a floated window, and we
change desks; the object persists in this case. We don't want to record
though, otherwise a floated window on an inactive desk could be there
indefinitely.

We stop recording a floated window if it is no longer on the active
desk, and start recording again if we move to a desk with a floated
window.

Test: ash_unittests WindowFloatMetricsTest.FloatWindowDuration
Bug: b/258473579
Change-Id: I9e037e23920d8d7c2cd9453ddba9e3e8da971436
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4019123
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Gavin Williams <gavinwill@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077150}