Fix device scale factor being applied to viewport client hints twice.
When getting the viewport size from RenderWidgetHostView, the device
scale factor was already implictly applied, so when it was explictly
scaled by this factor, it was being applied twice. This CL fixes this
bug and adds a histogram to estimate the impact this bug had.

Bug: 1394231
Change-Id: Ib4df4fc3b04d93787791069309d42d5d435b6a1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4064089
Commit-Queue: Max Curran <curranmax@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077151}