Add zero back to ppapi/PRESUBMIT.py
While looking at the presubmit --all bot's output for easy fixes I
noticed a huge number of errors of this form - 689 in total as it turns
out:

  PPAPI Interface modified without updating NaCl SDK.
  (note that some dev interfaces should not be added the NaCl SDK; when in doubt, ask a ppapi OWNER.
  To ignore a file, add it to IGNORED_FILES in native_client_sdk/src/build_tools/verify_ppapi.py)

  usage: verify_ppapi.py [-h] sources [sources ...]
  verify_ppapi.py: error: the following arguments are required: sources

I focused on the first paragraph, which was unfortunate because the
second paragraph was trying to give me a hint. I couldn't reproduce the
errors on Windows so I switched to Linux, and eventually it all made
sense.

Change crrev.com/c/3810377 (which I reviewed) deleted the zero from
this line:
  range(0, len(nacl_sdk_files), files_per_command):
This changed it to:
  range(len(nacl_sdk_files), files_per_command):

Unfortunately that causes the checks to be skipped on Windows, and to
run as NOPs on Linux, roughly ~700 times, because nacl_sdk_files has
about 300 entries, and files_per_command is 25 on Windows and 1,000
elsewhere.

Easy fix. Put the zero back.

Bug: 1222512
Change-Id: Ie2aad6f4130089547e6cde0f9c6291e44bc106c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4065121
Reviewed-by: Derek Schuff <dschuff@chromium.org>
Auto-Submit: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Derek Schuff <dschuff@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077236}