Prerender: Simplify initialization of PrerenderHostRegistryTest
This is a refactoring CL and doesn't change functional behavior.

A super class of PrerenderHostRegistryTest (i.e.
RenderViewHostTestHarness) has TestWebContents, so
PrerenderHostRegistryTest doesn't need to create it. This CL removes
creation of TestWebContents and moves common routines to the test
harness for code simplicity.

Bug: n/a
Change-Id: I480568df403a26230a7655b1e9035402433d9f63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4063448
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077253}