Fix sequence check failure in GpuArcVideoEncodeAccelerator::Encode
Mojo callbacks are supposed to be called on the same sequence as where
the Mojo receiver is bound to. In GpuArcVideoEncodeAccelerator::Encode,
we register the callback to be run on VideoFrame destruction which may
happen on a different thread.

This change fixes the issue by posting the callback back to the
receiving thread.

Bug: b:260438825
Test: android.mediastress.cts.MediaRecorderStressTest#testStressRecordVideoAndPlayback
Change-Id: Ie3a4d5fc8454b4c242bc1eaea3999010295a89de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4065439
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Shao-Chuan Lee <shaochuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077268}