[Office upload] Open doc URLs in the web app directly
Override part of the decision process for web app link handling using
the existing ChromeOsWebAppExperiments class.

We add a special case where we open URLs in the web app if they match
the expected office app ID and the URL matches the extended scope,
and the previous URL is empty. This is the case when we open URLs
for docs in a new tab, from the ODFS extension.

We'll change how we do this in the future rather than override
this part of the app launch process, so consider this short-term
solution to demonstrate what is possible.

Bug: b:246462203
Change-Id: Id3365e3b54642c4731b17b96b6127510b826bc63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4060986
Reviewed-by: Glen Robertson <glenrob@chromium.org>
Reviewed-by: Maggie Cai <mxcai@chromium.org>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077297}