History Sync: Delete foreign visits on startup if feature disabled
If the feature syncer::kSyncEnableHistoryDataType is disabled, there
shouldn't be any foreign visits in the history DB. This CL makes sure
this is the case: On startup, if the feature is disabled, any existing
foreign visits get cleared. (They can exist if the feature was
previously enabled.)
In order to avoid a perf impact on startup, a new flag is introduced in
the meta table, "MayContainForeignVisits". It's set to true whenever a
foreign visit is added. The deletion on startup is only triggered if
this flag is true.

Bug: 1347733
Change-Id: Ifabe499ff87c2716b213508ad6008f08eefc8c5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4051080
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077324}