[//media/gpu] Hardcode image bind as decoder/client managed per-platform
//media/gpu video decoder code threads through a callback via which
various platform-specific decoders bind a GLImage. This callback takes
in a boolean parameter that specifies whether the client manages the
image (in which case the GPU service internals do not need to on-demand
bind the GLImage to the texture) or the decoder manages the image (in
which case the GPU service internals *do* need to on-demand bind the
GLImage to the texture).

As part of eliminating the GLES decoders needing to perform lazy
binding, I am looking to isolate where and on which platforms there is
still potential for such lazy binding. It turns out that the division
between decoder- and client-managed images is on a per-platform basis:

- Windows and Mac always specify that the image should be
  decoder-managed
- Other platforms always specify that the image should be client-managed

This CL removes the boolean parameter from //media/gpu's
BindGLImageCallback, instead hardcoding its semantics to do either
decoder-managed binding (on Win/Mac) or client-managed binding (on all
other platforms). This makes it much easier to reason about when this
flow has (and doesn't have) the potential to result in lazy binding.

There is no functional change, as this CL is merely codifying existing
per-platform differences.

Change-Id: I96e34113889c70f509ffbc7000dbe413ba37e61a
Bug: 1323341
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4051781
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077329}