Record a BFCache-blocking reason when the "Authorization" header is used.
If the page has the "Cache-Control: no-store" header then we check if any of the frames in the main frame's origin have been seen making a network request using the "Authorization" header. This indicates that sensitive information may be present in the page.

For now the purpose is to measure how often this occurs. In the future, if pages with CCNS are allowed to enter BFCache, this reason may become a blocker.

Bug: 1382407
Change-Id: I10f48985132a6f677ca8ead6f7962d3958de5a55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4044207
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Yuzu Saijo <yuzus@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Carlos Caballero Grolimund <carlscab@google.com>
Commit-Queue: Fergal Daly <fergal@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Auto-Submit: Fergal Daly <fergal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077361}