[Signin] Update ConsentLevel in test utility functions
This CL updates some 'identity_test_utils' functions to work on the sign
in consent level instead of sync. 'ClearPrimaryAccount' only blocks
until 'OnPrimaryAccountChanged' notification is fired if there is a sync
account. This is updated to block if there is a primary account
regardless of the consent level. Also,
|RemoveRefreshTokenForPrimaryAccount| is updated to remove the refresh
token of the primary account regardless of the consent level.

Change-Id: I74a7c0f6f469ed61b61d85fd09c1caf2f323173d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4065682
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077376}