[mparch] Add DidChangeReferrerPolicy() to NCI and RFHO
This CL is a part of the effort to split RFH and FTN.
It adds DidChangeReferrerPolicy() to NavigationControllerImpl and
RenderFrameHostOwner to avoid access to FrameTreeNode from
RenderFrameHostImpl::DidChangeReferrerPolicy(). Changing whether to
protect a URL in the navigation API is moved to
NCI::DidChangeReferrerPolicy().

Bug: 1350575
Change-Id: I9f6d106fb85ea0d3f44eddd276edce4f9c6a2ccd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4019055
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077377}