Privacy Sandbox Settings: Add logic to disable ps4 for restricted accounts
This CL couples out the common general code where the APIs should be restricted to a separate method. This method would be the central location which the other K-APIs would hook into to check whether they
can be turned on or not, on top of the existing IsSiteDataAllowed.

Bug: 1378703
Change-Id: Id7dbdb171f827540c33fa4232eb7b51904d99d69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4061556
Reviewed-by: Olesia Marukhno <olesiamarukhno@google.com>
Commit-Queue: Rohit Agarwal <roagarwal@chromium.org>
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/main@{#1077393}