History Sync: Ensure to DeleteAllForeignVisits() when disabling Sync
When Sync (or the History data type in particular) is turned off, all
foreign history should be deleted locally.

Usually, the signal to do this should arrive in HistorySyncBridge via
an ApplyStopSyncChanges() call with a non-null MetadataChangeList.
However, due to crbug.com/1383912, this signal currently doesn't
always arrive.

This CL implements a workaround: Calling DeleteAllForeignVisits() when
Sync's TransportState gets set to DISABLED. This results in some
unnecessary DeleteAllForeignVisits() calls, but these are cheap if
there are no foreign visits in the DB.
This CL also adds integration tests to cover the disabling-Sync case.

Bug: 1383912, 1347733
Change-Id: If30c6e32182176f3df1c9f02ff1af403ec6cd766
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4051148
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077417}