[ReadingList] Refactor ReadingListSyncBridgeDelegate::StoreLoaded()
No behavioral changes: pure refactoring, the patch just removes
ReadingListSyncBridgeDelegate::StoreLoaded() and leans on a completion
callback instead.

This function doesn't really fit ReadingListSyncBridgeDelegate as it was
the only case where it's about the underlying storage (having been
initialized), as opposed to all other functions that propagate changes
received via Sync to the model.

As this function is essentially a completion event for an
owner-initiated SetReadingListModel(), it can be replaced with a
OnceCallback. While doing so, it is also split into a dedicated
function, ReadingListModelStorage::Load(), in anticipation of the sync
bridge eventually getting decoupled from the underlying storage.

Some tests abused the public exposure of
ReadingListModelImpl::StoreLoaded() to inject the initial list of
entries, pretending it was the storage layer doing so. This was
questionable, as ReadingListModelImpl was already loaded at the time,
and hard to achieve after this patch (where the abused function becomes
private). Instead, tests should adopt a fake storage that can deliver
the desired initial state, which is now done with the newly-introduced
FakeReadingListModelStorage.

Change-Id: Ib72ac19f94a1486e5cf828d36e3ab19b02777f1a
Bug: 1386158
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4042184
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077419}