[DecoderContext] Hardcode BindImage() as decoder/client managed per-OS
DecoderContext::BindImage() takes in a boolean parameter that specifies
whether the client manages the image (in which case the GPU service
internals do not need to on-demand bind the GLImage to the texture) or
the decoder manages the image (in which case the GPU service internals
*do* need to on-demand bind the GLImage to the texture).

As part of eliminating the GLES decoders needing to perform lazy
binding, I am looking to isolate where and on which platforms there is
still potential for such lazy binding. It turns out that the division
between decoder- and client-managed images is on a per-platform basis:

- Windows and Mac always specify that the image should be
  decoder-managed
- Other platforms always specify that the image should be client-managed

This CL removes the boolean parameter from DecoderContext::BindImage(),
separating it into two variant methods that hardcode semantics to either
(1) set the image as needing binding by the decoder (on
Win/Mac) or (2) set the image as not needing binding (on all other
platforms). This makes it much easier to reason about when this flow has
(and doesn't have) the potential to result in lazy binding.

There is no functional change to production behavior, as this CL is
merely codifying existing per-platform differences. There is one
functional change in gl_oes_egl_image_unittest.cc in code that is
compiled only on Linux and ChromeOS: this CL converts a BindImage()
call that passes false (indicating that the image is decoder-managed)
into AttachImageToTextureWithClientBinding(). In this particular case,
it doesn't matter whether the call specifies that the image should be
decoder-managed or client-managed as the test subsequently ends
without performing any operations that would result in lazy binding.

Bug: 1323341
Change-Id: I9be84169aec8db95dffd153a2e628ed60d3555ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055269
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077446}