Fix race condition after failed pin login attempt
The problem was that we kicked off the request to learn whether PIN
login is locked out due to too many attempts at the same time as we
kicked off PIN authentication itself. This meant that the ordering of
these two requests was not determined. The problem had probably been
there for a while, but it only surfaced relatively recently with the
change to the auth-factor-based cryptohome daemon API. The problem was
noticed due to failures of the lockscreen.PinUnlockFail tast test.

Bug: b:260740697
Change-Id: I2d81a88c065d4396b9930abd7a6dbbc1795c9f87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4063760
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Martin Bidlingmaier <mbid@google.com>
Cr-Commit-Position: refs/heads/main@{#1077449}