[viz] Include guest view Surfaces in browser-driven frame eviction
The browser process compositor frame eviction logic only tracks and
evicts toplevel compositor frames. This means that any process in the
FrameTree hierarchy holding surfaces do not get its surface (and hence,
tiles) evicted unless specially accounted for.

This is the case for third-party frames with site isolation, where
RenderViewHostImpl::CollectSurfaceIdsForEviction() enumerates the child
frames. But the logic doesn't apply to guest views, until this commit.

Practically speaking, this means that in particular any PDF tab will
retain two sets of tiles which are never evicted as long as the tab is
alive. On a High-DPI screen, the cost is very high, several hundreds of
MiB with a handful of PDF tabs on a macbook, for instance (at least
350MiB with 6 PDF tabs, see document in linked bug). None of this memory
is required for correctness, and performance is unlikely to suffer even,
given the way PDF rendering works in Chromium.

This commit changes the recursion logic to also include child frames. It
should apply beyond PDFs, but solves the problem mentioned above. On a
14 inch macbook pro, it saves >50MiB of GPU memory per PDF tab.

Bug: 1360351
Change-Id: Id1394e540ac888486e588dc9ade982c6adb43feb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4061850
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077481}