[FedCM] Reset `has_pending_web_identity_request_` on primary page change
Currently it is possible that the pending requests from a page that is
bf cached or the RenderFrameHost otherwise not destroyed to block the
requests from a new primary page on the WebContents. Thus, we change
the data to be PageUserData instead of WebContentsUserData.

Bug: 1393388
Change-Id: I8521a80bbc0af7dcb98d5a76cb0d9518c4a80ad3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4053946
Commit-Queue: Nicolás Peña <npm@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077499}