[Page Zoom] Remove Zoom option from NTP and fix potential NPE
This is a quick CL to remove the App Menu "Zoom" option from the app
menu on the NTP or start surfaces. Zoom needs a webContents to function,
and so it should not be shown on a NTP. Selecting the "Zoom" option in
that instance would cause an NPE when the page zoom feature attempts to
get the webContents from a null Tab.

During this fix we also noticed the same issue existed for the image
descriptions feature, so that is fixed as well.

AX-Relnotes: N/A
Bug: 645717, 1232536, 1394503
Change-Id: Ibc33a8df67b65a05b6cf18e447264b0c900e6753
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4064594
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077516}