[ReadingList] Split ReadingListModelStorage's impl to dedicated file
No behavioral changes. In fact, the bridge's public API remains
unchanged as ReadingListSyncBridge continues to (temporarily) inherit
the implementation from ReadingListModelStorageImpl.

The goal is to start decoupling the storage from the bridge, by moving
logic related to the first to a dedicated file. The main API signature
changes are around ScopedBatchUpdate, which now hosts write operations
as member functions: this is because the underlying transaction's
lifetime (determined by ScopedBatchUpdate's lifetime) needs to be extended until the bridge's API (the newly-introduced
DidAddOrUpdateEntry() and DidRemoveEntry()) are completed.

Change-Id: Ib807443910b56dc927fe0905f42cb2542eb7aa34
Bug: 1386158
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4042326
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077522}