Fix regression with devtools and stuck profiling.
Workers will always return the associated WorkerScheduler task runners
where LocalDOMWindow return the AgentGroupScheduler if it was destroyed.
In crrev.com/4ce6c34b it was changed to query this from the execution
context. While this code worked for LocalDOMWindow it got a task runner
that wouldn't run for Workers if the parent worker got destroyed. Always
use the cleanup task runner which restores the old behavior.

BUG=1392680

Change-Id: Ib8b6553c6f71e487f809c7edba4c21f8bcef3ae7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4062298
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077557}