[Image Descriptions] Do not try to localize unrecognized icon types
This CL fixes a crash that can occur when we attempt to use the message
id for an icon from the enum type provided by the service. If the icon
type is not recognized, we return 0. The annotator should then continue
to the next item in the loop, same as with regular image message ids.
We were erroneously not continuing to the next item, with this CL we
fix that.

AX-Relnotes: N/A
Bug: 1394211
Change-Id: I85d399cea0fea5afced018adf7ef9ad57ff1bee7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4063845
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1077569}