Fixing undefined behavior in Android JNI bridge
Reading union elements not specifically set is undefined behavior
in C++, and jvalue null_value = {0}; only sets the first element
of the union (in this case jboolean).

This accidentally works right now since it is set to 0 and the
compiler sets this uninitialized space to 0. This was found when
looking for reads from uninitialized memory.

Bug: 1131993

Change-Id: I308d716fa2e90412a90a7951e67360e523651d98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4058936
Auto-Submit: Kristian Monsen <kristianm@chromium.org>
Reviewed-by: Michael Bai <michaelbai@chromium.org>
Commit-Queue: Kristian Monsen <kristianm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077573}