Limit swipe to show launcher gesture to only part of the shelf
We want to be able to recognize the swipe/fling action only on part of
the shelf that is closer to the launcher button.
Ideally, a user should be able to use the gesture in the area between
the home button and the first view on the scrollable shelf.
When the scrollable shelf is full, we allow a minimum width for the
shelf to recognize the gesture.

Bug: None
Change-Id: Ic7b192f5c284dba28103e44a67de61621d171c13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4024323
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077589}