Stop throttling frames dispatched by OffscreenCanvas
Currently there is a limit of 3 frames that can be in the round trip
between the OffscreenCanvas and its placeholder.  This limit does not
make sense because user code is allowed to retain snapshots of the
content, which will retain references to CanvasResource objects under
the hood. This is bad because it can cause CanvasResourceDispatcher
to stop sending frames when previous frames are not returned quickly.
We should still have a limit, as a guardrail to protect against memory
leaks in the event of main thread CPU starvation. The limit just needs
to be high enough to avoid impeding reasonable use cases.

Bug: 1393703
Change-Id: I4f1a4e893d75606903274a58d130f1174cf9539e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4060564
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077602}