[Multidevice Setup] Added ARIA attribute to Learn More link.
Added an aria-describedby attribute to learn more link in the multidevice setup page. The link node is dynamically created as it is embedded in a grdp string, so this must be done via a method.

Also added a comment describing similar functionality in the existing code that was not otherwise clear.

Bug: 1350648
Change-Id: I74ecc177be0631c8892084d54f0df4f433c027e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4053642
Commit-Queue: Mark Polner <polner@google.com>
Reviewed-by: Jon Mann <jonmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077626}