Simplify and rearrange some logic in RenderFrameImpl::BeginNavigation
Currently, RenderFrameImpl::BeginNavigation has a big nested block for
handling current-tab navigations with the BeginNavigation IPC,
followed by logic that handles downloads and then uses OpenURL for
everything else.

This CL rearranges this to get rid of the nested block and clarify
that BeginNavigation should be used in the common case.  Specifically,
we now handle downloads and OpenURL cases with early returns, and
proceed to BeginNavigation for everything else.  No behavioral changes
are expected.

Bug: 1394513
Change-Id: I50ee4be25241078fe5de18103b735df14b528efe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4064743
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077734}