[Read Anything] Pass callback, snapshot, and content_node_ids through AXTreeDistiller.
AXTreeDistiller::Distill() is called with a mojom::Frame callback which
is called once the distillation process completes. This CL passes
callback across the entire process pathway to ensure it is available
when distillation completes, including when distillation occurs
out-of-process.

Before this CL, AXTreeDistiller cached the callback and called it once
distillation completed (either in process via the algorithm, or
out-of-process via Screen2x).

If Distill() was called again with a new callback before the Screen2x
distillation had returned results, the cached callback was replaced,
and when Screen2x finished, it called the new callback. When the second
Screen2x call finished, it didn't have any callback to run, which
resulted in a crash.

Bug: 1266555
Change-Id: I13d0f1506c380a08d698ec5751f34cd3ce341c09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4067679
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Reviewed-by: Jocelyn Tran <jocelyntran@google.com>
Cr-Commit-Position: refs/heads/main@{#1077737}