WebUI: Use webui::SetupWebUIDataSource() in a few more WebUIs.
In a few cases webui::SetJSModuleDefaults() is used directly along with
calls to AddResources() and SetDefaultResource(). Calling
SetupWebUIDataSource() has the advantage of not using SetDefaultResource() which is an anti-pattern and known to create
confusion.

Moreover, eliminating remaining usages of SetJSModuleDefaults() makes it
a bit easier to keep track of which pages have TrustedTypes CSP checks
on/off, in preparation of changing the default behavior of
SetupWebUIDataSource().

Bug: 1098690
Change-Id: I2c9a83fd7afb12dae8a7d37a9a1972f893ae5f4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4063426
Commit-Queue: Demetrios Papadopoulos <dpapad@chromium.org>
Auto-Submit: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077758}