shortcuts: Add most remaining icons
In this CL, most of the remaining keyboard icons are added to the app.

In icons.html, I added a new `iron-iconset-svg` that has the correct
size attribute to match
https://source.chromium.org/chromium/chromium/src/+/main:ash/webui/common/resources/keyboard_icons.html (which is where we are getting the
icon SVGs from).

In fake_data.ts, I added some code to add all of the icons as fake
accelerators for the purpose of displaying them while developing.
The following two screenshots show all of the icons added in this CL,
using this fake data:
https://screenshot.googleplex.com/9TMjFt6E5LZdx9g
https://screenshot.googleplex.com/9ZBX64hrecbYfyy

In input_key.html, I updated the height/width styling to use the correct
variables for iron-icon.

In input_key.ts, I added mappings for all of the new icons.

Bug: b/216049298
Test: browser_tests --gtest_filter=ShortcutCustomizationApp*
Change-Id: I8ecee3e283d470506994bc368d15f538a49d65e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4052620
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Commit-Queue: Camden Bickel <cambickel@google.com>
Reviewed-by: Michael Checo <michaelcheco@google.com>
Cr-Commit-Position: refs/heads/main@{#1077768}