[Partial Translate] Update label resizing logic to remove empty space
Previous logic relied on using the preferred size of bubble view housing
the language names, as this is the one element that can dynamically
change in width as languages are updated. However, this could be 
inaccurate to the actual width of the bubble, and resulted in empty 
vertical space in some text-language combinations due to too much height
being allocated to the text label.
The root cause is that the |padding_view| in |translate_view_| always 
has a preferred width of 0, and so the preferred width of 
|tab_view_top_row_| is less than the actual bubble width when the latter
exceeds the bubble's minimum possible width. The bubble's minimum width 
is determined by the new ComputeLargestViewStateWidth. Thus, this min 
width can be used as the source of truth when the actual width of 
|padding_view| is greater than 0, and the preferred width of 
|tab_view_top_row_| can be used otherwise.

Bug: 1374038
Change-Id: I18cb0b03ea6dd5a0265cf5926567b547ab21f9e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4038864
Reviewed-by: Basia Zimirska <basiaz@google.com>
Commit-Queue: Anthony Cui <cuianthony@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077800}