float: Fix animations when transitioning to and from tablet mode p2.
P1 fixed some cases where the animation was being skipped on
tablet <-> clamshell transition. crrev.com/c/3914844.

This patch handles the clamshell to tablet screenshot which was
incorporated for performance reasons. It expands on the shelf hider to
temporary hide the float container while the screenshot is taken. The
actual float window itself will be on top of the screenshot and
animated using a cross fade animation.

Test: manual
Bug: b/252505039
Change-Id: I7d691308507384f48f6d6ba6e44df427be32b0f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4063008
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077855}