Remove calls of GetActiveUserProfile() on desktop in /chrome/browser/ui/media_router
`ProfileManager::GetActiveUserProfile()` is mainly meaningful on ChromeOS
devices. It's OK to use it on Android as well (it's not ideal, but should be
harmless). On desktop platforms however, this should not be used, as it may
create a new profile, which is very unexpected for the user.

This CL was uploaded by git cl split.

R=takumif@chromium.org

Bug: 1264436
Change-Id: I43d25352bf7973c5c6906ab10d2f5381e3ae9a85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4042330
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077934}