[GPU] Separate AbstractTexture::BindImage() into decoder/client managed
AbstractTexture::BindImage() takes in a boolean parameter that specifies
whether the client manages the image (in which case the GPU service
internals do not need to on-demand bind the GLImage to the texture) or
the decoder manages the image (in which case the GPU service internals
*do* need to on-demand bind the GLImage to the texture).

As part of eliminating the GLES decoders needing to perform lazy
binding, I am looking to isolate where and on which platforms there is
still potential for such lazy binding. It turns out that the division
between decoder- and client-managed images is on a per-platform basis:

- Windows and Mac always specify that the image should be
  decoder-managed
- Other platforms always specify that the image should be client-managed

This CL removes the boolean parameter from AbstractTexture::BindImage(),
separating it into two variant methods that hardcode semantics to either
(1) set the image as needing binding by the decoder (on
Win/Mac) or (2) set the image as not needing binding (on all other
platforms). This makes it much easier to reason about when this flow has
(and doesn't have) the potential to result in lazy binding. This CL
also similarly separates CommandBufferHelper::BindImage(), the only
caller of AbstractTexture::BindImage(), into
CommandBufferHelper::Bind{Decoder, Client}ManagedImage. In turn,
vda_video_decoder.cc now calls
CommandBufferHelper::Bind{Decoder, Client}ManagedImage() rather than
passing |false| and |true| (respectively) to
CommandBufferHelper::BindImage().

There is no functional change, as this CL is merely codifying existing
per-platform differences.

Bug: 1323341
Change-Id: I026f94983a03e109a7350344b74f3c8abe812de9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055306
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077945}