LayoutNG: Do not call UpdateLogicalWidth/Height for LayoutReplaced
We should not call UpdateLogicalWidth() and UpdateLogicalHeight() from
UpdateLayout() of LayoutReplaced subclasses.

We need ReplacedContentRect() with the last layout result and
ReplacedContentRect() with the new layout result in UpdateLayout().
So, this CL updates ReplacedContentRect() and functions used by it
so that they take source data (border-box size and border-padding
strut) as arguments, and UpdateLayout() passes
 - Data from LayoutBox fields, or
 - Data from BoxLayoutExtraInput.

This CL should have no behavior changes.

Bug: 1353190
Change-Id: I745832984710191572809c638633ff0984b35722
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4059933
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077973}