Replacing PasswordForm with CredentialUIEntry inside OnEdited
Whenever a password is updated through SavedPasswordsPresenter,
OnEdited is invoked. It is used to check for leaks/weak of edited
credential.

Before this change: it was invoked for all the duplicates in all cases.
After this change: it is invoked only once and only if username or
password values were changed.

Bug: 1381203
Change-Id: I8952fe074a07e5ce9325dc4bc444164a9000cbb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4066642
Commit-Queue: Viktor Semeniuk <vsemeniuk@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077981}