[Managed Prefs]  JavaScript Settings Toggle Disabled by Policy
When JavaScript was disabeld by policy and the toggle was clicked in
Site Settings, the browser crashed. The issue is that the code
asserted that it is not managed before enabling the toggle, but
it was managed so the assertion failed as shown in the bug stack trace.

This CL fixes this issue by disabling the toggle when the setting is
managed by policy.

Bug: 1380613
Change-Id: I309a312531eb119eee78c3f6ce2941fe960e2b45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4036412
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078015}