Fix GLFormat for RGBX
SkiaOutputDeviceGL was using GL_RGBA8 for RGX which didn't match the
Skia expectation of GL_RGBX8_ANGLE when it's supported. This caused the
SkiaOutputDevice::Draw() to fail when SkSurfaceCharacterizations were
not compatible. Fix GL format and add the DCHECK that draw was
successful back on Ozone.

Fixed: 1317468
Change-Id: I4e51f4d82066a811dd13203d61fd435024bf3e89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4067600
Commit-Queue: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078048}