cros: Check policy value on recovery eligibility screen
Check the value of RecoveryFactorBehavior policy to set the
recovery_factor_opted_in flag in the wizard context.

Unmanaged users will be able to change the opt-in value on the consent
screen. The default value is `true` - the default of the
kRecoveryFactorBehavior pref.

Managed users will not be able to change the opt-in value on the consent
screen. The value will be set from the policy.

Bug: b:239804500
Change-Id: I4dfe5cd559123219ad87b99e99da399c59a2c502
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4057947
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Anastasiia N <anastasiian@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078079}