FSA: Require user activation on renames without parent write access
Consider the following scenario:
- the site acquires a file handle via showSaveFilePicker()
- the site wants to rename this file

In this scenario, the site has write access to the file, but not to its
parent directory. After this CL, a rename in this case is allowed _only_
if the rename is in response to a user gesture.

Files within the OPFS always have write access, so this change
inherently only applies to files outside of the OPFS.

Bug: 1140805, 1394837, b/254157070
Change-Id: I8b57b10d20dc6a0641bd5fbc145e1a6df0deef2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4068259
Commit-Queue: Austin Sullivan <asully@chromium.org>
Reviewed-by: Daseul Lee <dslee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078085}