Fix content analysis resizing after a quick block or warn verdict.
This CL moves the recoding of the "first shown" to when the dialog is
actually first shown instead of when the ContentAnalysisDialog object
is created.  This CL also suppresses the success verdict if the user
performed a cloud analysis and the verdict returned before the dialog
became visible to show progress.

Bug: b:260335167
Change-Id: Iea1630cfade256aefc316df3078f5730c6a14d21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055747
Reviewed-by: Dominique Fauteux-Chapleau <domfc@chromium.org>
Commit-Queue: Roger Tawa <rogerta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078093}