[CloudAP SSO] Enable setting new trusted_param using navigation handle.
The new trusted param, add_cookies_from_browser, is used to determine
whether the network service should add cookies it receives from the
browser to the current request.

In a future CL, the NavigationThrottle for this feature will set the
trusted param to true when the feature is enabled. Until then, it will
always be false.

Bug: 1246839
Change-Id: I7a22ac4748506c82a457c4f44234e965b08c9a94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4022014
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Igor Ruvinov <igorruvinov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078107}