Add a device policy to set default keyboard brightness [1/2]
This first CL adds the device policy and prefs for the keyboard
brightness. Similar to the keyboard backlight color:
crrev.com/c/3884785.

The expected behavior is:
During a new user signs in on the sign in screen, the brightness level
is set to the policy value. Then when the user changes the brightness,
the value will be stored for the user (as a user pref). When switching
to another user, the brightness will be switched to the other user's
pref value.

On the sign-in screen, the brightness will be set to the value provided
by the device policy. The users are able to update the brightness on the
sign-in screen.

The usages of the prefs will be implemented in the next CL in the chain.

Bug: b:250969878
Change-Id: I99004a11a2832903827acbcaec510fe03a3b8193
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4059983
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Bowen Li <lbowen@google.com>
Reviewed-by: Roman Sorokin <rsorokin@google.com>
Cr-Commit-Position: refs/heads/main@{#1078254}