[User Education] Views Help Bubble buttons to use MD defaults
This was requested to put help bubbles in line with other material
surfaces. Previously, we had used a slightly larger font for buttons.

As a result, a lot of code regarding custom styles can actually be removed.

Baseline on pixel test for IPH with action buttons is also updated.

Bug: 1385955
Change-Id: I40245b2ef7b0cd0eda36db3f49197b0ca9d4c14a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4067389
Reviewed-by: Mickey Burks <mickeyburks@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078337}