[realbox] Deflakes realbox tests
based on local benchmarking, I see flakes every 10 or 20 runs with no
obvious rhyme or reason. I suspect the reason for flakes is partly if
not entirely related to matches not being rendered once the ACResult
is updated. This CL attempts to deflake or reduce flakiness of the
realbox tests by ensuring the matches are updates synchronously every
time the ACResult is updated.

Bug: 1369011
Change-Id: Ie48c7bab34cbf8a93cb3023021dbf4355ab1258c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4070784
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078364}