[Test Automation] Use specialized logic to send Mac browser accelerators
Apparently, the method we were using to inject accelerators on Mac for
browser windows was not working (though it did work for dialogs).

This CL switches to using ui_controls::SendKeyPress() on Mac when
sending keypresses to the browser window. We cannot use this approach
generally as it is not avaialble on Aura platforms in browser_tests
(only interactive_ui_tests installs the correct handlers).

Change-Id: I75a10f665349626573dc3ab34cdb71c43a343454
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4072413
Reviewed-by: Kuan Huang <kuanhuang@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078399}