media/gpu/v4l2: Add HEVC stateful decode support
This adds HEVC decode support to the V4L2 stateful backend. It is only
enabled when the HEVC parser and decoder has been enabled as part of the
build flags. I.e.

enable_hevc_parser_and_hw_decoder = true

This is currently not enabled on any V4L2-based devices.

For Qualcomm devices, it is necessary to group frame NALU's together
while writing the encoded bitstream to the OUTPUT queue. The SPS and PPS
must be submitted with a complete frame, which is different than H.264
for the same devices. The HEVCInputBufferFragmentSplitter method
AdvanceFrameFragment() handles this requirement.

Bug: b:232255167
Test: video_decode_accelerator_tests on trogdor after locally enabling
Test: the HEVC decode build flag.
Change-Id: I233d30134e56e914a8ea0f9fa2bd131147228294
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4068739
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Steve Cho <stevecho@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078422}