[text-decorations] Add & enable flag to disable text-decoration override
This adds and enables a flag to disable the non-standard text-decoration
color override logic.

This "quirk" was only ever meant to be for the <font> element (not for
the <a> element) and in quirks mode.

Firefox applies this quirk in a far reduced set of circumstances, and
likely isn't required by web content anymore.

This is hooked up to a finch kill-switch so we can disable if necessary.

See:
https://github.com/whatwg/html/pull/8568
https://github.com/whatwg/html/issues/8566

Bug: 1394817
Change-Id: Icdf86b1099c81571729767b56a7b3aaee7c85e7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4066694
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078424}