Propagate device enumerate failures from MediaDevicesManager to JS
Add a result_code value to the response of
MediaStreamManager::EnumerateDevices and the associated mojom interface,
indicating if the enumeration succeeded, or the reason it failed.
Propagate this down to reject the enumerateDevices() promise in the JS
API. Also package the response up into a single mojom struct to make the
interface smaller while we're here.

Before the effort tracked in crbug.com/1313822, these errors would be
hidden by the camera capture service, and an empty list of cameras would
be propagated back up the stack as if it were a success. I've changed
all other callers of the enumerateDevice API (eg in getUserMedia,
AudioContext etc) to now handle an error in the same way, so no other
codepaths should change behaviour yet - TODOs have been added to track
propagating these errors further in appropriate ways.

Bug: 1313822
Change-Id: I457746c6ebd391d184c6dd2796127ac6dd80ed4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055531
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Tony Herre <toprice@chromium.org>
Reviewed-by: Michael Wilson <mjwilson@chromium.org>
Reviewed-by: Derek Schuff <dschuff@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078494}