Reenable TranslateInfobarTestCase::*HTTP* EGTests
HTML content language is only considered if the language
detection of the content fails or if it agrees with the
language detection.

In the current test, the actual language was english and
the HTML content language was fr leading to "undefined"
language.

Set the content of the test web page to 123456 so that
no language is detected and the HTML language is used.

Bug: 1328970
Change-Id: Ibdc38b75adad4b2415e1b5c2f5a7df4858a8de2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4069731
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Quentin Pubert <qpubert@google.com>
Cr-Commit-Position: refs/heads/main@{#1078497}