FLEDGE: Run a parallel auction if needed based on k-anon mode.
If we're configured to enforce k-anonymity, the alternate auction will keep track of the non-k-anon winner.

If we're configured to simulate operation of k-anonymity, the alternate auction will keep track of what the winner would be if enforcement were on.

The previous work in this are on BidderWorklet is changed some by this CL since it sometimes wasn't distinguishing between main and alternate bids being identical vs. alternate bid being a failure.

(This is up to the level of AuctionRunner, the service simply ignores it; it may make sense to pull the flag reading into the service).

Bug: 1234419
Change-Id: I67766d04f70fac926853800b1dbaf23c245fba09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3941227
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Russ Hamilton <behamilton@google.com>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078556}