Log warning when write protect is missing from machine_info
The value is written in the file at device boot and is
populated from crossystem. It would be unexpected if we don't find
the value in the file but find it in crossystem when Chrome is
asking. Log the warning to have data in the logs in that case.

BUG=b:249498455
TEST=None

Change-Id: Ifec949c609230e17395e8d49aca688e3bde53f28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4076247
Commit-Queue: Igor <igorcov@chromium.org>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078579}