Fix flaky chrome://metrics-internals test
The test is flaky on Chrome-branded builds on Windows.

This is because:

1. Apparently, in Chrome-branded browsertests builds, we are
considered on the "Stable" channel.
2. Windows is sampled at 10% on Stable for UMA

So the test fails 90% of the time on official builders (when we are
sampled out, metrics service is disabled).

This CL fixes this by simulating being sampled in.

Bug: 1394384
Change-Id: I6e47b0d21ab660740ec5d02ef8a3e768d57872f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4073404
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Auto-Submit: Luc Nguyen <lucnguyen@google.com>
Commit-Queue: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078584}