[ios] Fix discover feed menu anchor
This change fixes the anchor for the discover feed menu so that the
alert view based menu correctly points to the button.

The bug was caused by the fact that the button is in a container
which is itself contained in the FeedHeaderViewController's view.
Because of this, the coordinate system for the button's frame was not
the same as the coordinate system for the FeedHeaderViewController
view's frame.

Screenshot with fix:
https://screenshot.googleplex.com/9USMeJyEDvTdgmv.png

Fixed: 1298471
Change-Id: If2f785fc72f5c4db534f634e4547ee557ec41414
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4066396
Commit-Queue: Scott Yoder <scottyoder@google.com>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078625}