[Test Automation] Correctly report queued step when test times out
Previously, when an InteractionSequence was destroyed, in the aborted
callback, the current step would be cited in most cases. This meant
that, for example, if an InteractiveTest timed out waiting for a step,
the step that had already executed would be cited instead of the pending
step where the timeout occurred waiting for.

This change adds "sequence destroyed" to the list of failure reasons
that cite the pending step rather than the current one, remedying the
problem.

It is not a complete solution to debugging information not being as
complete as possible on a timeout, but it should be a vast improvement.

Change-Id: I68f0cb7be534d6a8dac724a5f8aef6995e89024c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4072417
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078683}