[iOS][Omnibox] Fix RTL constraints update
The popup constrains its text and icon to the omnibox named guide.
The constraints are applied when the view is attached to a window.

These constraints should change when RTL changes, the variation 2 of the popup uses `isRTL` to determine the offset of the icon.

This CL adds a second update when RTL changes.

Bug: 1384528
Change-Id: Id76e20b4b3888310180e9546ff6ff5f00ad1fe9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4061083
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Auto-Submit: Christian Xu <christianxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078696}