Anchor chooser dialogs to extension icon
WebUSB and WebHID display a device chooser dialog when a page requests
access to a device. This CL moves the dialog for extension pages so the
dialog is anchored to the extension icon instead of the lock icon in
the omnibox. The extension icon is temporarily pinned while the chooser
is open.

Bug: 1360960
Change-Id: Ie0d55189b59f06e5c7a2c3b57b676930d0c72791
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3936591
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078717}