gif_recording: Refactor CaptureLabelView to host a drop down button
This CL refactors CaptureLabelView so that it can host two buttons
instead of one. The first button is the capture button, when pressed
capture will be performed. The second button, which shows only when
in video recording mode, and multiple video formats are supported
(e.g. webm, gif, ... etc.). This button will trigger a menu from which
the video format can be selected (not in this CL).

The button should behave the exact same way as it does today if the
GifRecording feature is disabled.

Demo is attached to the bug.

Bug: b:261032943
Test: Manual, existing tests.

Change-Id: I22b922879063a9cdc27446eb45f9425c0586ac41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4065019
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Michele Fan <michelefan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078729}