Disable Media Foundation for Clear on older AMD drivers
Edge Watson data shows a spike in crashes in atiuxp64 when experimenting
with MF for Clear enabled.

This change adds a new GPU Blocklist option for MF for Clear & creates
a rule for AMD drivers equal to or below the highest AMD driver we've
seen reported so far (26.20.15002.61).

Bug: 1394484
Change-Id: I27d0b89366e9e8ca867eac1d0c435225953046fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4062911
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Commit-Queue: William Carr <wicarr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1078791}