Defer subresource header registration until prerender activation
Currently if a subresource on a prerendering page returns an
attribution header, we handle the header immediately and try to bind
an associated interface for ConversionHost which would cause the
prerendering to be cancelled due to mojo capacity control to ensure
the prerender correctness. The attribution header is not handled in
this case.

The header handling should be deferred until prerender activation to
avoid that, which would also help prerender to improve the coverage.

Bug: 1394420
Change-Id: Ie8c62d342e958f4170cb857b67c739c22dcb8c7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4060280
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Nan Lin <linnan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078799}