[Omnibox] Add holding space for the status icon
Before the status icon background is added, the icon layout is
  [4dp space] [icon] [4dp space].
So when the icon is GONE, there still 2 [4dp space] there.

After the status icon background is added, the icon layout is
  [background [icon] background].(background is 32dp, icon is 24dp).
And when the icon is GONE, the background is GONE as well, otherwise
there will be 32dp space left.

This CL add a custom status icon, when the status icon and its
background are GONE, a 8dp space we add to the view, so the behavior
will be same as before.

Bug: 1383114
Change-Id: I380f418c6b0a084eec316ee96f975066756461bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4031711
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Commit-Queue: Gang Wu <gangwu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078803}