[Layout Manager] Ensure tabClosed doesn't refocus
When tabClosed is called sometimes the tab that is to be selected was
already focused and the user may have already focused elsewhere.
LayoutManagerChromePhone then ends up requesting focus again. This is
particularly true when closing an undoable tab as it first triggers a
focus in tabPendingClosure and again in onFinishingTabClosure. For now,
fix the issue of refocusing a previously focused tab from
onFinishingTabClosure by moving the refocusing logic to StaticLayout
and focusing when the tab is selected or changed.

Why does this currently happen?

We currently call tabClosed() in 4 cases:

1. tabPendingClosure: a single tab was be closed, but can be undone.
2. multipleTabsPendingClosure: a bulk set of tabs was closed, but
   can be undone.
3. tabRemoved: the tab is no longer in the model, but may exist
   elsewhere. Called on the tab being removed from the model.
4. onFinishingTabClosure: the tab is closed and cannot be undone.
   Called once per closed tab regardless of source, but after the tab
   is permanently closed. This is a catchall for tabs that were not
   able to be undone so didn't trigger any of the other calls. However,
   it is called too late for undoable tabs so cannot be relied on as
   the sole signal.

There is also willCloseTab, willCloseMultipleTabs, and willCloseAllTabs,
but these are called before the tabs are removed from the model, which
is too early.

This could all be simplified to a single event if we had a single
"tab closed" event that was sent when a tab was no longer in the
TabModel. Think of it like an earlier onFinishingTabClosure (e.g.
onStartingTabClosure) which is called regardless of cause; however, that
currently doesn't exist. I'm happy to work on that as an alternative
or follow-up to this, but it is more complicated.

The current behavior is error prone as all cases need to be checked,
but the distinction is also important for some use cases such as: Undo,
recent tabs, etc. Hence why there is currently no unified event.
Perhaps there is an opportunity to make these more complicated events
internal to tabmodel/ so that this complexity isn't leaked elsewhere
although that requires careful consideration.

Bug: 1394372
Change-Id: I0664c3dd35dbebca187b5568380cb52038eff4df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4062732
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078813}