Add use counter for non-standard values of appearance
All keywords not listed in [1] are not standard appearance values and
should be deprecated for interoperability. Since this deprecation will
cause slight web changes, we want to add a use counter first to assess
the magnitude of this change before landing an intent to ship.

[1] https://drafts.csswg.org/css-ui-4/#typedef-appearance-compat-auto

Change-Id: Ia5f9ecb21064d5d4d5b3a3b1ca7bfda95a94053e
Bug: 981719
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4076301
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Di Zhang <dizhangg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078869}