[blinkpy] 15/N: Update from CI results
The build resolver for `update-metadata` and `rebaseline-cl` relied on
the `builders.json` config to tell if a builder was 'try' or 'ci'. This
is problematic when a builder has the same name in both the 'try' and
'ci' buckets because `builders.json` is keyed on the builder name.

Instead, allow `update-metadata` users to specify the bucket explicitly
with an optional `{ci,try}/` prefix in the `--build` parameter
(defaulting to 'try'). This syntax resembles that of other tools like
lucicfg and buildbucket.

This is a no-op change for `rebaseline-cl`, which continues to assume
only try builders.

Also, add a build range syntax "<builder>:<start>-<end>" so that it is
easy to update expectations over a longer time horizon (which gives a
stronger signal of test flakiness).

Bug: 1299650
Test: git cl issue 3601551
Test: ./blink_tool.py rebaseline-cl --verbose --dry-run --patchset=15
Change-Id: I2ae30e92e7ed0b3c3b00eff43b6fc256d1469603
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4076036
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1078899}