[Web Payment] Use test installer for Android payment app browsertest
Before this patch, the Android payment app browsertest was installing a
payment handler for https://play.google.com/billing payment method
identifier using the PaymentInstruments API that is planned to be
removed. The test installer could not be used as is in this test,
because it sets the payment method identifier to be the origin of the
payment handler's service worker JavaScript file.

This patch switches the Android payment app browsertest to use the
test-only installer tool, which now has an additional parameter for a
custom payment method identifier, i.e., https://play.google.com/billing
in this case.

Bug: 1327265
Change-Id: Ibbbcaef5de06bbfa4eeb9e75385cb45e96140aad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4076485
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078950}