[QuickStart] Add BootstrapOptions Logic to AuthenticatedController
This CL adds the constructor and the Payload generation code for
BootstrapOptions. The payload is generated according
to go/cros-quick-start-messaging-specs.

go/cros-quick-start-base-client-class for more information wrt the BaseClientClass and Derived Clients

NOTE: There is a TODO to clean up the NearbyConnection once we have implemented the correct way to pass it through. The current implementation is to unblock work on the clients. We will also be cleaning up TargetDeviceClientBase in another CL.

TEST=added unit test

Bug: b/260633661
Change-Id: I90a94aca5ac77245fb3da0e88617df3ef67da511
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4062647
Reviewed-by: Michael Hansen <hansenmichael@google.com>
Reviewed-by: Curt Clemens <cclem@google.com>
Commit-Queue: Jason Rhee <jasonrhee@google.com>
Cr-Commit-Position: refs/heads/main@{#1078966}