autocorrect: record enable by default transitions
This CL updates PrefChangeRecorder to handle new possible autocorrect
pref state transitions. With the inclusion of the EnabledByDefault
preference, we can now move from Default to EnabledByDefault, and
EnabledByDefault to Disabled.

We do not record the transition between EnabledByDefault to Default as
this would require a branch of logic added that is NOT guarded by a
feature flag.

BUG=b:257146732

Change-Id: Idab05f1c55c96a6d8bca468cbdcae6d8f26fb7f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4054752
Reviewed-by: Mehrab N <mehrab@chromium.org>
Commit-Queue: Curtis McMullan <curtismcmullan@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079047}