[APS] Restrict AppPreloadService to unmanaged regular accounts
For the initial launch of APS, we will only be preloading OEM apps on
regular, unmanaged/unsupervised accounts. This restricts the service so
it does not start for other account types, preventing it from making
network calls to the APS server when it is not necessary.

Bug: b/259494622
Change-Id: Id4ab62eed2224c0edbccfdbc1163cf65cc31e3f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4073604
Auto-Submit: Tim Sergeant <tsergeant@chromium.org>
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Jeevan Shikaram <jshikaram@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079071}