[Office upload] Implement installation of web app from setup dialog
Install the web app via ExternallyManagedAppManager. This will attempt
to retrieve the manifest in the background, and will fallback to
installing it based off metadata from the website if the manifest cannot
be found, which can happen if the user isn't logged in to Microsoft 365.

Bug: b/259869338
Change-Id: Id843d083b600f0a06bc61aeb06f6979b169449b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4040389
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Austin Tankiang <austinct@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079090}