[UserSessionManager] Switch CHECK to a condition in profile creation
The refactoring of `ProfileManager::CreateProfileAsync()` changed the behavior on failure during the process of profile creation (crrev.com/c/3804886); it used to return an enum to designate failure and now returns a nullptr profile.

A crash was reported due to the fact that the profile creation happened while the browser process was shutting down, causing the returned profile to be null and crash on the `CHECK(profile)`.

This CL changes the CHECK to an early return for nullptr profiles to protect against the crash.

Bug: 1381160
Change-Id: I82c5826285615bb88b59b5f0b97d9d5f2c72b52d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4008359
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1079138}