[Translate] TranslateRanker only created for user-visible profiles
The translate ranker is no longer built for System profiles and Ash
non-user profiles.
It is still built for Guest profiles, because the Translate feature is
available in these profiles.

This CL also cleans up histograms:
- remove expired histograms
- renames Translate.Ranker.Model.Status to
  Translate.Ranker.Model.Status2 as it is impacted by the new behavior.

It is not really clear whether Translate.Ranker.Model.Status2 is still
useful or not. There is no evidence that it is still used.

Bug: 1284659
Change-Id: I1f439e969ba338f7b3c7504b78f5b31ca63689b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4028662
Commit-Queue: Charles Zhao <charleszhao@chromium.org>
Auto-Submit: David Roger <droger@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Charles Zhao <charleszhao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079142}