[GPU] Separate set_is_bind_pending() into {set, clear}_bind_pending()
AbstractTexturePassthrough::set_is_bind_pending() takes a bool
indicating whether to set or clear the texture as needing binding. To
simplify readability and analysis of the code, this CL separates the
method into two no-parameter methods: one that sets the texture as
needing binding, and one that clears it. This makes it much simpler to
see and understand the flows that actually set the texture as needing
binding.

Bug: 1323341
Change-Id: I21f1804f132566fd4cb159f12cc05ccf4d414f40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4078926
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079208}