[Passwords] Store DB primary key in PasswordForm
This CL is adding the a new int field to the PasswordForm struct that
carries the DB primary key.
This will provide an organic way to uniquely identify passwords in
other layers such as UI.

The field is add an optional field since it will be absent for forms
that aren't yet stored in the DB (e.g. added from settings) or on
platforms that don't use the logins database as a storage layer.

This also updates the PasswordStoreChange() class by omitting the
dedicated field for carrying the primary and relying on the new field
inside PasswordForm

Bug: 1330906
Change-Id: I76e2e4856767ab491de28b1482f0685d1e1074b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4075772
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079227}