Add a builder for Ash SystemNotifications
Most notifications don't need several of the fields required by the
ash::CreateSystemNotification factory function.
This leads to a lot of duplicated and empty code when creating a system
notification. In other instances the factory function isn't flexible
enough leading to a lot of ternary operators or keeping variables
around solely to set them conditionally somewhere in the function body.

To reduce the duplicated code for notifications going forward this
introduces a builder class that allows reusability within a class and
can be integrated into the code flow of a function where things
relevant to the content of a notification happen.

Bug: b:248211321
Change-Id: Ieb6e243ed2c7b27c87cfdb4226dc50af42b8dadf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4057149
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Auto-Submit: Christoph Schlosser <cschlosser@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079269}