[Android Bulk Tab Edit] Fix race condition when fetching thumbnails
When undoing grouping via select tabs the restoration for the thumbnail
of the final tab ungrouped is racy. This is because its state needs to
be updated each time a tab is ungrouped and fetching single thumbnails
are faster than multi-thumbnails. Fix this by checking that the fetcher
associated with the request for the thumbnail is the same as the current
one in the model before setting the thumbnail.

All these requests should probably run on a separate sequenced task
runner to avoid such issues in the future, and to facilitate cancelling
in-flight task but this is sufficient for now.

Bug: 1395467
Change-Id: I308969a8ab5879c20cf003eeed8157133e33bc4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4076411
Reviewed-by: Fred Mello <fredmello@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079295}