Reland "Add tests to verify page load metrics are recorded in cases of crashes."
This is a reland of commit 4cf270351c08ec51555b4a3f215a7199aac5289e

The previous check-in is reverted as the test case that simulate crash
caused by memory exhaustion fails the Linux Asan Msan tests.

Upon investigation, it turns out, with ASAN builder, the
ExhaustMemory() method which runs a loop in the renderer process does
not exhaust memory and keeps running, the navigation does not commit
and the waiter on navigation hence times out.

This CL removes that test case and is to reland other test cases. That
test is to be added back once a proper way to test it is found.

Original change's description:
> Add tests to verify page load metrics are recorded in cases of crashes.
>
> Browser tests are added to verify page load metrics are recorded in
> cases of crashes. Crashes are simulated by navigating to chrome debug
> urls and by shutting down render process.
> During all these crashes, the destructor of the PageLoadTracker is
> called and page load metrics are hence recorded.
>
> Bug: 1383971
> Change-Id: Ia909820cf3237974e888471028a9fe8119b92720
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4027239
> Commit-Queue: Hao Liu <haoliuk@chromium.org>
> Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1075581}

Bug: 1383971
Change-Id: I34a0e12ebb4d752b56a3a6872076a422d9865c51
Include-Ci-Only-Tests: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4068381
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079361}