Center the drag image under the mouse pointer
Currently, drag images are positioned so that the mouse pointer is
over the upper left corner. Besides being weird, it's not useful,
as the mouse pointer image obscures the image of what's being dragged.

Therefore, center the image under the mouse.

Improvements of aligning the image to match what's already on the
screen will need to await further Views improvement, but this is a fix
that can be done now.

Fixed: 856987
Change-Id: I6eb2040b1946f004389dbdd9c35c52f5481ae6a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4073789
Auto-Submit: Avi Drissman <avi@chromium.org>
Reviewed-by: Jayson Adams <shrike@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079396}