Remove references to /opt/chromium/chrome_sandbox.
Based on https://crbug.com/312380, it is no longer necessary to
provide a setuid sandbox, but the testing code is still setting it to
/opt/chromium/chrome_sandbox. This is a non-hermetic binary that is
installed to the test machines via puppet. If the setuid sandbox
environment variable is set, it causes fatal error if it is set to a
non-existent file, which can happen if setup on an Android machine fails
since this results in it reporting itself as a Linux machine and it does not contain the sandbox binary. Since the binary is not actually
necessary, this removes an unnecessary failure mode from the automated
testing.

Bug: 1334743
Change-Id: I1155d8e77e2417b34bf1a03afbb6e821bb4ecb85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4077067
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Mark Seaborn <mseaborn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079418}