Wait for NetworkConnectionTracker initialization in ReportSchedulerTimer
Otherwise, the timer starts observing the network connection tracker
before it is initialized and never receives a call to
OnConnectionChanged post-initialization, which in turn prevents
AdjustOfflineReportTimes from being called, which in turn prevents
MaybeSet from being called.

This is likely one of the causes of crbug.com/1383243.

We also introduce a sequence checker and scoped observation for code
health.

Bug: 1395817, 1383243
Change-Id: I2c4eaa33f3d03fa2336ae4c9259fcc687ea2d380
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4080846
Reviewed-by: Alex Turner <alexmt@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079544}