Fix the anchor rect for Stylus tray bubble
This cl sets the correct anchor rect when the shelf is hidden for the
stylus tray bubble.

Another approach is to fix it in the `TrayBubbleView`. This seems like
a generic fix for all the tray bubbles. However there are some issues
with this approach:
    1) we need to pass in a raw pointer of the shelf view, which is
not always accessible from each bubble.
    2) The position from the`shelf()->GetSystemTrayAnchorRect()` does
not apply for all the tray bubbles. It is the edge of the work area,
which doesn't count the position of the tray. See this screen
recording for the styles tray:
https://drive.google.com/file/d/1cHZyGFc9IB5dFiHAPbtPL86eR3edTEW7/view?usp=sharing
The bubble shows from the keyboard accelerator is on the right side of
the screen. But the bubble should be anchored on the top of the stylus
tray.
See the draft cl of this approach:
https://chromium-review.googlesource.com/c/chromium/src/+/4074865/1/ash/system/tray/tray_bubble_view.cc

Screen recording after fix:
https://drive.google.com/file/d/1GKDtzpvTLy5jW6pwJ6BzthYB-nZHvx1n/view?usp=sharing
(see the screen recording before fix in the bug)

Bug: b/258690409
Change-Id: I96395d0d368aac2ccef25fd33cfb03d2cba14e79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4077046
Reviewed-by: Andre Le <leandre@chromium.org>
Commit-Queue: Jiaming Cheng <jiamingc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079602}