[floss] Fix flaky behavior in BluetoothSocketFloss
BluetoothSocketFloss will DCHECK if it is destructed in a different
thread than where it is created (because it is originally kept as
a scoped_refptr). Make sure both the successful and failing socket
callbacks have a reference to the socket so the destructor runs in the
ui thread.

For listening sockets, we need to a bit more:
* Keep a reference to the accepted socket on the listening socket (keeps
  it from going out of scope while a socket thread is active).
* Make the tasks on the socket thread cancelable and cancel tasks during
  disconnect (again to keep the socket from going out of scope in the
  socket thread).

There is still a race in listen even after these fixes but it is
improved after this patch.

BUG=b:243420879
TEST=device_unittests --gtest_filter=BluetoothSocketFloss\*

Change-Id: I9196d0e36b4b11ca892a3aaa19b94dbfc4b0b187
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4060397
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Cr-Commit-Position: refs/heads/main@{#1079603}