Pass `sRGB` values to inspector frontend
* `SkColor4f` holds colors in sRGB in a non-clamped
fashion. So, the values that are out of bounds of
sRGB can represent colors in other color spaces

* `Y` in `XYZ65` color space gives the relative
luminance of the color which we use while
making the contrast calculations.

In here, we're passing the rgb values of the
color to inspector frontend which then
calculates the luminance of the color
in the frontend (by partially converting
it to XYZ65 color space) and finds
the contrast ratio.

In addition to that, we're passing backgroundColorCssText to
the frontend too so that we're able to render the background
correctly for the contrast info preview ("Aa" text in the background)

Bug: 1384525
Change-Id: Ie2ec158a19ac4ffd64445625b907686ca3c354f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4081067
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079719}