Speculative fix for flake in scrolling test
This patch replaces 8 calls to waitForAnimationEndTimeBased with waits
for a scrollend event.  though scrollend is experimental it has been
enabled for some time in testing and is close to I2S.  Each time based wait adds 200ms to the test duration, contributing 1.6s to the total run time of the test.  To make matters worse. Tests run on debug builds tend to be 3x slower than their release counterparts.

This patch also switches from onMouseClick to test_driver, which is the preferred API since supported on WPT.

Bug: 1395295
Change-Id: I8313dc903af37e06e65f4e0498647f82bf15a49e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4081117
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079840}