Reuse server endpoints by default
Server endpoints are now reused on Mac and Linux. On Windows this is not
possible due to limitations of the underlying mechanisms. Thus, this
behavior has been isolated as an implementation detail of the Windows
endpoint connector.

This change also allowed some of the interfaces to be cleaned up.

Fixed: 1395178
Change-Id: I8e1e00b82cc602fbd7c1e41e2f6536435f779c17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4070125
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Cr-Commit-Position: refs/heads/main@{#1079872}