ash: Match deskbar layout with desks spatial order
Prior to this CL, when RTL was enabled, the desk bar was laid out in
this way:

  [Library] [NewDesk] [Desk 3] [Desk 2] [Desk 1]

However, the spatial order of desks, as seen when switching between
desks was still LTR. For example, switching from Desk 1 to Desk 2 would
scroll to the right.

With this change, desks in the bar will match the spatial order. So
when RTL is enabled, the desk bar will look like this:

  [Library] [NewDesk] [Desk 1] [Desk 2] [Desk 3]

BUG=b:247912921

Change-Id: Ifd4a5acb3761397a211f4a5b01efc05e1fc92ef9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4076793
Commit-Queue: Daniel Andersson <dandersson@chromium.org>
Reviewed-by: Yongshun Liu <yongshun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079937}