Specify AttributionDataHost registration-eligibility at creation time
This allows the AttributionDataHostManagerImpl to avoid counting a data
host as being in "source mode" when the renderer knows that it will only
ever be used to register triggers. Prior to this CL, all data hosts
were considered eligible to register sources until the first trigger
arrived.

This reduces the number of triggers that need to be buffered while
waiting for the number of source-mode data hosts to reach 0, and will
make it easier to remove the requirement that a single redirect chain is
only allowed to register sources or triggers but not both.

Bug: 1395047, 1309176, 1309173
Change-Id: I54846a5b3e1236a60b51a02441abe9aa2afcfc93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4071920
Reviewed-by: Nan Lin <linnan@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079950}