Split IDP sign-in status into its own permission
Prior to this CL the identity provider sign-in status was stored in
the  ContentSettingsType::FEDERATED_IDENTITY_SHARING permission. This
was a hack as the sharing permission is keyed by the relying party
requester but the IDP sign-in status is keyed by the identity
provider URL. This CL introduces a new dedicated permission -
FEDERATED_IDENTITY_IDENTITY_PROVIDER_SIGNIN_STATE - for storing the
IDP sign-in status.

BUG=1384288
TEST=FederatedIdentityIdentityProviderSigninModeContextTest.*

Change-Id: I0c532f79e19c0369dc6d5dd99731d189de74a3e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4068860
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Peter Kotwicz <pkotwicz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079956}