Fix the system nudge position with auto-hide shelf
This bug is caused the by the wrong state check in the calculation.
After logging in the state of the shelf is still with lock state
when the nudge is rendered.

Here I recorded Dark/Light nudge as an example in the screen records
below:

before fix (see the wrong position from 25s) :
https://drive.google.com/file/d/1U2a4nGHwCQptHgi00n6gAgUxQ3Dn9ZlN/view?usp=sharing

After fix (see the corrected position from 28s):
https://drive.google.com/file/d/1VU9WuzJo_T0sdsGK0YLXeUhr3flROoaZ/view?usp=sharing

Bug: b/251738095
Change-Id: I5823ce07d6a0b1a581ba81a5b16e2f2c1cca7984
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4076039
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Jiaming Cheng <jiamingc@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079967}