Emit important UMA histograms earlier behind feature flag.
Currently, some important histograms used as filters are only emitted when closing a log. In the case of an unexpected Chrome termination, those histograms will not be emitted. Consequently, associated logs will not be tagged.

In this CL, changes are made so that important data/histograms are emitted when a new MetricsLog is created in OnDidCreateMetricsLog(). In this way, if Chrome unexpectedly terminates during a log, those histograms will be in .pma files on disk. Next Chrome session will be able to send the log to the server as an independent log and it will be properly tagged. In the case of OnDidCreateMetricsLog() fails to emit histograms, they will be emitted in ProvideCurrentSessionData() called in CloseCurrentLog().

Note that for some histograms, the value will change during the time span of one log, the value at the creation of the log might be different from the value at the closure of the log.

Those histograms will also be emitted in DisableRecording() which is called during shutdown so that the next session will include them in the independent log.

All related unit tests and browser tests are parameterized to test the behaviour with both the feature enabled and disabled.

Currently, only histograms used as filters are emitted earlier, but changes can be made later to emit histograms earlier when possible.

Bug: 1367008
Change-Id: I70218315f1c674d28e04deda4f8ff4ceb9fcd697
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3988404
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Ariel Zhang <arielzhang@google.com>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Cr-Commit-Position: refs/heads/main@{#1079977}