Storage buckets: handle bucket deletion more gracefully.
When checking or updating bucket policies, check whether the bucket
has been deleted (e.g. via the clear browsing data dialog). Reject
promises in this case.

This also fixes a bug where BucketManagerHost used the name to track
the BucketHost. This was a bug because a name can be reused after
deletion. It now uses the bucket ID, which can never be reused.

Bug: none
Change-Id: Ib49196b3ef656c32c2922499db665ca4887a026b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4077246
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080021}