Expose button with aria-haspopup=dialog as a regular button
Note that we are making this change ahead of the spec change, because
it's an obvious usability improvement that keeps coming up. Will also drive the spec change.

Bug: https://github.com/w3c/core-aam/issues/51
Change-Id: Ie06afeae4118b868eaee6471f8bda9d31b25a027
AX-Relnotes: buttons with aria-haspopup=dialog will no longer use the button menu role, because it's both misleading to the user and to the screen reader, which might trigger the wrong mode.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4081366
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1080035}