[Test Automation] Switch to synchronous mouse operations.
Previously, MoveMouseTo, ClickMouse, etc. used a MultiStep where there was a callback that sent an event, which creates a problem where state transitions that come between the send-mouse-events call and the follow-up event could swallow a discrete event in the following step; e.g.:

  ClickMouse(), // if event comes in between steps...
  WaitForEvent(...),  // ... this won't trigger

What we are doing is indistinguishable from just having a single step with a RunLoop, except that in very specific cases, we wanted to avoid a nested RunLoop because of the way drag-drop operations happen.

This CL switches from an asynchronous to a synchronous approach, eliminating the swallowed event problem; each Mouse verb is now a single step with a callback that performs the mouse action and then returns. It does not appear to break any existing tests, though further work will need to be done to ensure that drag-drop functions correctly in all cases on all platforms.

Follow up steps:
 - Ensure that mouse events are delivered correctly on Mac (see
   crbug.com/1396074)
 - Change Windows mouse handling so that events that initiate or
   complete a drag are sent from a different thread (crbug.com/1396441)

Bug: 1396164

Change-Id: Id0328811e996de094a7c62609c2323209a1cd5a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4081876
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Kuan Huang <kuanhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080082}