Fix flaky tests due to failing to delete product directory.
The fix consist of not having the unit tests log into the `updater.log` file.

The tests defined by TaskSchedulerTests initialize the logging for
unit tests to log in the updater directory in the `updater.log`.
The log file is created with sharing permission for read and write but not delete.

However, the integration tests must delete the product
directory including the log file, as part of enforcing test pre and post conditions.

When integration tests are shuffled so that in the same batch of tests they are run after any of the TaskSchedulerTests, the product directory can't be cleaned up because of a sharing violation file IO on the updater.log file. The integration test will fail.

In addition, this CL allows the unit tests to log directly into
${ISOLATED_OUTDIR} so that unit tests logs can still be captured if
needed, for diagnostic purposes.

Bug: 1392429
Change-Id: I387ae68b28686db5a2f447557790159b3575e717
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4083942
Reviewed-by: Xiaoling Bao <xiaolingbao@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080133}