[Fast Pair] Add timeout for writing account key
Add a timeout for writing the account key to a fast pair device, if the
timeout is exceeded, then writing the account key should fail.

Tested: Added a unit test for the timeout and manually tested that the timeout is called and the chrome://histograms is updated.
Change-Id: I4d178940e8fe9b1f0eb20110870cfbb0ed110847
Bug: b/254903408
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4060584
Commit-Queue: Alex Kingsborough <akingsb@google.com>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Daniel Classon <dclasson@google.com>
Cr-Commit-Position: refs/heads/main@{#1080143}