[Fast Pair] Implement discovery notification ban logic
Adds ban logic for discovery notifications for initial and subsequent
scenario based on go/cros-fast-pair-notification-parity-design. Adds a
ban list for model id and protocol for a device, and when a user
dismisses a notification, update the state and the expire timeout
in the ban list. When the device is found by the scanner, check for
the state and timeout before showing a notification for the device.

TESTED=Verified on DUT that dismissal by user bans a notification for
the following durations:
 - first dismissal: 2 seconds
 - second dismissal : 5 minutes
 - third dismissal: until session ends, until toggle is reset

Bug: b/249596355
Change-Id: I45c940d49b034ab6ff2e1f85921880a4c5e018a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4052302
Commit-Queue: Juliet Lévesque <julietlevesque@google.com>
Reviewed-by: Daniel Classon <dclasson@google.com>
Cr-Commit-Position: refs/heads/main@{#1080170}