[NTP][Panorama] Add Background Collections to Categories page
Added the function for getting the list of background collections to the customize chrome page handler and called it in the category page constructor.

Added the BackgroundCollection object to the mojom file for the customize chrome handler and replaced the instances of object that were holding its place in app and categories.

Changed some of the instances of category naming to collection to differentiate clicking on collections (which were added here) and what will happen when classic chrome or upload image is clicked in the future.

Currently, it is just showing on the page as a list of plain text names of the collections. Styling will be added in a future CL.

Screenshot: screenshot/aDqvMoyv78qXdNH.png

Bug: 1384251
Change-Id: If1dcd9cc5212225a7aad42d3bbf8475cb41370a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4067891
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Riley Tatum <rtatum@google.com>
Cr-Commit-Position: refs/heads/main@{#1080225}