Remove expectation on number of GetPendingReportsForInternalUse calls
For the purposes of this test, it's irrelevant how many times this
method is called. What we care about is that the internals UI does the
right thing in issuing a ClearData call and then showing the updated
reports.

We achieve this by removing the exact series of expected calls to
GetPendingReportsForInternalUse and instead always invoking its callback
with a persistent vector of reports that is cleared by the mock
ClearData call.

Bug: 1394360
Change-Id: Idff8f497f7042a457a29d0b40511e5564fe2a48a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4085963
Quick-Run: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Nan Lin <linnan@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080328}