ProfileManager: consistent terminology "created" vs "loaded"
Created/Deleted: the profile is created or deleted from disk
and from UI surfaces. Creating or deleting a profile is a manual
operation from the user, and has significant implications.

Loaded/Unloaded: the profile exists on disk and in the UIs, and is
loaded into a RAM as a C++ `Profile` object. Whether a profile is loaded
or unloaded is mostly a technical/performance consideration and is
usually not very important for the user.

Change-Id: I2f441f9fa7983eb0b933970b24b28cd720768b49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4070745
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080353}