Parameterize EUK key name by Key Algorithm
This CL changes the name of EUKs to be parameterized by their algorithm
(currently RSA or ECC). To maintain backwards compatibility, the naming
of RSA keys is not changed (left as "attest-ent-user"). ECC keys are
suffixed with "-ecdsa".

This naming change solves a problem that is exposed by the adding of an
Algorithm parameter to the chrome.enterprise.platformKeys.challengeKey
API (http://crrev/c/3983593, WIP). If a caller requests a user-scoped
ECC key to be registered, but the existing EUK is RSA, the existing RSA
key will instead be returned. This CL solves this problem by separating
storage of RSA and ECC EUKs.

BUG: b:252786598
Change-Id: I68dab0fb0d84bcb4041978f26af96237113e24ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4044045
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Kevin King <kcking@google.com>
Reviewed-by: Michael Ershov <miersh@google.com>
Cr-Commit-Position: refs/heads/main@{#1080374}