[UPM-Desktop] Password export - add functional export.
This CL allows to export passwords and closes the export dialog
automatically if the export succeeds.
Error handling and progress bar are to be added in follow-up CLS.

Some WebUI code is copied and tests from the old UI in settings (that
will be deprecated after the new UI is launched), sharing the code
between settings WebUI and Passwords WebUI is unfortunately not
possible, because we need to modify it slightly and deal with
PasswordManagerProxy class being different in two WebUIs.

Bug: 1394416
Change-Id: I2d1615e4135d726c2187b8ba7e11ddf112d32b71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4079110
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Commit-Queue: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#1080388}