Persist tab thumbnails through discards
This CL introduces a mechanism for WebContentsObservers to be notified
when a WebContents is about to be discarded.

The first user of this mechanism is ThumbnailTabHelper, and it uses it
to move the thumbnail data over to the new WebContents to be able to
keep showing thumbnail data in the tab strip even though a page is
discarded.

Bug: 1336798
Change-Id: I7ac8994dd1deea873665a6d37377840a33d89234
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3991486
Reviewed-by: Alison Gale <agale@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080390}