ash: Change window stack method for overview drag
When the drag of an overview item is started, the window, along with the
item widget, are brought to the top of the desk container. This is to
ensure that they appear on top of the desk bar, as well as other
windows. The code was previously only stacking if the item widget was
alive *and* the item widget and the overview item window shared the same
parent.

This should normally be the case, but there may be some odd edgecase
that results in the stacking not happening. This appears to have been
the case for b/260755896.

This change removes the condition so that the overview item window is
always stacked on top.

BUG=b:260755896

Change-Id: I2b9dbdf90216b1f39832a88b524c70af8d3cbbf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4083931
Commit-Queue: Daniel Andersson <dandersson@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080395}