Change updateScrollableContents to be event-driven for initial resizing
There has been a long-standing issue crbug.com/657336 where for the
initial resizing of an iron-list, updateScrollableContents polls on a
10ms interval indefinitely if an iron-list is not rendered (for example
when the settings section it is in is not active). Today this affects
the startup sites list and the tab discard exceptions list in the
performance settings.

This CL changes cr-scrollable-mixin to use a ResizeObserver to know when
the parent element of an iron-list changes it's scrollHeight, instead of
polling it on an interval. The added test is for the case when a list is
initially hidden, and then becomes visible, and it passes for both the
old and new versions of the mixin.

Fixed: 1392785, 657336
Change-Id: Icbb198ee5368c2038bd92cbd95c92d095e64615b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4072125
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Charles Meng <charlesmeng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080422}