Allow share-this-tab-instead for self-capture
If the application specified {preferCurrentTab: true},
we detect that the current tab is a reasonable choice.
We therefore expose the button that lets the user
switch to sharing the current tab.

Note that for many applications, choosing the current tab
is undesirable. For example, in the context of
video-conferencing applications, it would often produce
a "hall of mirrors" effect.

If the application specified {surfaceSwitching: "exclude"},
that takes precedence, and the share-this-tab-instead button
will NOT be exposed.

Bug: 1395479
Change-Id: Icc20ce7dde023296cc654782f3a1a919ea1cdd67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4076374
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Auto-Submit: Elad Alon <eladalon@chromium.org>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Elad Alon <eladalon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080471}