Only allow the legacy process launcher format for Chrome `cmd`
The legacy command layout format is deprecated with this CL, because it
can be used to interpret and/or execute unrelated registry entries. The
only legacy process launcher format that is now allowed is for versions
of Google Chrome 110.0.5435.0 and lower with a command id of `cmd`.

Bug: 1384737
Change-Id: I8d9bcda47c1dc02c34277c8dd6d3a91314a30694
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4084521
Reviewed-by: Xiaoling Bao <xiaolingbao@chromium.org>
Auto-Submit: S Ganesh <ganesh@chromium.org>
Commit-Queue: S Ganesh <ganesh@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080486}