immersive fullscreen mac: overlay widget hit test
When the overlay widget and the browser widget overlap, right clicks
are not being handled. This results in the tab background context menu
not being displayed when "Always Show Toolbar in Full Screen" is
disabled.

This CL adds a hit test for the overlay widget and returns HTNOWHERE to
route the event appropriately.

Bug: 1369641
Change-Id: I19e47037815398fc4f69d3fe8781b8bc3ac6d49e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4080541
Commit-Queue: Tom Burgin <bur@chromium.org>
Reviewed-by: Keren Zhu <kerenzhu@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080488}