Add Linux UpdateService launcher
Launch the UpdateService process as a child of the client when
necessary. Similar to the change for UpdateServiceInternal. This will
not work for system level installs invoked across a user boundary but
is required to continue prototyping other aspects of the Linux update
client.

Additionally, the RPC unittests for UpdateService have been refactored into multi-process tests.

Bug: 1375890
Change-Id: Ie457116484f6353051bf04af21c29aa2a6652884
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4081450
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080574}