Add the shadowrootmode attribute, and trigger streaming only if present
This CL adds the new "shadowrootmode" attribute for use with
declarative Shadow DOM. See the discussion at [1] for more details,
but this will be the new standardized attribute to enable DSD.
This CL triggers the other new behavior, streaming, based on the
presence of this new attribute. In this way, the old behavior and
the new behavior can both co-exist while developers migrate from
the old attribute to the new one, and behavior will not change
until they opt in.

The virtual/streaming-declarative-shadow-dom-disabled test suite
was enhanced with a copy of all of the "old" tests using the
"shadowroot" attribute, as they were before [2] landed. In this way
the virtual suite maintains the functionality of the old syntax.
This virtual suite also runs the "new" tests, to ensure no inadvertent
changes happen for the flag-disabled case.

[1] https://github.com/whatwg/html/pull/5465#discussion_r1026760650
[2] https://chromium-review.googlesource.com/c/chromium/src/+/3967571

Bug: 1042130,1379513,1396384
Change-Id: I5a888287b7f313d13dd28c8b4b0e36ae4feea512
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4039504
Commit-Queue: Mason Freed <masonf@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080576}