Implement activity bits for Linux user-scope
Support activity bits by providing implementations for the stubbed out
util functions. On Linux it is not safe to assume that all of the home
directories for users on a system are subdirectories of /home/. Thus,
a system-scope implementation will require more research.

Bug: 1384968
Change-Id: Ibb8bc2f123502d856194fe24421e9f8942cc84d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4080203
Reviewed-by: Xiaoling Bao <xiaolingbao@chromium.org>
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Cr-Commit-Position: refs/heads/main@{#1080592}