infra: Apply COMMENT_LEVEL_RESTRICTED to all chrome trybots
This prevents the CQ from posting snippets/summaries of build failures
to gerrit for CLs that have an internal chrome trybot treated as a
mandatory builder for chromium/src CLs. It will instead print a
comment to gerrit that simply says "this build failed".

This is to prevent logs/traces from internal bots being posted on
public CLs. This is considered non-negotiable, and is applied to any
chrome trybot registered in chromium/src.

Bug: 1259887
Change-Id: Idd158071b5235c620f10ca60eabc857959cf47ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4086690
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080595}