[BREFs] Apply RuntimeFeatureChange diffs to *ReadContext
When we receive a map of modified feature values via IPC from the
renderer, we should be able to alter the state found in the browser's
RuntimeFeatureStateReadContext to reflect these new values. This
CL makes sure that this ability is limited to only this IPC case.

Change-Id: I8eed3cc1bcace83eb54bda097220b2aca1a73449
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4081379
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080600}