Handle high max buffer_size in AudioLatency::GetExactBufferSize()
Previously AudioLatency::GetExactBufferSize() was expecting that
max_hardware_buffer_size is always less than any max_allowed_buffer_size
that can be passed to this function. But that assumption is not always
correct. For example AudioManagerFuchsia sets
max_hardware_buffer_size=48000, which is more than kMaxWebAudioBufferSize
(i.e. 8192). In that case the function was failing DCHECK in debug
builds and returning 0 in release builds. Updated it to handle this case
correctly.

This partially solves the problem described in crbug.com/1395234.

Bug: 1393243
Bug: 1395234
Change-Id: Ic4a696fae0ec70af34accde785510bf0ecf0db24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4054038
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080624}