[GridNG] Remove unnecessary parameters from NGGridLineResolver
When the line resolver was static, it was necessary to plumb auto repeat
and `subgrid_span_size` values all the way through from the layout
algorithm. But now that the line resolver is a member of
`NGGridPlacementData`, we can use member variables upon construction and
query those instead. This also allows us to remove some duplicated
properties from `NGGridPlacementData` and instead query the line
resolver.

This will make it much easier to handle auto-repeats for subgrids.

Bug: 618969
Change-Id: I0a98833198867d2599e1385a46325809c570436c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3990256
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080667}