[Test Automation] Allow passing window hint to ui_controls mouse actions
There is a race condition between window activation/key status on Mac
between appkit and windowserver, which means the current way of
determining where to route mouse messages when using
ui_controls::SendMouseMove* and ui_controls::SendMouseEvents* sometimes
doesn't work correctly.

Most tests will know which window they want to send a mouse event to.
Because of this, we need a way to pass that window as an optional
hint. We add it as an optional parameter at the end of the ui_controls
methods, and it is used by the Mac implementation if provided (otherwise
the existing selection logic will be used instead).

This information is now available for other platforms, though this CL
does not further pipe it into UIControlsAura. We have created
crbug.com/1396661 to evaluate follow-up and determine whether some
Aura platforms could benefit from using the hint as well.

Bug: 1396074
Change-Id: I9c021c33646b397886f9f6bcc810941e946024ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4087381
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Jayson Adams <shrike@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080671}