Fix test failures caused by new SWAManager Demo Mode dependencies
crrev/c/4076557 introduces an IsDeviceInDemoMode check to be called when SystemWebAppManager is instantiated, which happens during TestProfile
creation. IsDeviceInDemoMode relies on InstallAttributes being
faked/instantiated, and the demo_mode.config pref being registered.
The latter is temporary, as the demo_mode.config pref is being
deprecated and all usages in code removed.

The ideal solution would be to just not check app enablement if we're in tests, since no SWAs get installed by default in tests. Unfortunately, all apps need to be checked for enabled status regardless of whether we're in tests (crrev/c/3750829), so we have to check if we're in Demo Mode to verify whether the Demo Mode SWA is enabled.

Test: CQ in child CL that enables DemoModeSWA feature
Bug: 1384427
Bug: b/255803718

Change-Id: I5fc6d0fc0e007a63edabda1a101a1ab396430341
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4084447
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Commit-Queue: Jackson Tadie <jacksontadie@google.com>
Reviewed-by: Alberto Juarez <albertojuarez@google.com>
Cr-Commit-Position: refs/heads/main@{#1080674}