Lift the DawnWireServices class declaration from .cc to .h.
I would like to use DawnWireServices from a fuzzer because it has a handy direct reference to the DawnClientSerializer. To keep things tidy it would be helpful to lift the declaration of DawnWireServices from webgpu_implementation.cc to webgpu_implementation.h, which is what this CL proposes.

Change-Id: I600458a2f0e11aa30e1ad65d76a6e4885b39a9bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4084073
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Chris Bookholt <bookholt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080689}