Files F2: Add keyedKeepFirst concurrency model
The keyedKeepFirst() concurrency model avoid starting a new
ActionsProducer if multiple requests with the same key are done.

It starts a new run if a request with a new key is done or the request
of the same key has already finished.

This will be used for FileTasks to keep the current behavior of only
requesting new FileTasks when requested with a new set of FileKeys.

Bug: b:228112698
Change-Id: Iceda83403464a84a0fc70fde7a67f560b2acabab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4083033
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080706}