Add --qs-add-fake-cast-devices to populate quick settings Cast UI
Passing this command line switch adds fake cast devices so that a
developer can work on simple changes to the quick settings UI without
having a Cast device on the local network. For example, this is
convenient for the linux-chrome emulator build.

Remove an existing hack that added a device scroll list item when the
QsRevamp flag was enabled. The switch does a better job of populating
the UI.

Bug: b:252872586
Change-Id: I9b781b1e088b6778d6b95082cf6ed5a97c91e4a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4086247
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080711}