Clarify where the flakiness comes from.
According to the flakiness dashboard, mac tend to be flaky for UKM tests
and fuchsia tend to be flaky on UMA tests.  Others seems not be affected
so much.
Let me clarify which platform is flaky for what so that we can run tests
as much as possible.

Bug: 1395715
Change-Id: Ibcffc3604d176e2f15af0831fbcba28b18f224d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4082137
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080730}