infra: Bump CQ experiment of linux-chromeos-compile-chrome to 100%
After a surge, the pool appears to be about 50% utilized after
bumping the CQ experiment to 50%:
http://shortn/_Yrl2V725xz

So perhaps the bot pool has *just enough* machines for 100%. (Will
likely want to throw a few more in for safety if that turns out to be
the case.)

So this bumps it to 100% to test that out. Will probably see another
surge after this lands. So many want to land it during off-hours.

Bug: 1259887
Change-Id: I5afa89704100f0f3d92d38c15def7e55e24eaefe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4087326
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1080786}