Files F2: Move FileTasks static methods to functions
Move annotateTasks() and getDefaultTask() to plain functions to make it
simpler to re-use in the Store/Reducer.

These functions are used by the FileTasks.crate() and the reducer will
do a similar job to process the FileTasks to the Store. The same for the
constant `INSTALL_LINUX_PACKAGE_TASK_DESCRIPTOR`.

Rename method getTaskItems() to getAnnotatedTasks() to better reflect
the content returned by this method.

Refactor the parameters of updateContextMenuTaskItems_() to make it
easier to switch between the Store and non-store implementations.

Bug: b:228112698
Change-Id: Ic03fc16d8b77cd9b871bf822964ca28c2f24da52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4082948
Reviewed-by: Wenbo Jie <wenbojie@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080792}