[iOS] Delete HistoryStateOperationsTest.TitleReplacement
This was intended to test that calls to replaceState
correctly update the current NavigationItem's title. But this
was based on a misunderstanding of the behavior of replaceState.

replaceState's second argument was originally spec'd to be
a title, but no browser supported replacing the title this way.
The spec has since been updated to treat the second argument as
a no-op.

So this test is expecting replaceState to do something that
it actually doesn't do, which explains why this test has never
passed (it has been disabled since it landed).

This CL deletes this incorrect test.

Bug: 677356
Change-Id: Ide3df017ced7e0770f1fb4661746190f49aab8d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4087442
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Auto-Submit: Ali Juma <ajuma@chromium.org>
Reviewed-by: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080904}