[iOS] Fix TabGrid bottom toolbar visibility
In some cases, when moving in and out of the SearchMode, the visibility
of the TabGrid bottom toolbar was wrong.
Simplify the logic to make sure it is working as expected.

Fixed: 1399519
Change-Id: I1ea38193a23e16124c84ba62482575958186f04f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4088565
Reviewed-by: Louis Romero <lpromero@google.com>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1080947}