Modify the battery discharge granularity metrics.
This CL removes the version of the metric that reflected the capacity
of the battery in favor of Power.MaxBatteryDischargeGranularity, which
was renamed to Power.BatteryDischargeGranularityRelative, and fixed
the unit label.

This CL also adds another version of the metric that is reported in
milliwatt-hours.

Finally, the reporting code was moved from the PowerMetricsReporter
class to BatteryDischargeReporter, as the metrics are related to
the battery state.

Bug: 1191045
Change-Id: I1f282226e8a38f2f6758a47feb59723b56c46e2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4072725
Commit-Queue: Olivier Li <olivierli@chromium.org>
Reviewed-by: Olivier Li <olivierli@chromium.org>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080951}