Multiplanar: Add support to SkiaOutputSurfaceImpl::MakePromiseImage
For multiplanar shared images, we need to make promise image from
YUVA texture in SkiaOutputSurfaceImpl::MakePromiseImage. This follows
logic similar to MakePromiseSkImageFromYUV which won't be needed later
on for YUVDrawQuads as TextureQuads will be created instead for YUV
format as well. This follows that now for multiplanar formats we
fulfill texture per plane and also get the GrBackendFormat per plane
in case we need to create a fallback image and so have vector<>
of formats for creating fallback image.

This change also updates GetGrBackendFormatForTexture function to
take in a SharedImageFormat with plane index to get the appropriate
GrBackendFormat. This is not supported yet for Vulkan/Dawn and needs
further investigation (added TODOs), but Metal texture can be created
per plane and hence added support for it in SIFormat util function.

Bug: 1366486
Change-Id: Ib699e30574d0bbd01ea4b4556e7200f8e2f3d828
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4080197
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080980}