[iOS][ContentSuggestionHeader] Fix shift up animation
Content suggestion headers can be updated during the scroll, causing
`pinnedOffsetY` to be invalid. When this happens during the animation,
the tiles are not scrolled to the top causing the omnibox to be hidden
by the `PrimaryToolbarView`. In that state, the omnibox's popup and the
keyboard are still visible.

If the animation is not interrupted and `pinnedOffsetY` changed
during the animation, shift the omnibox to the top at the end of the
animation.

Bug: 1320052
Change-Id: I7d1f23ac34ca903f5964236da33642687e2c24ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4079149
Reviewed-by: Adam Arcaro <adamta@google.com>
Auto-Submit: Christian Xu <christianxu@chromium.org>
Reviewed-by: Quentin Pubert <qpubert@google.com>
Commit-Queue: Christian Xu <christianxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080993}