[Select to Speak] Speculative fix for flaky tests
Select to Speak tests started flaking more frequently after
crrev.com/c/4014940, which changed how the Select to Speak
dialog was forced to be not-shown. This tries another way
to ensure the dialog is not shown, so that speech will
not be blocked on dialog accept/cancel.

Bug: b:261774383
Test: Existing
Change-Id: Iac3dcebc9229557f2d0184fca3c39d3cf17fae3b
AX-Relnotes: N/A
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4085547
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/main@{#1081007}