Animate tablet mode search box from expanded to closed state
Show the search result page during the animation to closed instead of
hiding immediately. Also, fix the clip rect calculation so that the
closing animation looks correct.

Finally, add a unittest which ensures that the search result page is
shown and animating for the expanded->closed transition.

Bug: 1381599
Change-Id: I783b7d47d37d5be4d72a0220f3aa2651836ca3f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4018005
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081018}