[UPM] Extract store backend consumer bridge to a separate class
Split the PasssowrdStoreAndroidBackend JNI bridge in two parts
- native -> java bridge (existing bridge)
- java -> native bridge (consumer bridge)

This is a preparation for moving the GMSCore interaction out from the UI
thread. Such change will make all native -> java calls executed on the
background working thread while operations callbacks will be called by
the GMSCore on the UI thread. It is preferred to keep to not allow a
single object to live on several threads.

Bug: 1394715
Change-Id: Ibbfb40727e089fdb41ebac9529cdf44112d11360
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4079256
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Maxim Anufriev <maxan@google.com>
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081033}