Make ParseCertificate compatible with std::make_shared
We don't need a private destructor anymore. Unlike scoped_refptr,
std::shared_ptr makes the ownership store the caller's decision, not the
class's. If someone manages to[*] get an unshared ParsedCertificate,
there's nothing wrong with them freeing it.

Additionally, since std::make_shared is meaningfully different than
adopting a bare pointer in std::shared_ptr, use a pattern like
base::PassKey to make std::make_shared work. Except, as this code is due
to be extracted, we can't actually use base::PassKey, so I've just done
the equivalent thing manually in the class.

[*] It's not actually possible to get such a thing, because the
constructor is private and the factory functions all return shared_ptr.
It might be worth changing that, so some callers can embed the structure
directly rather than indirect with a refcount, but I haven't looked at
doing that here. If we were to do that, it could reduce also the places
this API leaks shared_ptr into the rest of Chromium. Something to ponder
as we work on this code.

Bug: 1322914
Change-Id: I8b32a37c0417d802cd6a14b183924bf411363066
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4084894
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081058}