Define a new CI waterfall called "chromium.cft".
This creates a new CI (continuous integration) waterfall
called "chromium.cft". It defines a new builder group for
the CfT (Chrome for Testing) subproject, and sets up
the first builder for it ("mac-rel-cft").

It also changes the mac-rel-cft tryserver to mirror the new
CI builder, rather than copying mac-rel exactly as it was
doing before. As part of this we also create a
tryserver.chromium.cft config.

Bug: 1396097
Change-Id: I19806163a0da527863309a3595197b916a140c70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4081509
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/main@{#1081078}