Extension Telemetry Persister Shutdown Bug
The Extension Telemetry Persister saves in-memory telemetry data to disk at shutdown. The check for the presence of telemetry data is incorrect however. This error causes a null report serialization attempt in ExtensionTelemetryService::Shutdown.

This CL fixes ExtensionTelemetryService::SignalDataPresent, the check for the presence of telemetry data. A unit test is also added to check this edge case.

Bug: 1398577
Change-Id: I88afb9456a25a7fcf6aa99be62678c9ff2919be7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4084886
Commit-Queue: Adam Psarouthakis <psarouthakis@google.com>
Reviewed-by: Anunoy Ghosh <anunoy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081083}