[GPU] Convert GLImage*::FromGLImage() to GLImage::ToGLImage*()
We are in the process of restricting GLImage's public interface to
ensure that new usage doesn't creep in while we are eliminating its
existing client usage. GLImage::GetType() poses a barrier to doing so:
- This method is used in subclasses' static FromGLImage() methods.
- Because those methods are static, moving GetType() to protected means
  that those methods can no longer access it.
- Solving that problem would require declaring those static methods as
  friends, which would require putting the definition of the
  subclasses *before* the declaration of GLImage itself. That is, of
  course, not possible.

To solve this problem, this CL does the following:
- Moves the 5 GLImage<Foo>::FromGLImage() methods to instead be
  GLImage::ToGLImage<Foo>(). This is possible while just
  forward-declaring the subclasses in gl_image.h.

This will then enable moving both GLImage::GetType() and
GLImage::FromGLImage<Foo>() out of the public interface in a followup
patch.

Bug: 1382031
Change-Id: If8d0cedacf71288605827d693eb8281f41667de0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4078919
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Auto-Submit: Colin Blundell <blundell@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081099}