Miss the cache when breaking before a fragment with low break appeal.
If a fragment doesn't fit in the fragmentainer, we should of course
break before it. This normally happens automatically, but in some cases
the machinery needs an extra nudge. FinishFragmentation() will in some
cases clamp the break (inside) appeal down to kBreakAppealLastResort.
We need to miss the cache if we find such layout results in the cache,
since the break appeal may no longer be valid, when we have moved over
to a new fragmentainer. Otherwise we will never let the node fit in the
previous fragmentainer, not even after making the fragmentainers taller.
This leads to correctness issues, and also freezes during column
balancing, since no fragmentainer stretch amount will be enough to pull
the node back to the first fragmentainer.

Bug: 1395408, 1399449
Change-Id: Ic0fc032b0b7022dfb9e5c35af75e09d8b3f50316
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4085506
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1081158}