cros_next: add knob switch component
This cl adds a new knob switch component. A knob switch is a
switch widget with a circle knob on a rounded rect track. When the
switch is clicked, the knob will slide from one side of the track
to the other side. The selected state will also be changed.

KnobSwitch class extends the views::Button class and allows clients to
define a switch callback when the selected state changes. For ease of
building the shape of knob and track, the cl adds a private background
class called ThemedFullyRoundedRectBackground which can paint a fully
rounded rect filling in its view's local bounds. The circle shape of
knob can be treated as a special case of fully rounded rect.

The final effects are attached in the bug comments and the unit test
will be added in the following CL.

Bug: b/260715410
Change-Id: I881c56fdcdef5e9f61207b857912761ecc188c45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4060951
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081256}