Fix challenge selection a11y issue
This CL fixes a challenge selection a11y issue where if the user
attempts to select a radio button using a screen reader, it does not
register it. This CL updates the registering to be when the check is
changed instead of when the view is selected, which will register both
clicks and screen reader focuses.

Based on discussion with pkasting@ and pbos@, the common pattern for
keeping the callback list subscription is to store it in a member vector
so that the subscriptions do not go out of scope.

Bug: 1370329, 1399353
Change-Id: I5633a34db0e54f4fe123aaf95085d9f203e00ce4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4087641
Reviewed-by: Siyu An <siyua@chromium.org>
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081267}