Retry: Start JapaneseSettings migration on connecting to IMEService.
This makes use of the impl in cl/491527557 to do the actual migration, as in go/e14s-jp-sys-pk-v2.

This sets the chromium prefs for use in the ChromeOS settings page. The plan is to do the conversion the other way - prefs to Mojom JapaneseSettings object on Japanese decoder startup at a later point.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4054261
Commit-Queue: Keith Lee <keithlee@chromium.org>
Reviewed-by: Curtis McMullan <curtismcmullan@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Cr-Original-Commit-Position: refs/heads/main@{#1080651}
Change-Id: I68a4c6bfbc776cc6ab410ea0623389526b23d530
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4091013
Cr-Commit-Position: refs/heads/main@{#1081339}