CCA: Add fallback logic when choosing default aspect ratio for photos
Previously, we always prefer 4:3 aspect ratio as the default one for
photo modes. However, there are some external cameras which only have
low resolutions for 4:3 but have higher resolutions for other aspect
ratio. As a result, the default photo resolution on such devices will
output photos with low quality.

This CL adds a fallback logic to select other aspect ratio if the prefer
one does not have a resolution which is equal to or larger than 720p.

Bug: b:238398331
Test: Manually
Change-Id: If5b9d2f8dd64429cf942609059f456a55d213418
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4081156
Commit-Queue: Wei Lee <wtlee@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081371}