Wait before launcing the BiDi Mapper
This is a quick fix to address the customer-facing issue when user asked for BiDi connection. The proper solution should be in waiting for the initial navigation to be finished.

Tested manually on mac by creating new sessions with `webSocketUrl` and making sure the sessions were created with the BiDi Mapper.

Bug: chromedriver:4295
Change-Id: I50a57ce2a5dfe4032417a3fd525fb6debdf1996e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4084864
Reviewed-by: Vladimir Nechaev <nechaev@chromium.org>
Commit-Queue: Maksim Sadym <sadym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081393}