[FilesTiering] Create pin_manager_ on remount of DriveFS
Currently the pin_manager_ is created when the feature is first toggled,
unfortunately this means the mojom::DriveFs* ptr that it is passed is
invalid if DriveFS gets remounted. To avoid this happening, tie the
creation of the pin_manager_ to the lifetime of the mount point. This
also adds changes that ensure the `search_query_` mojom::Remote is
property bound before trying to call any methods on it.

Fixed: b:261785512
Test: Deploy to DUT, restart Drive and starting bulk pinning
Change-Id: I51034900d1621b9f14a5f9dd854578887d5d1511
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4087589
Reviewed-by: Bo Majewski <majewski@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081397}