Expose latest RecoveryIds in chrome://cryptohome.
RecoveryId is a unique identifier of cryptohome recovery operation. It
will be visible to users for transparency (so that a user can check if
the local recovery ids on their devices match their ledger entries).
This cl uses a new dbus call for getting extended info on an AuthFactor
and exposes latest RecoveryIds in chrome://cryptohome for debug purpose.
Similar mechanism will be used to expose user's RecoveryId history in
Settings.

Bug: b:192786922
Change-Id: Ie630f35014e08738cddb5599853b853153ed8974
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4085025
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Commit-Queue: Maciek Slusarczyk <mslus@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081402}