History Sync: Make sure only "good" URLs are added to the database
History code has the concept of URLs that should or should not be added
to the history DB, via HistoryClient::CanAddURL / CanAddURLToHistory.
This is checked on the HistoryService layer.
Since Sync interacts directly with the HistoryBackend, it bypassed this
check, and thus could sometimes add "unwanted" URLs to the DB, if they
someone ended up on the server.
This CL plumbs CanAddURL() through to the HistoryBackend, so it can
reject unwanted URLs from Sync.

Bug: 1364576
Change-Id: I52ed0a083426a327303d458f4f178f0d07d1e63a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4085425
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081458}