Delete the long-disabled CertVerifyProcMacTest.MacCRLIntermediate test
This test hasn't worked in years, and we've now switched to the builtin
verifier on Mac. The old CertVerifyProcMac still exists as a fallback
for a little while longer, but we're not going to spend the time to
care about why this test doesn't work, just delete it.

Fixed: 671889
Change-Id: I36531c2e39df543a1a0c3bc622457d311a9673e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4090198
Reviewed-by: David Benjamin <davidben@chromium.org>
Auto-Submit: Matt Mueller <mattm@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081504}