Delete long-disabled CertVerifyProcMacTest.MacKeychainReordering test
This test hasn't worked in years, and we've now switched to the builtin
verifier on Mac. The old CertVerifyProcMac still exists as a fallback
for a little while longer, but we're not going to spend the time to
care about why this test doesn't work, just delete it.

Fixed: 867174
Change-Id: Ica2ce7650d21d5469b2f168f2df087f98e94775d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4091266
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Queue: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081533}