CloseAll: Windows continue to have parents after close
This CL fixes how we handle the asynchronous closing of windows in
CloseAll. Windows cannot remain in the closing desk's container as they
are being closed asynchronously because the closing desk's container id
may need to be reused immediately. We previously solved this by simply
unparenting the windows immediately after we start their asynchronous
close processes. However, the windows should still have parents when
they are finally being destroyed. This CL fixes that by instead
reparenting the closing windows to the
kShellWindowId_UnparentedContainer container. This means that the window
will still have a parent when it is being destroyed. We also add a unit
test to verify that the window continues to have a parent after the
asynchronous close process is initiated.

Bug: b:261892928
Change-Id: I7e8d8655a2c8fb3dbcc1556d2c5f926173c06796
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4090405
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Ben Becker <benbecker@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081553}