Fixing drag-and-drop issues on Mac in pages with out-of-process iframes
This commit fixes an issue with WebDragDestMac where the process id and
view id for blocking same-page drag-and-drop between out-of-process
iframes (crbug.com/666858) were not being cleared once the drag ended.
This caused WebDragDestMac objects to become "tainted" with a specific
process id, so that only drags targeting that specific process id in
that WebContents would work. In cases where the tainted process id
belonged to an inner frame, this would prevent the entire page from
accepting any drops at all. The fix clears the data when the drag ends,
allowing other iframes to accept drops again.

The call to isValidDragTarget is removed from the async
DropCompletionCallback, as it is not necessary. The isValidDragTarget
function checks the validity of the targetRWH as a drop target for the
specific drag operation, which has already been checked in the
performDragOperation function for the specific drop data and origin. As
such, the asynchronous nature of the callback does not affect the
validity of the content::DropContext in that regards. Additionally,
after the fix, the startDrag info may have already been cleared at the
time the callback is called, so the call to isValidDragTarget could not
remain in the code.

Bug: 1399569
Change-Id: Id4f4a4e7c98e04a7c55f2807f5fe35c0e8734020
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4088669
Commit-Queue: Avi Drissman <avi@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081578}