Fix paint holding on navigation
RenderWidgetHostViewAura::TakeFallbackContentFrom() was taking both the
fallback content and content rendering timeout. The call to
GetContentRenderingTimeoutFrom() causes issues on navigation as
RenderWidgetHostImpl::WasShown() resets [1] the fallback content if
there was a content rendering timeout already set. WasShown() runs
almost immediately after TakeFallbackContentFrom() so it sets the
fallback SurfaceId and then immediately resets it, causing only the
background color to show and thus a flicker of background color on
navigation instead of paint holding the old page.

It's worth noting that on navigation a new content rendering timeout is
set after WasShown() by RenderWidgetHostImpl::DidNavigate(). This causes
the fallback SurfaceId to be reset after 4 seconds as intended, making
the original call to GetContentRenderingTimeoutFrom() unnecessary even
if it didn't cause flicker.

This problem with paint holding applies to any case where the render
widget changes, so both cross-origin navigation (always) and same-origin
navigation (with BackForwardCache enabled).

RenderWidgetHostView::TakeFallbackContentFrom() is also used with
portals. When the portal renderer content is replacing the top-level
renderer content, TakeFallbackContentFrom() gets the top-level renderer
as fallback for portal renderer. We know the top-level renderer
has already loaded because it opened a portal which the user then
clicked on. The old top level renderer won't have a useful content
rendering timeout to inherit so removing
GetContentRenderingTimeoutFrom() should be fine for portals as well.

Make RenderWidgetHostViewAura match what RenderWidgetHostViewMac already
does [2] and skip the GetContentRenderingTimeoutFrom() call.

[1] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/render_widget_host_impl.cc;l=873;drc=d7202e3ac8c272c07bc5036cd32487d41d5edef8
[2] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/render_widget_host_view_mac.mm;l=1026-1027;drc=4d9f1c1280ee8474a61ff89260a81651c50b4fb6

Bug: 1338993
Change-Id: Ie2757a94537effb06f564c7cec118cb37c7583eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4057827
Commit-Queue: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081602}