Remove RenderViewHostChanged from Android WebView
RenderViewHostChanged is an old API which is deprecated and we would like
to remove. This CL moves the Android WebView code to rely on the newer
version which operates on frames - RenderFrameHostChanged. To minimize
difference in behavior, the new implementation does not consider any
frame which is not the primary main frame in a WebContents.

Bug: 1168562, 1394513
Change-Id: Id65ab408a31e0932ed463a3553bc69e00dbe439a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4087142
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081610}