Add ATTESTATION_NOT_AVAILABLE status
This new status will be used to indicate that attestation is not
supported by the device, so the attestation flow shouldn't be retried
when this failure occurs.

Nothing actually sets this status yet; this just updates all the
receivers to handle the new state.

Bug: b:254146401
Change-Id: Ib9abb24825d564d905aa702f9c02a6f607cc7dd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4061942
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081648}