Enable MediaCapabilities tests with type="webrtc"
Despite the fact that none of the formats used in these tests are
supported where `type="webrtc"`, we should still be exercising that
code path since it's the third possible value for that parameter.
In the future it would be good to add more tests (here or elsewhere)
where the expected return type is something other than "UNSUPPORTED"
for that.

Change-Id: I963a73a4a0ac5836a116ff5e2d45feb3defcc02d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4090626
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Johannes Kron <kron@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081665}