[NTP][Panorama] Add Themes page general styling for layout
Added the styling and html to support the layout of the page from designs. The text still needs i18n and the images still need to be filled with their respective pictures. This is just for layout.

The layout is uses 1 grid with 6 columns. The first 4 children of the grid span 3 so that there are 2 per row. The rest of the children span 2 so that there are 3 per row. To account for the wider space between the first 4 children, padding is added to the right side of the odd ones and left side of the even ones to add up to the difference in gap.

cr_grid was updated to utilize the css grid-gap property instead of setting --cr-grid-gap to both column and row. This way it doesn't hurt the pages currently using cr-grid while also giving the ability to set row and column gap to be different. Also added some more css variables for more control over the element styling.

screenshot: screenshot/9PYwV8HibMmeJgw

Bug: 1384251
Change-Id: Id83f16d24b980ce9a0406a95fa293ecc99541538
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4084451
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Riley Tatum <rtatum@google.com>
Cr-Commit-Position: refs/heads/main@{#1081687}