float: Fix crash when dragging a normal window to new desk
This crash can happen whenever dragging a normal window to another desk,
or onto the new desk button (also to another desk).

It happens because our scoped object that modifies stacking order for
drag, so that normal windows will appear on top of floated windows
during the drag, has a `Shutdown()` function that listens to the
window animation and deletes itself at the end.

When dragging an item to another desk, the window object is set to null
since it is removed from the current grid. There is no need to listen to
the animation, since to get to this state we would have to drag the
window above the desk bar, which means its animations will not hover
over the floated window, and we can restore the stacking order
immediately.

Test: manual, added regression test
Fixed: b/261757970
Change-Id: I2c56b6ac5e79d5eb64f15fe40c4cd5bc7aeb4b30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4089662
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081695}