Fix indenting in chromeos_strings.grd
This fixes the indentation of the <if block in chromeos_strings.grd.
When the entire block is indented by two spaces these presubmit warnings
are triggered:

You are modifying UI strings or their meanings.
To ensure the best translations, take screenshots of the relevant UI (https://g.co/chrome/translation) and add these files to your changelist:
  chromeos\chromeos_strings_grd\IDS_DIAGNOSTICS_NAME_SERVERS.png.sha1
  chromeos\chromeos_strings_grd\IDS_SCANNING_APP_EDIT_BUTTON_LABEL.png.sha1
  chromeos\chromeos_strings_grd\IDS_SCANNING_APP_FILE_SAVED_TEXT.png.sha1
  chromeos\chromeos_strings_grd\IDS_SCANNING_APP_REMOVE_PAGE_DIALOG_TITLE.png.sha1
  chromeos\chromeos_strings_grd\IDS_SCANNING_APP_RESCAN_PAGE_DIALOG_TITLE.png.sha1
  chromeos\chromeos_strings_grd\IDS_SCANNING_APP_SCAN_BUTTON_TEXT.png.sha1

This is because leading spaces in multi-line messages are significant,
and changing this is impractical. To avoid these warnings and to avoid
triggering translation requests the lines (after the first) in these
strings were pulled back to their original location.

Bug: 1394970
Change-Id: Ide621a619d3d97909ca7344428dce23c08ab497a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4090561
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081711}