[Files]: Adds a component for displaying the path of a file.
In order to allow the users to distinguish where the found files are
located, we are introducing a widget that shows the path of the selected
file. This change adds the element and modifies the search container to
active it when search is taking place.

Bug: b:4210716
Change-Id: Ib4a777714a6d94987c59afd66a01a5ea715d6c0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4067566
Commit-Queue: Bo Majewski <majewski@chromium.org>
Reviewed-by: Wenbo Jie <wenbojie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081761}