[Cast Streaming] Change the default frame refresh interval to 50ms.
Experimentation shows that reducing the refresh interval slightly
improves image quality, most noticeably after scrolling a text-heavy
page.

While there is some power impact when static content is cast for a
long period of time, it is not likely to be significant as encoding should be a no-op and frames should be small or empty.

This is put behind a default-enabled feature kCastFastRefreshFrames, which can be used to revert to the previous behavior if necessary.

The feature, the command-line flag, and the refresh_interval parameter
can all be removed after this launches to stable.

Bug: 1394392
Change-Id: Ie68a2b2bf86ee9fe2f1178686e06b47f7a65c7da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4093497
Commit-Queue: Mark Foltz <mfoltz@chromium.org>
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081783}