splitview: Fix bad divider position computation
This CL fixes a 1-DIP gap between snapped windows, which disrupts the
multiresize functionality. One specific scenario where this arises is
when the length of the work area is odd, which can happen on a real
device when the display scale factor is not 100%. Here is the problem:
  int snap_width = divider_end_position * snap_ratio;
  int next_divider_position = snap_position == SnapPosition::kPrimary
                                  ? snap_width
                                  : divider_end_position - snap_width;
snap_ratio is a float, and snap_width is rounded down because of the
implicit typecast to int. That means divider_end_position - snap_width
is effectively rounded up. Therefore, next_divider_position is rounded
down for a left snap, but up for a right snap. This inconsistency
means different divider positions for a left snap versus a right snap.
The fix is to make snap_width a float so that the rounding is performed
at the end of the computation of next_divider_position.

Fixed: b/262011280
Test: ash_unittests WindowStateTest.SnapWindowOddWorkAreaLength
Change-Id: I37ff40eaa29ee614208f87314e1af139912548e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4093936
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081784}