[omnibox][doc] Limit low quality docs.
Currently, the server determines which docs are shown and their scores.
Server suggestions are notoriously aggressive, often resulting in
showing 3 docs seemingly unrelated to the input. (see screenshot).

There're 2 use cases we want to support well:
1) Users use the omnibox to navigate to their docs quickly.
2) Users looking for a particular doc they've never visited before.
   E.g., they see someone presenting a doc during a meeting, and they
   want to open the doc locally.

This CL checks each doc whether
a) the user own the doc and
b) the input matches the doc title and owner (as opposed to inputs that
match the contents or contain typos).

If both or either are true, we continue showing the top 3 docs from the
server with their server scores. This handles case (1); even if the
input isn't exactly 'right', we're likely to include their desired doc.
This also partially handles case (2) since when looking for a doc the
1st time, users are likely to type the title and owner.

If neither are true, we show only 1 such doc. This too partially handles
case (2) in case the user input contains a mistake (e.g. typo or an
extraneous word).

Low quality docs still contribute to the overall limit of 3.

Ideally, we'd also allow docs the user's edited, commented on, or viewed
more than N times to likewise not be subject to the 'must be exact
title/owner input or will get limited to 1', but the server response
doesn't include such signals, so we rely on whether the doc is owned.

This CL also cleans up old scoring flags. It does not clean up the
scoring logic as we may still return to it. It also removes instances of
`SplitByColon()` as apparently `String16VectorFromString16` now splits
on colons.

Guarded behind `'DocumentProviderMaxLowQualitySuggestions'` param.

Bug: 982382
Change-Id: I7bf94a764069cb84305f851e0d342a23ed34b07a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4092642
Reviewed-by: Angela Yoeurng <yoangela@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081791}