[iOS] Fix keyboard dismissed on swiping from Find bar to type
When the Find bar is presented, VoiceOver correctly focuses the Find bar
text input field. However, when the user tries to swipe repetitively to
navigate to the keyboard, they have to go through the content area, at
which point the keyboard will likely be dismissed.

This CL solves this problem by replicating the technique used when the
omnibox popup is presented i.e. disable accessibility elements in the
content area when the Find bar is being presented. It does so by adding
two methods to FindBarPresentationDelegate:

- (void)findBarDidAppearForFindBarCoordinator:
    (FindBarCoordinator*)findBarCoordinator;
- (void)findBarDidDisappearForFindBarCoordinator:
    (FindBarCoordinator*)findBarCoordinator;

Since BVC is already the FindBarCoordinator's presentation delegate,
the coordinator simply lets BVC know when the Find bar is presented or
dismissed, so the BVC can disable unwanted accessibility elements. To
make the navigation through the onscreen-keyboard less confusing, the
secondary toolbar accessibility elements are also hidden.

This CL also sets the accessibility elements of the Find bar view
explicitly so the user can navigate to the Find bar buttons without
having to go through the search bar and share button. Finally, the
context string announcements are now slightly delayed, so they are not
immediately cancelled when the user taps the Previous/Next button in
the Find bar.

Video before CL
https://drive.google.com/file/d/15z7-eAFt0BXDKcwiPlDLxycEfoQasqOZ/view
Video after CL
https://drive.google.com/file/d/1aUClFq6z6eTE1Z4TnNpSLnEZOnZGYeWC/view

Fixed: 1370002
Change-Id: If840992e6f1d8c463bdf11ac0576762edcf3f728
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4075959
Reviewed-by: Louis Romero <lpromero@google.com>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081800}