arc: Show placeholder confirmation dialog for ARCVM /data migration
This CL introduces a dialog to ask the user's confirmation on ARCVM
/data migration. It is shown when the user clicks the "Update" button in
the notification introduced by CL:4023144.

The dialog explains that browser tabs and apps are closed when the
update starts, and asks the user to save all their work.

When the user clicks the "Start update" button, the migration status is
changed to kConfirmed, and Chrome is restarted. A migration screen to be
shown after the restart will be implemented by subsequent CLs.

When the user clicks the "Remind me later" button, nothing happens and
the user can continue their session. A notification will be shown again
in the next session.

Strings and icons are placeholder ones and will be replaced once the
final design decision is made.

Bug: b:258278176
Test: --enable-features=ArcVmDataMigration
Test: ash_components_unittests \
Test: --gtest_filter="ArcVmDataMigrationConfirmationDialogTest.*"

Change-Id: Ia4df2584d4c95b4b7283e4c6ab6568c2b680f245
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4022655
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081827}