Android PIP API: Fix pip for custom tab.
This CL fixed the issue where pip cannot be launched in custom tab.
The issue happens because we was checking the if the initiator tab is alive by finding its activity in app tasks. However, this doesn't work in custom tabs. Instead, we just need to check if the initiator tab and its activity is still alive.

Also removing Intent.FLAG_ACTIVITY_NEW_TASK flag, this is making the launch into pip feature not working properly for android T+. We can safely remove this flag because Android will move the activity to a new task when the activity entered picture-in-picture mode.

Bug: 1381672
Change-Id: I3d9723b261d5db92a780bacad8cbc08eb97fa060
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4092055
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Auto-Submit: Jazz Xu <jazzhsu@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081864}