[Kartoffel][Clank] Add the "More" button for Consent EEA
This CL adds the "More" button for Consent EEA, this button will be
shown when the content area is bigger than the scrollable area (Not all
the content is visible due to the screen size).
Clicking the button makes the consent to scroll down by one page; once 
the bottom is reached the button disappears.

Screen recording: crbug.com/1378703#c127

Bug: 1378703, b/61711738
Change-Id: I61f40e715c820e2eaa50f412c99fe15157370ab0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4085022
Reviewed-by: Rubin Deliallisi <rubindl@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Nicola Tommasi <tommasin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081892}