Fix FilterOperations::MapRectReverse() with blur filter
The MapRectReverse() function with blur filter was shrinking the bounds
rather than expanded them. This function is supposed to provide the
bounds needed as input to the filter and a blur filter requires pixels
outside it's bounds as input. The only usage of MapReverseRect() had a
special case to call MapRect() for blur filters, likely due to this bug.

Fixed: 1385154
Change-Id: I5fc930dff82516d1cc1b5fe07f4c4d0e59faace5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4083521
Commit-Queue: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Cr-Commit-Position: refs/heads/main@{#1081920}