Remove RenderViewHostChanged from VrWebContentsObserver
RenderViewHostChanged is an old API which is deprecated and we would
like to remove. This CL moves the VrWebContentsObserver code to
rely on the newer version which operates on frames. The code uses only
the RenderWidgetHost, which should be equivalent between the
RenderViewHost and the RenderFrameHost for the primary main frame.

Bug: 1168562, 1394513
Change-Id: I1220244bfeee1b9a6ad91439e62c745db105bb18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4092767
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081923}