float: Chrome window bounds are incorrect after entering tablet mode
The window seems a bit smaller, and is chopped of at the right edge if
a floated chrome window is transitioned to tablet mode.

This was happening due to the bounds getting changed in
`TabletModeWindowState::AttachState`, which gets called first, and
then the top chrome behaviour changes - bigger buttons for easier
tapping and therefore a larger window minimum size. In
`AttachState`, the bounds are still using the clamshell smaller window
minimum size.

The fix is to skip bounds update in `AttachState`, and delay to later
in `OnTabletModeStarted`, which runs after the top chrome behaviour
changes.

Fixed : b/261780362

Test: added test, manual
Change-Id: I6fa59c244b5ac29ab61a251ca5e5c179da7b48b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4089888
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081942}