Don't enqueue FrameTokens when no callbacks
PresentationTimeCallbackBuffer is given lists of callbacks to associate
with a frame.

Normally these fire upon receiving PresentationFeedback from Viz. Which
happens regularly.

We are investigating throttling OnBeginFrame when Viz is backed up, and
the last frame has not yet been Acked. This exposed a flaw in where we
enqueue the FrameTokens even when there is no callback to run.

In blink_web_tests there are two separate BeginFrameSources. One is the
standard Viz source. The other is a manual source ran by the tests in
SingleThreadProxy::CompositeImmediatelyForTest. This results in tests
being submitting frames regardless of Viz side back-pressure.

In these tests we can hit the
PresentationTimeCallbackBuffer::kMaxBufferSize of 60 FrameTokens.
Before Viz has responded with an OnBeginFrame. However only 2 of these
actually have real callbacks.

This change updates PresentationTimeCallbackBuffer to not enqueue
anything when there are no callbacks.

Bug: 1396081, 1399880
Change-Id: Id7bddd2253e0eec44c7510501579bdaf78829b28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4092899
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081970}