Check focusable cells for actionable nodes
This bug was found via a problem with the netflix app.
Prior to this change a focusable cell would never traverse to it's
children even if it has buttons inside of it. The result is that
certain buttons / other actionable nodes could never be reached.

AX-RelNotes: ChromeVox Users should now be able to reach nodes
within focusable cells if the nodes are actionable.

Bug: b/258768297
Change-Id: I8675531a7c3476dfad1d622eb316aef181ce522e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4093499
Reviewed-by: Anastasia Helfinstein <anastasi@google.com>
Commit-Queue: Jacob Francis <francisjp@google.com>
Cr-Commit-Position: refs/heads/main@{#1081975}