Fix memory destruction order in TestActivityTracker()
Memory owned by a sub-class can't be held in perpetuity in a raw_ptr
by its base class. The solution is to avoid inheritance, and just
make the former base-class a member.

This fixes one use case leading to an obvious dangling pointer, but
it is not likely the only one present.

Change-Id: Ic89d0672415ae1cc496773e1d8630df1eaa70bdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4093739
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081982}