[StartRefactor] Notify Start showing/hiding status without populating the StartSurfaceState.
Before the refactoring, the StartSurfaceMediator calls notifyStateChange() in setStartSurfaceState(), which will be removed after the refactoring is enabled. This leads to backPress handler no longer gets the state updates. In this CL, we split the notifyStateChange() and add the missing call when showing the Start
surface homepage.

Bug: 1347089
Change-Id: Ic7c241be37a79c9c242b74549eb39e0fc09cd390
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4076746
Commit-Queue: Xi Han <hanxi@chromium.org>
Reviewed-by: Lijin Shen <lazzzis@google.com>
Cr-Commit-Position: refs/heads/main@{#1082015}