personalization: user-select none
When using a touchscreen and selecting a wallpaper image from the grid,
it is very easy to accidentally drag to highlight. The user-select blue
overlay is close in color to the wallpaper image selected blue border.
To prevent confusion, set user-select none on images that appear in
grids. Also applies to screen saver albums.

BUG=b:259416992
TEST=click in whitespace and drag to highlight, images don't turn blue

Change-Id: Iccb80aa4994b909691c698def11249a44ec22347
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4086027
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Commit-Queue: Jeffrey Young <cowmoo@google.com>
Reviewed-by: Tao Wu <wutao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082023}