[Extensions] Properly allow localhost sources in extension CSP
Extension CSP should allow localhost sources as a viable entry, but
currently do not because the minimum CSP (which is always applied to
extension pages) does not include them. Fix this by adding localhost
and 127.0.0.1 to the minimum CSP for extensions, and add a regression
test.

Bug: 1247690
Change-Id: I99d97551dd6d7fe95aae5a81ebdcac6c6f573497
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3968883
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082082}