[FilesTiering] Start and stop pin manager at the correct spot
The pin manager was previously stopped on the OnUnmounted method,
however, this doesn't fire always. To ensure the pin manager is removed
at the appropriate time, move it to the RemoveDriveMountPoint. This also
makes sure to start the `DriveFsPinManager` when it is enabled (and stop
it on disable).

Bug: b:261530520
Test: Deploy to DUT and enable bulk pinning
Change-Id: I00639b59894604de1e5ee7969865be8d098c31f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4096139
Reviewed-by: Bo Majewski <majewski@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082109}