webauthn: enable macOS BLE permission prompt for Ventura.
macOS 13 (Ventura) switched to System Settings and some of the URLs
used to open System Preferences to a specific page didn't work for
System Settings.  Because of this, the WebAuthn BLE prompt for
Bluetooth permissions was disabled on macOS 13.

This change updates the URL for Bluetooth permissions to work on macOS
13 and enables the WebAuthn prompt for macOS 13.

Bug: 1400159
Change-Id: Iee05bbc316513fefafc4f3ca86f491bca4146a97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4081245
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082153}