autocorrect: record enabled by default to enabled
Previously in crrev.com/c/4054752 we added logic to record whenever a
user transitions to the autocorrect "enabled by default" bucket for the
first time. We also recorded the transition from "enabled by default" to
disabled, but not "enabled by default" to enabled. This was based on the
assumption that the user would be switching autocorrect on and off via a
toggle. There is work being completed to switch to this toggle for the
autocorrect setting, but we should also support the case where users
will be interacting with the dropdown currently used for the settings
(off, modest, aggressive).

In light of the above, this CL adds the final possible transition a user
is capable of completing with their autocorrect preference, moving from
"enabled by default" to explicitly enabled (they set their dropdown to
"aggressive" as the "enabled by default" state will show as "modest" in
the autocorrect setting dropdown after crrev.com/c/4078669 was landed).

BUG=b:257146732

Change-Id: I108410121b89df7a8b7ceac4803eef96169384e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4087802
Reviewed-by: Darren Shen <shend@chromium.org>
Reviewed-by: Mehrab N <mehrab@chromium.org>
Commit-Queue: Curtis McMullan <curtismcmullan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082173}