[Fast Pair] Add pairing protocol to WriteDeviceToFootprints
Adds an optional field to WriteDeviceToFootprints for the device pairing
protocol, which is used in a follow up to log the final state of the
retroactive pairing flow. We will eventually need to add logic to
PendingWrites to also log successes of writing to Footprints, but this
is a P2 since its a rare scenario and not launch blocking like the
general retroactive flow is (tracking b/261917790)

TESTED= No unit tests since refactoring, no new added logic. Protocol
used in unit tests in follow up CL. Testing on DUT in follow up CL.

Change-Id: I0642dc72faf38964445042e133ca165aaa3af0fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4090008
Reviewed-by: Daniel Classon <dclasson@google.com>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Juliet Lévesque <julietlevesque@google.com>
Cr-Commit-Position: refs/heads/main@{#1082188}