[omnibox][grouping] Keeps 2nd column matches in zero-prefix state.
Repurposes the default-disabled feature to keep all suggestions with
headers to instead keep all suggestions in the 2nd column. This is not
the desired end-state. Though it unblocks prototyping and experimenting
with the 2nd column UI (Zarya) in NTP realbox. To enable ZInspire on
Android the appropriate page classification still need to be added.

Also removes build flags guarding the Adaptive feature on Mobile. Those
are not necessary and only make the code harder to read. Adaptive is
already guarded by a feature flag.

Also organizes the code (no functional changes) to better reflect
different treatment of typed and zero-prefix cases with respect to the
grouping framework.

Bug: 1395416
Change-Id: Ib3e16c66b5eca74ac4008092230968d49cabf919
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4077274
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1082220}