Fix volume down behavior
Check mute level first and adjust the volume later.
Also avoid to update the output_volume_ directly.
Defer updating the output_volume_ to SetOutputNodeVolumePercent.

When you decreased the volume from 4 to 0, it would mute the device.
The behavior cause applications cannot decrease the volume of a single
device to zero without muting all other devices.
The correct behavior is for the volume to decrease to zero and then mute
when you press another volume down key.
The root cause is DecreaseOutputVolumeByOneStep updates output_volume_
directly when you decease volume from 4 to 0.
Therefore IsOutputVolumeBelowDefaultMuteLevel will become true since the
volume is 0.

BUG=b:259872514
TEST=Deploy to DUT and manually test with volume down key.
Low-Coverage-Reason: The feature flag is default enabled,
the code in the else part will be abandoned in the future.

Change-Id: Id3a2de1c1058d490485efecd4600263f6fb9bb13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4079403
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Ching Yun Chang <whalechang@google.com>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082259}