Init supported_formats with an initializer list
This is a minor cleanup CL which uses an initializer list to set the
vector of supported video formats in the WebrtcVideoEncoderFactory
instead of emplacing each one individually in the c'tor. This lone
exception is H.264 which requires a function call to determine whether
it is supported.

Change-Id: Ibb2eca293cf082340576ded93cedc743504e9cdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4098650
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082273}