[Lacros] Add log to detect Lacros window activation inconsistency
go/lacros-getting-stuck-minimized-or-activated

Though seemingly unrelated, we received two reports that may stem from the same root cause regarding the Lacros activation states.

#Lacros stuck in the activated state

In crbug.com/1284930, we see that there exists a set of steps that can cause Lacros to be trapped in the activated state. The reproducing steps are clearly stated in the bug description.

#Lacros stuck in the minimized state

In b/256952679, we see that there also exists a (yet to be known) set of steps that can cause Lacros to be trapped in the minimized state.

#Why these two are related

In the first issue, getting stuck in the activated state, the fix CL is to patch IsBrowserActive() of BrowserAppInstanceTracker and return true under a special case. Interestingly, if we return true under all cases (i.e., hardcoding return true inside IsBrowserActive()), and we build and deploy Lacros on a DUT, we will be hitting the exact issue as in b/256952679, that is, Lacros will always get stuck in the minimized state unless we alt-tab our way out.

Another proof that these two issues are interconnected is that, in both issues, once we get stuck, we will be able to observe a difference between two values maintained by two systems, that should stay the same the whole time. More details below.

#How do we tell if we get stuck

Currently there are two systems that track the activation statuses of the Lacros windows. One is the views::widget::IsActive() from within the Ash process, the other is BrowserWindowInstance::is_active who mirrors the internal states of  BrowserAppInstanceTracker from within the Lacros process.

In either issue, be it stuck in minimized or activated, the readings of activation statuses from the two systems will become different. As a result, we can leverage this difference to monitor the occurrences of Lacros getting stuck in either minimized or activated state.

#What needs to be done

We still are not able to reproduce b/256952679, even though through the aforementioned hardcoded trick we can reproduce very similar behaviors from local. By adding an error detection log as proposed in crrev.com/c/4056599, we will be able to tell from future user reports whether users hit relevant issues again, and if so, learn from those reports and figure out reproducible steps.

Bug: http://b/256952679
Change-Id: I6b579afd8dae243ed9ee6c8176835ce2fb4fccce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4056599
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Di Wu <diwux@google.com>
Reviewed-by: Alexander Bolodurin <alexbn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082275}