Introducing PostProcessing Layer.
Client will specify in the OutputConfig the format in which they want
output to be. PostProcessor Layer provides the results from the model
evaluation to the client in the format they specify.

This CL implements the post processing logic based on the type of
classifier.
 - Implements GetClassificationResults.
 - ClassificationResults could eithier be from BinaryClassifier,
   MultiClassClassifier or BinnedClassifier.
 - Tests

Future CLs will contains:
- GetRegressionResults  implementation.
- Integration GetClassificationResults API for client to use
  PostProcessing Layer.

Change-Id: Ia208a474e149b1e28644a87c43337a450643a4c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4073467
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Ritika Gupta <ritikagup@google.com>
Cr-Commit-Position: refs/heads/main@{#1082304}