Add vertical offset to text selection magnifier.
Use a background filter offset to allow text selection magnifier to be
displayed at a slight vertical offset to the text caret position. In
order to do so, this CL adds a cc::FilterOperation type for offset
filters and renames the FilterOperation drop_shadow_offset_ to a more
generic offset_.

Note that damage computations for the filter operations are not yet
handled correctly, see crbug/1379125.

Bug: b:244116654
Change-Id: I87e5313495f9460f75ebfd6f051d1a79c8b96f4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4005088
Reviewed-by: Darren Shen <shend@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Commit-Queue: Michelle Chen <michellegc@google.com>
Reviewed-by: Sam McNally <sammc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082322}