Reduce extension download retry backoff while Kiosk is launching
When there is a communication issue with CWS extension downloader will
retry with exponentially growing backoff, which blocks Chrome OS Kiosk
launch flow to maximum about 30 minutes (progressively longer backoff
periods of up to 10 attempts, with the maximum backoff between attempts
of 10 minutes). We'd like to set a cap on the backoff duration so that
the downloader retries more frequently and fails faster while Kiosk
is launching.

Bug: b:231914992
Test: unit_tests and manual
Change-Id: Idf56b6fcf651c3aa2da0fee1aec13cae069ca0dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4039913
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Yi Xie <yixie@chromium.org>
Reviewed-by: Ben Franz <bfranz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082326}