[NTP][Panorama] Connect Classic Chrome UI to page handler
I haven't added the image of classic chrome to the tile. I'll add it in later after the CL goes through for adding the same image to the preview on the first page of the Customize Chrome side panel UI. That way I can reuse it.

Also, updated the image container for dark mode and updated focus state
for the tiles to match  NTP.

screenshot of the focus state: screenshot/AeyaQHab2GcZ2ZV

Bug: 1384245
Change-Id: I50e3e5aa711df4ea75b5823b93895b5d2814f9bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4093360
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Riley Tatum <rtatum@google.com>
Cr-Commit-Position: refs/heads/main@{#1082329}