[Kartoffel][Clank] Add more button for Notice EEA
The CL adds the "More" button for Notice EEA, the button will be
shown when the content area is bigger than the scrollable area (Not all
the content is visible due to the screen size).
Clicking the button makes the consent to scroll down by one page; once
the bottom is reached the button disappears.

Screenrecord: crbug.com/1378703#c135

Bug: 1378703, b/262211601
Change-Id: Ib152f58049e2fef47e71dba9c7907f397ef6cb0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4092244
Commit-Queue: Nicola Tommasi <tommasin@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Rubin Deliallisi <rubindl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082332}