[Android][Privacy Guide] Rename test for UA in Sync's back button in PG
This CL changes the name of the test for the user action on the back
button of the History Sync card in PG from "testSyncCard_backButton" to
"testSyncCard_backClickHistorySyncUserAction". This change is done so
the test name is consistent with the others on PrivacyGuideFragmentTest.

Bug: 1238896
Change-Id: Ie556cdda9666f8fdeb659e3e0eec21e54bbbbdde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4092606
Commit-Queue: Filipa Senra <fsenra@google.com>
Reviewed-by: Andrey Zaytsev <andzaytsev@google.com>
Cr-Commit-Position: refs/heads/main@{#1082367}