[Android][Privacy Guide] Generic method to trigger metrics on next PGMDT
This CL creates a generic method to trigger metrics on the next click in
PrivacyGuideMetricsDelegateTest. This CL also makes use of this method for the tests of the MSBB card on the next click. The tests for the other cards - Sync, Safe Browsing and Cookies - will also make use of this method.

Bug: 1238896
Change-Id: I0916063c902684ffde51c87ccedf58ab26e412cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4094418
Reviewed-by: Andrey Zaytsev <andzaytsev@google.com>
Commit-Queue: Filipa Senra <fsenra@google.com>
Cr-Commit-Position: refs/heads/main@{#1082371}