[TTFCC] Reorganise the content of the sheet
The main layout of the BottomSheet is reorganised so it can be more
usable or very small screens. The icon and the title of the sheet are
added to the same RecyclerView so they will be scrolled together with
the card items (and also the button in the future CL).
The two views at the bottom (for scanning a card and managing the payment methods) are added to a footer element.
This change will make the height calculation easier.

Screenshot: https://screenshot.googleplex.com/6LS5DG9rXnEQTdJ
HSV link: https://hsv.googleplex.com/6283660136284160

Changes:
- The main layout of the bottom sheet is reorganised: there is no more
image because it's moved to the header item layout and the footer holds
the bottom two element
- A new layout is added to represent the header item (contains an image
and a title)
- A new item type is added for the header

Bug: 1247698
Change-Id: Ib1eb6def542af7bde0d69c96ba7edf8337f53ea4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4084867
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082378}