[Kartoffel][Clank] Add "More" button for Notice ROW
The CL adds the "More" button for Notice ROW, the button will be
shown when the content area is bigger than the scrollable area (Not all
the content is visible due to the screen size).
Clicking the button makes the consent to scroll down by one page; once
the bottom is reached the button disappears.

Screenrecord: crbug.com/1378703#c136

Bug: 1378703, b/262227341
Change-Id: I09b402f41c345aaa6f3e76edc80b1b6564f368ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4091927
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Nicola Tommasi <tommasin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082411}