[MSC] Cleanup in capture_policy_utils.cc.
This CL merges the `BUILDFLAG` checks for ash and lacros to a
`BUILDFLAG` check for ChromeOS. This CL also replaces the usage of
`GetWebsiteSetting` in `IsGetDisplayMediaSetSelectAllScreensAllowed`
with `GetContentSetting`. This is a more canonical way of checking
permissions as it implements the interpretation of the values in
the setting and maps it to `ContentSetting`.

Bug: 1300883
Change-Id: If16a50eee2913e8f089e9a8f7b7d8751102a9f46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4092349
Commit-Queue: Simon Hangl <simonha@google.com>
Reviewed-by: Elad Alon <eladalon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082419}