Add UMA tracking failed responses to audio service GetDeviceDescriptions
Add a quick histogram to track the rate of times the audio service
successfully returns the list of audio devices, vs the number of times
it fails eg due to crashing and an empty result is propagated up to
MediaDevicesManager et al.

A fuller treatment which actually returns a result code might look
something like crrev.com/c/4097129, but that's getting pretty big, and I
want to get some stats in before the 110 branch on Thursday.

Bug: 1400480
Change-Id: I622a07ad8a831fb718614a0c68d32a4d5ed2d960
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4098905
Reviewed-by: Olga Sharonova <olka@chromium.org>
Reviewed-by: Johannes Kron <kron@chromium.org>
Commit-Queue: Tony Herre <toprice@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082439}