ash: Fake effects and effects-management enhancements
SimpleToggleEffect changes:

(1) Assigns a default icon and a11y ID. If a fake effect is needed
with a different icon/ID, make another subclass of SimpleToggleEffect.

(2) Callback is the base class OnEffectControlActivated(), subclass
callbacks removed and the ability to assign any other callback is
removed.

Remove FakeVideoConferenceTrayController constructor body from the .cc
file, "default" it in the header.

Add a function to VideoConferenceTrayEffectsManager for retrieving
set-value effects.

Add function to VcHostedEffect for retrieving the number of states,
and make this function and GetState() const.

Bug: b/259838099
Change-Id: I87c23ab585e58188a7a90ab549a01b6ddd8bc9ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4093453
Reviewed-by: Andre Le <leandre@chromium.org>
Commit-Queue: Roger Tinkoff <rtinkoff@google.com>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082466}