[ios] Allow getting weak pointers on ChromeBrowserState objects
The lifetime of Objective-C objects is difficult to reason about
as they are reference counted, it is easy to extend lifetime of
an object by keeping a reference to it, and ARC makes this even
harder because the compiler can decide to autorelease objects
which extends their lifetime until the next autorelease pool drain.

This makes interacting with C++ model object difficult as the UI
elements may be semantically "destroyed" (i.e. the objects have
been removed from the view, ...) but block may still access to
their properties, including pointer to C++ model objects.

By allowing weak pointers to ChromeBrowserState objects, it become
easier to ensure that the UI has no dangling pointers. This is
simpler than other alternatives such as explicitly clearing the
property on the UI objects (by an explicit call to a `-shutdown`
method) and less brittle (i.e. it is difficult to ensure that a
`-shutdown` method is always called before an UI object is no
longer needed, or before the `ChromeBrowserState` is destroyed).

Bug: none
Change-Id: I5608f99912cadaee6910e099a85c18335b639870
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4096900
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082520}