[infra] Use a consistent ordering for definition arguments.
The modifications to have lucicfg fmt apply an ordering to the
definitions is not ready for automatic use because we need to be able to
avoid reordering lib code. This change manually applied changes to the
files defining builders with a local .lucicfgfmtrc so that builder configs can be migrated in an automated fashion without that functionality being ready.

Bug: 868153
Change-Id: Ifb47a996a6cae6b7a6ddbb27416226b1b0d52870
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4097866
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1082524}