[DIPS] Get site engagement from the DIPS database.
Add cookie_mode to DIPSRedirectChainInfo, and has_interaction to
DIPSRedirectInfo. Unlike the other members of those classes, they are
not const. They get set by DIPSService later since they are not known
at the time of construction. They're optional<> so we will know if they
are accidentally not set.

Make DIPSRedirectContext take a DIPSRedirectChainHandler instead of a
DIPSRedirectHandler; DIPSRedirectChainHandler takes a whole vector of
DIPSRedirectInfoPtrs instead of one redirect at a time.

DIPSService now calls DIPSStorage::Read() (asynchronously) and checks
whether the site has an interaction timestamp or not, instead of
calling the SiteEngagementService.

The UKM metric SiteEngagementLevel will now only contain the values 0
and 1, instead of the full range of the EngagementLevel enum.

DIPSService no longer depends on SiteEngagementService.

Bug: 1386250
Change-Id: I54753699c440b321e75e167ac6159fe3943ac7a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4089941
Commit-Queue: Ryan Tarpine <rtarpine@chromium.org>
Reviewed-by: Joshua Hood <jdh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082544}