Delay done showing for instant layout animations.
While the other three transitions are driven by the Layout, start
showing is driven by the manager. This creates a problem when a layout
that does not have any animation tries to finish showing in the middle
of the #show() call, but before observer get the start showing signal.

To fix this, the manager now blocks done showing signals while it is
in the middle of a start showing operation, and runs the done showing
logic afterward.

Bug: 1385523
Change-Id: I7a47a6557463636cb43a630c42ee827ea3d0bd67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4097749
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082642}