Revert "ax_ui: make AXPosition keeping an anchor as an AXNode instance"
This reverts commit 86543a2742f2de6bc6c989c48d8723ded3759273.

Reason for revert: this change introduced a reliability
regression, tracked with crbug.com/1394743
The premise of the original CL is that we can simply always get the
AXTreeId from the AXNode stored on the AXPosition using a weakptr.
However, after deleting an AXNode on which Narrator's focus is, we call
AXTreeObserver::OnNodeDeleted and expect to get the node id and tree
id on that position, but the AXNode is not there anymore, since
it was stored using a weak ptr.

Original change's description:
> ax_ui: make AXPosition keeping an anchor as an AXNode instance

> Change-Id: I7ae44cbcd789683715fc3fd32e832ed100845e2b
> Bug: 1362839
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3989063
> Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
> Commit-Queue: Alexander Surkkov <asurkov@igalia.com>
> Cr-Commit-Position: refs/heads/main@{#1076372}

Change-Id: Iec38187317e46d38faaa475d1966c182b3c0a53e
Bug: 1394743
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4098338
Commit-Queue: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1082685}