Update TextureLayerChangeInvisibleMailboxTest Ack Handling
When working on shifting the timing of DidReceiveCompositorFrameAck to
align with OnBeginFrame, we noticed a race condition on TSAN for this
test.

DidReceiveCompositorFrameAck was checking for a side-effect of a
callback ResourceReleased. However both of these are PostTasks, and it
is possible for the Ack to be processed before the callback.

This change updates the test to have those two methods ensure that both
have occurred for test teardown.

The rest of the DidReceiveCompositorFrameAck checks have been moved to
DidCommitAndDrawFrame, which reflects the actually timing we care about.

Bug: 1399880
Change-Id: I98d276c13ec06eb1b2b8ba1d70f6f2d72a3000df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4098425
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082696}