[Wayland] Ensure DND buffer size is a multiple of scale
When using a scale factor of 2.0 on Sway ToT, this fixes a crash when
trying to DND an image or text that happens to have an odd width or
height. The issue is Wlroots will not allow attaching a buffer to a
surface when the buffer's size isn't a multiple of the surface's scale
factor.

R=nickdiego

Bug: 578890
Change-Id: I283ed27a7d0051b895e0eed3b9a74d83feac3a37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4099423
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082777}