[NamedMojoIpcServer] Pass Windows security descriptor in ctor
The Windows endpoint connector creates the server endpoint with a
security descriptor that only makes sense in the CRD setup, so this CL
cleans up the code a bit and allows the caller to pass the security
descriptor to the constructor directly, via a new struct called
"EndpointOptions". Server name and message pipe ID are also moved to
the new struct, which makes future extensions of the options more
manageable.

Bug: 1378803
Change-Id: If1703fa6be8631406faf898b40ec3a3d3cedfaf5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4099965
Reviewed-by: Noah Rose Ledesma <noahrose@google.com>
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082792}