cros_next: fix the wrong disabled icon color
We recently add a new argument icon_changed in
IconButton::UpdateVectorIcon method such that we only need to
update the image model for disabled state when the icon is
changed. However, this argument should be set to true when we first
call it in ctor otherwise there will be no image model for disabled
state.

The screenshot of disabled icon buttons before fix is:
https://screenshot.googleplex.com/8H6p3mnbtDadnYW

The screenshot of disabled icon buttons after fix is:
https://screenshot.googleplex.com/8xSABhrUhZN8qjN

Bug: b/262398402
Change-Id: I3142df8c37792d8a32a75bd9557a2ffcc50a50e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4104640
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Reviewed-by: Connie Xu <conniekxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082793}