[a11y] Fixed issue with propagation of disabled state in DateTime
There was a bug with the disabled state of a `DateTime` element not
being propagated down to its descendants which cause issues where
some accessibility ATs were not announcing the
elements as disabled.

For example, with Narrator on Windows, Scan Mode navigation would
not announce the inner spinbuttons as disabled in the following example: data:text/html,<input type=date>.
The disabled state was only announced for the date time element.

The fix consists of two parts:
First, `AXObject::ComputeDescendantOfDisabledNode` was not returning the
correct value, as it was never actually querying the parent's disabled
state. So a check was added so this was actually being done now.

Second, `AXObject::IsDisabled()` was returning false for the children of
disabled `kDate` objects, so a check was added so that elements that
have a shadow host that is disabled and are a control
return true in `AXObject::IsDisabled`.

Change-Id: I9bff6c3b2708d4b0390eaf1c43fc1ab897256392
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4083950
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082808}