Add deletion in bookmarks++ bulk edit mode
Adds a delete button to the bookmarks++ side panel footer in edit mode,
with delete functionality. Does not yet include a warning dialog for
deleting bookmarks with associated powers, as price tracking is the only
integrated power so far and it does not require this dialog.

Bug: 1365606
Change-Id: I11f1dae63e18e48ed94d30bf16539736418ac182
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4090688
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Emily Shack <emshack@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082844}