[QuickStart] Refactor TargetFidoController->AuthenticatedConnection
This CL refactors TargetFidoController to be a part of the AuthenticatedConnection class.
The design inspring this refactor can be found in
go/cros-quick-start-base-client-class. The idea is to consolidate all of the target device client logic in with AuthenticatedConnection to avoid code bloat of having multiple classes.

TEST=tested via unit tests

Bug: b/260010484
Change-Id: I2d576328977653fdf4b5377adcb03d888fed6aee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4048884
Reviewed-by: Michael Hansen <hansenmichael@google.com>
Reviewed-by: Curt Clemens <cclem@google.com>
Reviewed-by: Brittany Hartmire <bhartmire@google.com>
Commit-Queue: Jason Rhee <jasonrhee@google.com>
Cr-Commit-Position: refs/heads/main@{#1082852}