Skip HTTPS-First Mode upgrades for hosts with HSTS
Before this change, HTTPS-First Mode would intercept all navigations
and attempt to upgrade them to HTTPS (by serving an internal redirect
and tracking their success) before HSTS could apply to the request.
This meant that an HTTPS upgrade could fail and the HTTPS-First
interstitial would show for sites that have HSTS (for example, due to
a slow page load) -- this interstitial is confusing and not useful to
users, as a site with HSTS explicitly does not support HTTP (and so
does not support HTTPS-First Mode's fallback).

This change adds an additional lookup when HTTPS-First Mode checks if
it should intercept and upgrade the navigation. After this change, the
HttpsOnlyModeUpgradeInterceptor will make a call to the network
service to check if the hostname has HSTS active, and if it does skip
upgrading the request and allow HSTS to handle the upgrade.

Bug: 1294328
Change-Id: I9c5344f51f225d5689f77af1274b281ffd6a8798
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4093485
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082874}