Adjust stylus event to allow handwriting in nearby inputs
The computed touch action for an input element allows stylus writing
only in the editing area of input box. This makes it slightly tough to start writing in many search fields that have a lot of padding outside the editing area and inside the Input box.

Adjust the pointer event to nearest input in some fixed area around the pointer down to improve the chances of writing in input fields.
This is useful to support the upcoming split-merge ('|') stylus gesture,
which is highly likely to start outside the editing area.

Bug: 1398127
Change-Id: I957ea0e4bd1de51f95ac915a4c11882fae39d4a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4085362
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Mahesh Machavolu <mahesh.ma@samsung.com>
Cr-Commit-Position: refs/heads/main@{#1082918}