Move FileSystemEntryURLLoader security check to the UI thread.
Security checks inside ChildProcessSecurityPolicy are currently allowed
on both UI and IO threads. Security team has a goal of enabling better
security checks which are only possible on the UI thread.

This CL moves the actual security check earlier, while the code is still
executing on the UI thread. Once those are performed, the result is
passed in as a parameter to the IO method which continues the creation
of the FileSystemEntryURLLoader. The semantics remain exactly the same
and this approach avoids adding more thread hops.

Bug: 1286497
Change-Id: I15f438d834b8f25285ec690c91b35a0b1d8f3a6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4104641
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082927}