Make ForLayerComparison() somewhat clearer.
We drop the conditional XOR mask in favor of a simple if, and add some
comments.  As a tiny bonus, the new code also allows us to return
a simple uint64_t, making the surrounding code generation a bit simpler.

Performance is neutral.

Change-Id: I4d9ae846acc4dab2a5e2f430fc3063ec6b7569b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4095031
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083010}