Add an interface that allows responding to ICE observer events. (#3/n)
The new interface allows blink to respond to the ICE observer events raised by the ICE controller, and also to initiate actions. This lets blink manipulate the connections used by the ICE transport through an ICE controller. Observers get a handle to this interface when attached to an ICE controller.

Bug: 1369096, webrtc:14131
Change-Id: I52783feac1603b39d65a9b63bc1dcfc8e9c89e66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4079121
Commit-Queue: Sameer Vijaykar <samvi@google.com>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083014}