[//media] Don't build GLES2DecoderHelper on Android
This file is used only by the VDA implementations, which are not used
on Android. Eliminating its being built on Android enables followup
simplifications in //gpu code.

It should likely be possible to move the one method of
CommandBufferHelper that is called on Android into the client and
eliminate CommandBufferHelper being built on Android as well. However,
there is some fiddliness there as CommandBufferHelper has nontrivial
logic with respect to the destruction of its |stub_| ivar, so I didn't
pursue that in this CL.

Bug: 1323341
Change-Id: I8f39a46648181638c4603102620ee7c2120e42c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4097090
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083032}