[WebAuthn] Use TouchToFill sheet for non-conditional passkey sign-in
This change modifies non-conditional WebAuthn get requests on Android,
when the allowCredentials list is empty. Currently those requests are
passed to GMSCore for passkey selection. After this CL, the selection
will be handled by the Touch To Fill bottom sheet currently used by
the password manager and WebAuthn conditional UI.

The goal is to have the same UI experience for users invoking
passkeys, whether it is using a conditional or non-conditional
WebAuthn call.

After passkey selection, the bottom sheet is closed and the selected
key is added to the requests allowCredential list. It is passed to the
GMSCore API for the user verification step.

Bug: 1393167
Change-Id: Ic189be26cd0e937e061c4d83ad0893e407cbc95a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4056155
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083073}