[UPM][fix] Fix potential PasswordStore bridge init race condition
This CL fixes potential use-after-free that could occur during the
PasswordStoreAndroidBackend initialisation.

Dispatcher bridge java side is initialised on the background thread and
requires a reference to the java side of the receiver bridge. It was
possible for the native receiver bridge to be destroyed before the
reference to its java counterpart is acquired.

Now the reference is acquired on UI thread and binded for the later use
on the background thread.

Bug: 1394715
Change-Id: I5d4400c7c72b3f4cb63167e6e662f39678c00781
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4103582
Commit-Queue: Maxim Anufriev <maxan@google.com>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083079}