[WebAuthn] Add a null check for WebContents on Android WebAuthn requests
There are crash reports indicating the WebContents associated with a
given frame is sometimes null, which is dereferenced during conditional
UI requests.

The exact conditions under which the null WebContents manifests are
unclear, but this CL adds a check to abort the request when that
occurs.

Bug: 1399887
Change-Id: I225cc4dc1e8e43872fdbc29100f8790a788db1c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4103711
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083085}