Fenced Frames: [4/N] Fledge `runAdAuction()` yields a FencedFrameConfig
This CL doesn't fully implement the change advertised in the headline
(resolving a web-exposed FencedFrameConfig object in the Promise
returned from `runAdAuction()`); it is a refactoring in preparation for
that change.

This CL makes the RunAdAuction() IPC callback exclusively return an
optional `blink::FencedFrame::RedactedFencedFrameConfig` to the
renderer, where the `NavigationAuction` object is responsible for
pulling the internal URN off of this object and returning it to script.
This URN is identical to the URN that was previously returned to the
renderer via the same callback, hence the "refactoring" nature of this
CL.

To do this, we add a non-nullable URL member to the
blink.mojom.FencedFrameConfig struct, with appropriate validation and
unit testing.

A subsequent CL will modify the input dictionary to `runAdAuction()`
to give web developers the ability to express that they want a
full-blown config object returned to them in the `Promise` that
`runAdAuction()` resolves to, not just a URN. That CL will be easy to
write since the backing config will already be in place in the
renderer.

Bug: 1347953
Change-Id: I88a7ca0f8b59c7f31f9df406a8ef382c04d6c67d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4098955
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Garrett Tanzer <gtanzer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083114}