Removing SetLosslessEncode method from VideoEncoder classes
This setting hasn't been used in a number of years and as of the last
time I tested it, enabling it resulted in a noticeable slowdown in
encoding speed. This CL removes the method from the interface and
classes since it isn't currently used and I don't foresee us using it
any time soon (plus it removes any potential for confusion between
lossless_encode and lossless_color).

Change-Id: Icc74e63d852e8e1bf7cb029a19d6928d1b97714c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4104843
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083194}