Add short names for the ci.chromium.cft console.
I had thought that Milo would default to showing the builder
name for a short name, but apparently it just doesn't show
a name at all if no short name is specified.

This CL adds "short" names that match the builder name.
They're not very short, but this is okay since we're only
putting three builders on the console for now. We can add
categories and real short names if we need them.

Bug: 1396097
Change-Id: I4ac09bf16d68232b64413c7a0a1b84bfa54e633d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4104966
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Wanda Mora <morawand@google.com>
Cr-Commit-Position: refs/heads/main@{#1083207}