[Fast Pair] Stop GATT connect timeout only if correct service discovered
The GATT connection timeout timer stops when GattDiscoveryCompleteForService is called from the adapter if the service discovered is the Fast Pair service for the device associated with the FastPairGattServiceClientImpl which the function belogn.

This function is called for every observing the adapter, regardless of whether the service discovered belongs to the device to which the client is associated, so this is checked inside of the aforementioned function.

The call to stop the timer is moved inside of the if branch checking this condition from its erroneous position outside of it.

      Initial pair was successfully completed on DUT. Given that the
      GATT client is fast pair protocol agnostic, this is sufficient
      evidence that the change doesn't affect the GATT connection.

Test: unit tests in this CL and test on DUT:
Fixed: b/262270537
Change-Id: Ifdd684a0fe1b55eaca4f37dc9d774e3a3a49053e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4099546
Reviewed-by: Juliet Lévesque <julietlevesque@google.com>
Commit-Queue: Brando Socarras <brandosocarras@google.com>
Cr-Commit-Position: refs/heads/main@{#1083215}