Update androidx espresso to 3.5.0.
Current version is 3.2, which is 2019 time frame and not compatible with
accessibility_test_framework 4.0, which we're trying to add. Of note is
we don't actually use the androidx espresso code, so this change will
not do anything that affects currently running tests.

Bug: 1351758
Change-Id: Id469ff78dc5378edfcbc909220af740f4ca46ef3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4105680
Commit-Queue: Benjamin Joyce (Ben) <bjoyce@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Auto-Submit: Benjamin Joyce (Ben) <bjoyce@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083246}