Extensions: SidePanel: Don't set the optional path while enabled is true
Document the expectation that the API will respect an API call that
doesn't have a path even though enabled is true. A case can be made for
the expected behavior either way. Instead of saving what the API was
called with, the API could instead error if the optional path is missing
when enabled is true.

Fixed: 1400829
Change-Id: I0d0c9c6eb8e036506a43aab06893d94ead42fc11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4103650
Reviewed-by: Kelvin Jiang <kelvinjiang@chromium.org>
Commit-Queue: Solomon Kinard <solomonkinard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083255}