[eche] Removing the sorting of recent apps based on timestamp.
As the items that come from the phone does not have the timestamp
data, sorting the apps based on the timestamp will lead into purging the
app that are used on the phone when the list is sorted.
After this change, the list of apps are kept in a LIFO fashion.

Test= Locally built chrome and tested like this: Opened app A and app B on the
phone. Then connected phone hub. At this time we had app A and B at the head of
the list. Then streamed app A. Now app B is still in the list (before the change
local build was not behaving this way and after the new change it was fixed.)

Bug: b/262195624
Change-Id: I6e25902e687321bc8b95b34e3143c8bb30189245
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4099258
Reviewed-by: Jon Mann <jonmann@chromium.org>
Commit-Queue: Abbas Nayebi <nayebi@google.com>
Cr-Commit-Position: refs/heads/main@{#1083256}