[Widget] Respect foldable device measurements.
This change updates the Widget code to respect the foldable device
measurements on Android S and beyond.
When an array of sizes is supplied by the launcher, and Chrome is
running on a modern enough OS that supports creation of RemoteViews
mappings, the size hints are used to construct widgets.

Bug: b/244273263
Change-Id: Ib57baba78d11619b07a7b42a110fe9a95d2797e9
Fixed: b/244273263
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4104085
Reviewed-by: Gang Wu <gangwu@chromium.org>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1083311}