webauthn: show better error message for missing passkeys.
Currently, if no credentials match the platform authenticator then we
remove it as a mechanism. If there are no mechanisms left, we show an
error about being unable to support the type of security key needed.
But that's confusing when the site wanted the user to sign in using the
platform authenticator but there weren't any matching credentials.

Thus, if "internal" was a candidate transport for any credential and no
mechanisms remain after filtering, tweak the language on the error
screen to match the strings form Android.

Bug: 1396550
Change-Id: I838f95016f3692ba2233ca009d10c7c39eaca5b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4108469
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083328}