System Nudge Time To Action Metric
go/nudge-time-to-action

Implement metric that tracks the time it takes for a user to perform an
action informed by a System Nudge.

Starts measuring time when the nudge is shown and records the Nudge
Catalog Name in one of the time buckets available when the user
performs the expected action. Time buckets are "Within1m", "Within1h"
and "WithinSession".

b: 256189359
Change-Id: Idfca0230c801087211ee1eba528f7012eb9d71e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4063247
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Kevin Radtke <kradtke@chromium.org>
Reviewed-by: Kevin Radtke <kradtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083335}