[cc-side-panel] Only check selected color
This CL also makes the color circle smaller if selected so that the
color circle plus the selected ring equals the size of an unselected
color. To facilitate that, this CL moves the check mark wrapper into the
color element so that the checked bit only has to be set once.

https://screenshot.googleplex.com/AFjfoGFah7KPX9E

Bug: 1384229
Change-Id: Icc197cd6cd8c1b3aa1bba06f18ed3c026fd9cab1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4104289
Commit-Queue: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Paul Adedeji <pauladedeji@google.com>
Cr-Commit-Position: refs/heads/main@{#1083338}