immersive fullscreen mac: pin titlebar on reveal with sub-windows
When the titlebar is fully revealed this CL pins the titlebar if a
titlebar lock is active. This allows top chrome and its sub-windows to
stay in place even if the mouse leaves the top chrome area, preventing
misclicks of the omnibox results window and extensions.

https://drive.google.com/open?id=1qSVAMjp_MlIPhJPQ2_VO-WznAidCvI8f&authuser=0&resourcekey=0-qrZcY0jBswhBAa_u7-sC1g&usp=drive_link

Also rename NSWindow member variables from *widget_ to *window_.

Bug: 1396532
Change-Id: I8f58a6651cb89a9023c489990d8571a9c1466b7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4093914
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Tom Burgin <bur@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083359}