[Lacros] Load Lacros according to LacrosSelection policy.
go/lacros-selection-policy

Use LacrosSelection policy to decide which Lacros binary to load: rootfs
or stateful. If the policy is set to user_choice or unset, the user can
decide which one to load with lacros-selection flag, or load the
freshest binary version wise.

If the policy is set to rootfs or stateful, the corresponding Lacros is
loaded and lacros-selection flag is ignored and hidden.

The effective policy value does not change during the user session. The
cached value is used for decision making. If the policy changes during
the session, the new value will be applied on the next login.

Bug: chromium:1363880
Test: manual on device
Test: unit_tests --gtest_filter=BrowserUtilTest.*
Test: unit_tests --gtest_filter=BrowserLoaderTest.*
Change-Id: I561156913647e5658291321c8b9afa84c0855cfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4065943
Commit-Queue: Artem Sumaneev <asumaneev@google.com>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083416}