[ SavedTabGroup ] Change update observer function
Changes the update observer function to take an optional tab_id
parameter.

With this change we now only update the ModelTypeStore with the data
that has actually changed. This is more efficient compared to the
previous implementation since we no longer update an entire group and
its tab contents when any change / update was made.

Additionally, this fixes a bug where calling RemovedTab would not
actually remove the tab from the ModelTypeStore. This was evident when
a user would restart the browser and still see a tab that had removed
persisting in the group.

Bug: 1343409
Change-Id: I3d32ad52351df6b5c2c25f43b9c6a402c7c0f4dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4032404
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083435}