[Media Router] cleanup dangling pointer in MediaRouterMojoImpl
This patch is a follow-up to a previous cleanup patch, where the
ownership semantics for MediaRoutesObserver is cleaned up to avoid a
dangling pointer crash.

This patch switches to a checked ObserverList pattern for this observer
and ensure that this observer is always destructed on the UI thread,
resolving potential concerns with NotifyOfExistingRoutesIfRegistered
being called with an observer while it is being unregistered on a
different thread, potentially causing a crash.

Bug: 1381416
Change-Id: I241fc6dc7238cde52d2c52adabb51097ba89903d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4076037
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: george benz <gbj@google.com>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083438}