Remove RenderViewHostChanged from MetricsWebContentsObserver
RenderViewHostChanged is an old API which is deprecated and we would
like to remove. This CL moves the MetricsWebContentsObserver code to
use the newer RenderFrameHostChanged API and considers only the primary
main frame, which is the case where RenderViewHostChange would also be
dispatched.

The RenderFrameHostChanged event for the itniail RenderFrameHost is
dispatched too early for the metrics observer to see it, but it is
called for all subsequent changes. The reason why this change is safe
is that the constructor of the MetricsWebContentsObserver is explicitly
calling RegisterInputEventObserver with the primary main frame and
it takes care of this initial event not being observed.

The CL goes a small step further to unregister the listener when the
WebContents object it observes is destroyed, so we can avoid having a
stale entry in the input event observer list.

Bug: 1168562, 1394513
Change-Id: Id70f63efe0aba2b0819778040abedf86866c32d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4093119
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083443}