[journeys] Attach a "takeover" Action to HistoryClusterProvider match
This CL unconditionally attaches a HistoryClustersAction action to the
HistoryClusterProvider match.

It doesn't need to be guarded by a flag, because the Action execution
already takes the user to the right place by checking the flag.

It also makes the "action"-centric metrics work for the Dedicated
Suggestion row, as there's actually an action attached.

This CL also hides the "takeover" actions from rendering.

Bug: 1327076
Change-Id: I49ee9480cdde11a3c570dd16f0cd637023a32161
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4103705
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083473}