Notify the state of the drag handle with Chromevox when focusing on it
Previously with the Chromevox a11y is on, there is no announcement that
mentions the state of the shelf apps when the drag handle is focus,
that is, the users won't know if the apps in the hotseat is shown or
hidden. This cl adds additional states to the drag handle view so that
it can announce if the shelf apps are showing or not.

Bug: b:261648688
Change-Id: I8a64f10329d79f639a0ba288411f068dd9a40576
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4104481
Commit-Queue: Wen-Chien Wang <wcwang@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083499}