Settings WebUI: Make policy tooltips accessible to screen readers
Having `display: none` on the tooltip for all policy indicators in this
component was causing it to be skipped by screen reader. The focusable
icon in cr-tooltip-icon has an `aria-describedby` set to the tooltip
text, so the tooltip needs to be visibly hidden but still accessible
by screen readers.

Fixed: 1347653
Change-Id: If1716cd0eb174f12433254feb949e8437f64ddc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4104244
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083500}