webview: fix init order in startChromiumLocked.
We should be draining the run queue last, once all initialization is
done, as the purpose of the run queue is to hold tasks that require
initialization to be complete. A few things got added after it; move it
back down to the bottom and add a comment for future reference.

Change-Id: I49b0ad4e6f641d6afce39e15b92bcbc1c90efccb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4107108
Auto-Submit: Richard Coles <torne@chromium.org>
Commit-Queue: Michael Bai <michaelbai@chromium.org>
Reviewed-by: Michael Bai <michaelbai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083501}