Add enterprise.platformKeys.challengeKey method
Add a new method to the enterprise.platformKeys extension API to allow
specifying the algorithm of the key to generate (RSA/ECDSA). The new
`challengeKey` method takes options as a dictionary to simplify adding
new parameters in the future.

A design doc for this proposed method can be found here:
https://docs.google.com/document/d/1xUhrnceJUHTK3MGYobvV4iG-NP7Q9ytddcHtPxUzo18.

For broader context on why we are adding ECDSA support to
`challengeKey`, see http://go/cros-ecc-gecc.

Bug: b:252786598
Change-Id: I897f4e345e49efc16be15745d442a44e7f37268c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3983593
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Kevin King <kcking@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083573}