[UPM] Split PwM settings updater bridge into dispatcher and receiver
This CL does not change behavior.

This is a preparation for moving the GMSCore interaction out from the UI
thread. Such change will make all native -> java calls executed on the
background working thread while operations callbacks will be called by
the GMSCore on the UI thread. It is preferred to keep to not allow a
single object to live on several threads.

The same refactoring for PasswordStore was done in crrev.com/c/4079256.

Bug: 1394715
Change-Id: Idbe78df5e27d37f31e26cf9f88d443231e7f0e33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4098884
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Commit-Queue: Maxim Anufriev <maxan@google.com>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083667}