Generic DevTools FormLabelForNameError issue.
Using the for-attribute, <label>s may specify the ID of their
associated control. However, we see that a significant number of
websites incorrectly specify the name of a control instead.
This CL creates a new DevTools issue to warn developers about this
mistake. It utilizes DevTools' GenericIssue and is emitted within
Autofill in the renderer processes.

This CL only affects the Chromium part, and a follow-up will make the
necessary changes in DevTools-frontend. The result will look something
like this: https://photos.app.goo.gl/GAnAjomZfaz2FXcYA (strings are
still subject to change).

Technically, the CL makes the following changes:
- Add a new GenericIssueType FormLabelForNameError to DevTools'
  browser_protocol.pdl.
- Add the ability to emit GenericIssues from the renderer process via
  WebLocalFrames, based on what crrev.com/c/3856871 suggested (thanks!).
- Augment the GenericIssueDetails by a DOM.BackendNodeId, which will be
  used to link the issue to the <label> element in the UI.
- Rename WebNode::GetDevToolsNodeIdForTest() to
  WebNode::GetDevToolsNodeId(), since it's (I believe) currently not
  possible to access the node ID from outside Blink. According to
  http://crrev.com/c/chromium/src/+/3680459/comments/cb2b19e1_389ac54b,
  this should be fine.
- Emit the issue from within the Autofill code. Autofill's label
  inference code already has the necessary logic to associate forms
  with their labels, even if the name instead of the ID is used.

Bug: 1349788
Change-Id: I0deba8295fac46f1f04b2525ce75f8eb085f19f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4027566
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1083678}