[UPM] Move password settings operations execution from the UI thread
All password settings accessor operations that involve GMS Core API
communications are moved from the UI thread.

Execution on the background sequence is handled by helper class.

PasswordManagerSettingsService lives on the UI thread. It instantiates
PasswordSettingsUpdaterAndroidBridgeHelper (helper) which also lives on
the UI thread.

Helper instantiates two JNI bridges:
- One (dispatcher_bridge) invokes backend operations
  ((Get|Set)PasswordSettingValue), this bridge only manages calls from
  native to java.
- One (receiver_bridge) for handling operations callbacks
  (OnSettingValueFetched etc.), this bridge only manages calls from java
  to native.

Native parts of both bridges are instantiated on the UI thread.
However, the dispatcher_bridge is initialised (via Init() method) on
the background thread.

Java part of the dispatcher_bridge is instantiated on the background
thread during the native part initialisation (Init() method).
Note that JNIEnv is not thread-safe and calls to the java object
should be done on the same background thread. This is guaranteed
by the helper which executes all operations on dispatcher_bridge by
posting tasks on the background sequence.

Both native and java parts of the receiver_bridge live on the UI
thread. This bridge is only called from the Java side and forwards
all calls to the settings service which also lives on the UI thread.

The same change for PasswordStoreAndroidBackend was implemented in
crrev.com/c/4066003

Bug: 1394715
Change-Id: Ib5600150a2c4f17dd971955de1b9640b84a15d85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4105085
Commit-Queue: Maxim Anufriev <maxan@google.com>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083700}