[WebAuthn] Abort WebAuthn request when Manage Passkeys selected
As of https://crrev.com/1083073, discoverable credential WebAuthn get
requests on Android use the Touch To Fill bottom sheet for account
selection.

This sheet includes the ability to launch password management by
dragging the sheet upward and selecting 'Manage Passkeys'.

Unfortunately, when a user does that the WebAuthn request is left
outstanding, and the promise does not resolve. This means if the user
returns to the page after dismissing the password manager, they cannot
once again attempt to sign in.

This change causes the promise to resolve with a Not Allowed error
when passkey management is opened, which is the same as if the user
dismissed the sheet.

Bug: 1393167
Change-Id: I2ec7babef5b9095eb155380f7ff8f45564c29aaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4108254
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083714}