[Passwords] Add cross-origin iframe as a WaitForUsernameReason.
Forms in cross-origin iframes are not filled on the page load, there
is a check in the renderer code that prevents that (see
IsInCrossOriginIframeOrEmbeddedFrame in PasswordAutofillAgent).

This CL makes the following changes for forms in cross-origin
iframes:
1) PasswordManager.FirstWaitForUsernameReason metric is recorded
2) the code that fills on the page load is not invoked anymore
(previously it was invoked but always failed)

This CL does not remove the check on the renderer side, as it's more
powerful security-wise and covers cases such as: Frame of origin A,
hosting Frame of origin B, hosting frame of origin A.
On the browser side we don't have access to frame hierarchy and cannot
catch such cases.

Bug: 1401319
Change-Id: I29d98fce3bbf6ea9dbe56c26e9f05f5e21bd73db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4100290
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#1083720}