Store time in PSTs as a Timestamp datatype
Currently in PSTs some time fields are stored as ints and some are
stored as strings. We want to be consistent in how we store time in
PSTs so we will store all time fields with a Timestamp datatype,
which consists of a single int representing microseconds since the
Windows epoch.

Bug: 1395819
Change-Id: I60e01b316a457619bd3e08ede4c4965159aed802
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4067220
Reviewed-by: Steven Valdez <svaldez@chromium.org>
Commit-Queue: Ryan Kalla <ryankalla@google.com>
Cr-Commit-Position: refs/heads/main@{#1083726}