[iOS][signin] Reenable and fix policy-related signin tests
Re-enable /testSyncTypesDisabledPolicy with a few changes. Don't
verify the content of the sign-in promo in Settings which isn't
relevant for the test.

Stop setting local prefs in /testSigninDisabledByPolicy. This probably
had cross-talk with other policy-related tests in the past because
local prefs can persist across tests if something goes wrong during
the cleanup.

Bug: 1279995
Change-Id: Ie05b23fc8b1a693417e569a4e6b6aa897f0c5f6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4107608
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#1083728}