Use CrdSessionType in StartCrdSession remote command
This replaces the `curtainOffLocaUserSession` boolean which was
previously introduced. This change makes this command more consistent
with the other FetchCrdAvailability remote command, where we do not
mention 'curtain' but we do talk about the `supportedCrdSessionTypes`.

Note that this `curtainOffLocalUserSession` was never used server side,
so it is safe to completely drop support for this boolean.

Bug: b/261569067
Test: unit_tests "*Crd*" and manually deployed
Change-Id: Ib7521e09190b35b834ef4cda0763be67229e7592
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4092163
Reviewed-by: Roman Sorokin <rsorokin@google.com>
Reviewed-by: Ben Franz <bfranz@chromium.org>
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Cr-Commit-Position: refs/heads/main@{#1083766}