dPWA: Deflake IsFileHandledBySite check on Mac.
This method used to be implemented by enumerating all the apps that can
handle the specific file and then checking if the expected app is
amongst those. That however depends on LaunchServices having finished
indexing the new app already. While we do register new apps with
LaunchServices when we create them, it appears there is still some delay
until this query starts succeeding.

The LSCanURLAcceptURL method does not seem to care if the app URL has
been indexed yet by LaunchServices, so by switching to that method we
should solve these flaky failures (this won't however help for other
flakiness that might also be caused by this delay in LaunchServices
recognizing the newly created apps, although I'm not sure if this is
the cause of more problems).

Bug: 1399439
Cq-Include-Trybots: luci.chromium.try:mac12-arm64-rel
Change-Id: Ida31d58f5049815ea933f52a9127af09b1327a91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4108080
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083799}