Fenced frame: Add tests for navigating nested configs
A previous CL introduced the `fence.getNestedConfigs()` API[1], and
another one allowed fenced frames to navigate configs[2]. This CL adds
testing coverage for being able to navigate nested configs created with
FLEDGE.

Updates the `generateURNFromFledge` helper function to also take in an
optional list of urls to use as the nested config urls. When set, the
urls will be read into the list of nested configs, which can be used
to navigate a fenced frame. This CL also updates the FLEDGE logic used
by the test to support component auctions.

[1]https://chromium-review.googlesource.com/c/chromium/src/+/4035572
[2]https://chromium-review.googlesource.com/c/chromium/src/+/4086678

Change-Id: I1803ff00069ab092679a10a4765b8b8cd90773da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4099227
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Cr-Commit-Position: refs/heads/main@{#1083854}