[Extensions] Clean up Proxy API constants.
This CL cleans up proxy_api_constants.h my moving constants closer to
where they are used. This is a step in moving much of the functionality
in this api directory lower in the stack.

The CL also changes the signature of one of the helper functions. This function previously return bool to indicate success or failure. However, the function always returned true, so the function now returns the actual value it's constructing.

Bug: 883570
Change-Id: I67a158faa17075a5b297330fe1bbdc3c882d764f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4108077
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083857}