[User Education] Do not duplicate config when local config is present.
This allows setting up a single, default local feature promo config in
components/feature_engagement/public/feature_configurations.cc which is
triggered if the feature is enabled with no configuration given.

The alternative is that a feature engagement configuration must be
duplicated in every Finch configuration and every test that needs to
enable the feature, or if the feature is enabled from the command line.

Bug: 1381673
Change-Id: I316da3b5e0e5aaf5ce5347c2fa8474b2ec6328a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4098151
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083861}