breadcrumbs: replace disk read with GetEvents()
This change makes BreadcrumbPersistentStorageManager get the list of
breadcrumb events from BreadcrumbManager, rather than reading them from
disk.

The BreadcrumbManager was recently made a singleton, which makes it now
possible to treat it as a single source of truth for the log of
breadcrumb events.

Fixed: 1360583
Bug: 1163762
Change-Id: Ib65f7d5cb46903a7b9dbf9a881ea3806a75c45ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4083935
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083887}