GPUCanvasContext: sever link with swap buffers on destruction
Otherwise the WebGPUSwapBufferProvider could keep a pointer to the
GPUCanvasContext and call methods on it after it was destroyed. This
happened when the WebGPUSwapBufferProvider was giving images to an
onscreen OffscreenCanvas and the GPUCanvasContext was GCed in between
the PrepareTransferrableResource and the callback that the compositor
was done with the resource (which later calls into the client).

Fix this by Neuter()ing the SwapBufferProvider in the GPUCanvasContext
destructor. It is ok to do this in the GC object destructor because
WebGPUSwapBufferProvider is not a GC object.

(cherry picked from commit 410d3aeadd061b15c4654012102cce45f326ab4b)

Fixed: chromium:1406265
Change-Id: Ib78af5ce8f32fd1ba5718043c20fa652065e2ea4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4181166
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Cr-Original-Commit-Position: refs/heads/main@{#1094672}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4201046
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/branch-heads/5481@{#733}
Cr-Branched-From: 130f3e4d850f4bc7387cfb8d08aa993d288a67a9-refs/heads/main@{#1084008}